% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_bms_login.R
\name{brapi_bms_login}
\alias{brapi_bms_login}
\title{post /token}
\usage{
brapi_bms_login(con = NULL)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}
}
\value{
A list object of class 'brapi_con' with the authentication token filled.
}
\description{
Login into a BrAPI compatible database.
}
\details{
Although the POST /token call for authentication has been deprecated
after version 1.1 of the BrAPI specification, its functionality is still
retained for compatibility reasons with the
\href{https://bmspro.io/}{Breeding Management System (BMS Pro)} of
\href{https://integratedbreeding.net/}{Integrated Breeding Platform} to allow
users to obtain a valid token from within R and use with additional BrAPIR
v2 functions.
}
\examples{
\dontrun{
# Create a connection object for BMS
con <- brapi_db()$bms_test
class(con) <- c("list", "brapi", "brapi_con")
# Add a username between the double quotes
con[["user"]] <- "" # put your username in between the double quotes
# Add a password between the double quotes
con[["password"]] <- "" # put your password in between the double quotes

# Login
con <- brapi_bms_login(con = con)
}

}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.1/Specification/Authentication/Authentication.md}{BrAPI v1.1 Specs GitHub}
}
\seealso{
Other brapiutils: 
\code{\link{as.brapi_db}()},
\code{\link{brapi_canInternet}()},
\code{\link{brapi_checkCon}()},
\code{\link{brapi_connect}()},
\code{\link{brapi_db}()},
\code{\link{brapi_serverinfo_expiration}()},
\code{\link{brapi_serverinfo_show}()},
\code{\link{is.brapi_con}()},
\code{\link{print.brapi_con}()},
\code{\link{print.brapi_db}()}
}
\author{
Maikel Verouden
}
\concept{brapiutils}
