% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_canInternet.R
\name{brapi_canInternet}
\alias{brapi_canInternet}
\title{brapi_canInternet}
\usage{
brapi_canInternet(url = "www.google.org")
}
\arguments{
\item{url}{character; A url to check. Default is \href{https://www.google.org}{www.google.org}.}
}
\value{
NULL when internet access is available or ERROR when internet access is not available
}
\description{
Check for internet connection.
}
\references{
\href{http://stackoverflow.com/questions/5076593/how-to-determine-if-you-have-an-internet-connection-in-r}{Stack Overflow} and \code{\link[curl]{has_internet}}
}
\seealso{
Other brapiutils: 
\code{\link{as.brapi_db}()},
\code{\link{brapi_checkCon}()},
\code{\link{brapi_connect}()},
\code{\link{brapi_db}()},
\code{\link{brapi_serverinfo_show}()},
\code{\link{is.brapi_con}()},
\code{\link{print.brapi_con}()},
\code{\link{print.brapi_db}()}
}
\author{
brapir generator package
}
\concept{brapiutils}
