% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_connect.R
\name{brapi_connect}
\alias{brapi_connect}
\title{brapi_connect}
\usage{
brapi_connect(
  brapiDb = NULL,
  secure = FALSE,
  protocol = "http://",
  db = "127.0.0.1",
  port = 80,
  apipath = NULL,
  multicrop = FALSE,
  commoncropname = "",
  user = "",
  password = "",
  token = "",
  granttype = "password",
  clientid = "brapir",
  bms = FALSE
)
}
\arguments{
\item{brapiDb}{list; Connection object for a BrAPI compliant database created by the \code{as.brapi_db()} or \code{brapi_db()} functions. Default is NULL.}

\item{secure}{logical; Whether or not a secure connection ("https://") should be used. Default is FALSE, other possible value is TRUE.}

\item{protocol}{character; Scheme identifying the protocol to be used to access the resource via internet. Default is "http://", other possible value is "https://".}

\item{db}{character; Host name address of the BrAPI compliant database to be used. Default is "127.0.0.1" (defines localhost).}

\item{port}{integer; Port number to be used to access a BrAPI compliant database via internet. Default is 80.}

\item{apipath}{character; Path identifying where the Breeding API has been implemented. Default is "".}

\item{multicrop}{logical; Whether (TRUE) or not (FALSE) the database is a multicrop system. Default is FALSE, other possible is TRUE.}

\item{commoncropname}{character; Common crop name to be used, especially when the database is a multicrop system. Default is "".}

\item{user}{character; Username to be used, when logging in is required. Default is "".}

\item{password}{character; Password to be used for the specified user, when logging in is required. Default is "".}

\item{token}{character; Token to be used for authorization, will be passed with every call in the header. Default is ""}

\item{granttype}{character; Grant type. Default is "password".}

\item{clientid}{character; Client identification string. Default is "brapir".}

\item{bms}{logical; Whether (TRUE) of not (FALSE) the database used is from the Breeding Management System (Integrated Breeding Platform). Default is FALSE, other possible value is TRUE.}
}
\value{
An BrAPI compliant database connection object of class brapi_db.
}
\description{
Create a BrAPI connection object, which contains all parameters to connect to
a BrAPI compliant database via internet.
}
\details{
If the brapiDb argument is used to specify a connection object as created by the \code{as.brapi_db()} function or from the white list in the \code{brapi_db()} function, it will override the other parameters; except for user, password and token.
}
\seealso{
Other brapiutils: 
\code{\link{as.brapi_db}()},
\code{\link{brapi_canInternet}()},
\code{\link{brapi_checkCon}()},
\code{\link{brapi_db}()},
\code{\link{brapi_login}()},
\code{\link{brapi_serverinfo_show}()},
\code{\link{is.brapi_con}()},
\code{\link{print.brapi_con}()},
\code{\link{print.brapi_db}()}
}
\author{
brapir generator package
}
\concept{brapiutils}
