% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_db.R
\name{brapi_db}
\alias{brapi_db}
\title{brapi_db}
\usage{
brapi_db()
}
\value{
An BrAPI compliant database connection object of class brapi_db.
}
\description{
A white list of known BrAPI compliant databases.
}
\details{
Extended the function call with $database_name. White listed values for
database_name are: apiary, bms_test, cassavabase, cropontology, eu_sol,
genesys, germinate, icrisat_prod, imagebreed, mgis, musabase, ricebase,
snpseek, sweetpotatobase, terraref, testserver, t3s, tto, ttw, yambase.
}
\seealso{
Other brapiutils: 
\code{\link{as.brapi_db}()},
\code{\link{brapi_canInternet}()},
\code{\link{brapi_checkCon}()},
\code{\link{brapi_connect}()},
\code{\link{brapi_imagebreed_login}()},
\code{\link{brapi_imagebreed_logout}()},
\code{\link{brapi_login}()},
\code{\link{brapi_serverinfo_show}()},
\code{\link{is.brapi_con}()},
\code{\link{print.brapi_con}()},
\code{\link{print.brapi_db}()}
}
\author{
brapir generator package
}
\concept{brapiutils}
