% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_callsets.R
\name{brapi_get_callsets}
\alias{brapi_get_callsets}
\title{get /callsets}
\usage{
brapi_get_callsets(
  con = NULL,
  callSetDbId = "",
  callSetName = "",
  variantSetDbId = "",
  sampleDbId = "",
  germplasmDbId = "",
  page = 0,
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{callSetDbId}{character; required: FALSE; The unique database identifier
of the \code{CallSet} to be retrieved.}

\item{callSetName}{character; required: FALSE; The human readable name of the
\code{CallSet} to be retrieved.}

\item{variantSetDbId}{character; required: FALSE; The unique database
identifier of the \code{VariantSet} to be retrieved.}

\item{sampleDbId}{character; required: FALSE; The unique sample database
identifier of the \code{VariantSet} to be retrieved.}

\item{germplasmDbId}{character; required: FALSE; Return only call sets
generated from the Sample of this unique germplasm database identifier.}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is
\code{page = 0}). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Gets a filtered list of \code{CallSet} JSON objects.
}
\details{
Gets a filtered list of \code{CallSet} JSON objects.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_callsets(con = con)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.0#/Call\%20Sets/get_callsets }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-genotyping: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets_callSetDbId}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_search_calls_searchResultsDbId}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_get_search_variants_searchResultsDbId}()},
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId_calls}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variantsets_variantSetDbId_callsets}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets_variantSetDbId}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_search_callsets}()},
\code{\link{brapi_post_search_calls}()},
\code{\link{brapi_post_search_variantsets}()},
\code{\link{brapi_post_search_variants}()},
\code{\link{brapi_post_variantsets_extract}()}

Other Call Sets: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets_callSetDbId}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_post_search_callsets}()}
}
\author{
Maikel Verouden
}
\concept{Call Sets}
\concept{brapi-genotyping}
