% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_callsets_callSetDbId.R
\name{brapi_get_callsets_callSetDbId}
\alias{brapi_get_callsets_callSetDbId}
\title{get /callsets/\{callSetDbId\}}
\usage{
brapi_get_callsets_callSetDbId(con = NULL, callSetDbId = "")
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{callSetDbId}{character; required: TRUE; The unique database identifier
of the \code{CallSet} to be retrieved.}
}
\value{
data.frame
}
\description{
Gets a \code{CallSet} by ID.
}
\details{
Gets a \code{CallSet} by database identifier.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_callsets_callSetDbId(con = con,
                               callSetDbId = "callset01")
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.0#/Call\%20Sets/get_callsets__callSetDbId_ }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-genotyping: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_references_referenceDbId}()},
\code{\link{brapi_get_references}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_search_calls_searchResultsDbId}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_variants_searchResultsDbId}()},
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId_calls}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variantsets_variantSetDbId_callsets}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets_variantSetDbId}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_samples}()},
\code{\link{brapi_post_search_callsets}()},
\code{\link{brapi_post_search_calls}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_variantsets}()},
\code{\link{brapi_post_search_variants}()},
\code{\link{brapi_post_variantsets_extract}()},
\code{\link{brapi_put_samples_sampleDbId}()}

Other Call Sets: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_post_search_callsets}()}
}
\author{
Maikel Verouden
}
\concept{Call Sets}
\concept{brapi-genotyping}
