% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_commoncropnames.R
\name{brapi_get_commoncropnames}
\alias{brapi_get_commoncropnames}
\title{get /commoncropnames}
\usage{
brapi_get_commoncropnames(con = NULL, page = 0, pageSize = 1000)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is 'page'= 0).
Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be returned.
Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Get the Common Crop Names
}
\details{
List the common crop names for the crops available in a database server.

This call is \strong{required} for multi-crop systems where data from multiple
crops may be stored in the same database server. A distinct database server
is defined by everything in the URL before "/brapi/v2", including host
name and base path.

This call is recommended for single crop systems to be compatible with
multi-crop clients. For a single crop system the response should contain
an array with exactly 1 element.

The common crop name can be used as a search parameter for Programs,
Studies, and Germplasm.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_commoncropnames(con = con)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Core/2.0#/Common\%20Crop\%20Names/get_commoncropnames}{BrAPI SwaggerHub}
}
\seealso{
Other brapi-core: 
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations_locationDbId}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_programs_programDbId}()},
\code{\link{brapi_get_programs}()},
\code{\link{brapi_get_search_lists_searchResultsDbId}()},
\code{\link{brapi_get_search_locations_searchResultsDbId}()},
\code{\link{brapi_get_search_people_searchResultsDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_trials_searchResultsDbId}()},
\code{\link{brapi_get_seasons_seasonDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_serverinfo}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_get_trials_trialDbId}()},
\code{\link{brapi_get_trials}()},
\code{\link{brapi_post_lists_listDbId_items}()},
\code{\link{brapi_post_lists}()},
\code{\link{brapi_post_locations}()},
\code{\link{brapi_post_people}()}
}
\author{
Maikel Verouden
}
\concept{Common Crop Names}
\concept{brapi-core}
