% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_events.R
\name{brapi_get_events}
\alias{brapi_get_events}
\title{get /events}
\usage{
brapi_get_events(
  con = NULL,
  studyDbId = "",
  observationUnitDbId = "",
  eventDbId = "",
  eventType = "",
  dateRangeStart = "",
  dateRangeEnd = "",
  page = 0,
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{studyDbId}{character; required: FALSE; Filter based on an unique study
identifier, in which the events occurred.}

\item{observationUnitDbId}{character; required: FALSE; Filter based on an
unique observation unit identifier, in which the events occurred.}

\item{eventDbId}{character; required: FALSE; Filter based on an unique event
identifier.}

\item{eventType}{character; required: FALSE; Filter based on an event type.}

\item{dateRangeStart}{character; required: FALSE; Start date for filtering
based on a date range. Coded in the ISO 8601 standard extended format,
where date, time and time zone information needs to be provided (check for
example \url{https://www.w3.org/TR/NOTE-datetime}
).}

\item{dateRangeEnd}{character; required: FALSE; End date for filtering based
on a date range. Coded in the ISO 8601 standard extended format, where
date, time and time zone information needs to be provided (check for
example \url{https://www.w3.org/TR/NOTE-datetime}
).}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is
\code{page = 0}). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Get the Events
}
\details{
Get list of events
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_events(con = con)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Phenotyping/2.0#/Events/get_events }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-phenotyping: 
\code{\link{brapi_get_images_imageDbId}()},
\code{\link{brapi_get_images}()},
\code{\link{brapi_get_methods_methodDbId}()},
\code{\link{brapi_get_methods}()},
\code{\link{brapi_get_observationlevels}()},
\code{\link{brapi_get_observations_observationDbId}()},
\code{\link{brapi_get_observations}()},
\code{\link{brapi_get_observationunits_observationUnitDbId}()},
\code{\link{brapi_get_observationunits}()},
\code{\link{brapi_get_ontologies}()},
\code{\link{brapi_get_scales_scaleDbId}()},
\code{\link{brapi_get_scales}()},
\code{\link{brapi_get_traits_traitDbId}()},
\code{\link{brapi_get_traits}()},
\code{\link{brapi_get_variables_observationVariableDbId}()},
\code{\link{brapi_get_variables}()},
\code{\link{brapi_post_observations}()}
}
\author{
Maikel Verouden
}
\concept{Events}
\concept{brapi-phenotyping}
