% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_germplasm.R
\name{brapi_get_germplasm}
\alias{brapi_get_germplasm}
\title{get /germplasm}
\usage{
brapi_get_germplasm(
  con = NULL,
  germplasmPUI = "",
  germplasmDbId = "",
  germplasmName = "",
  commonCropName = "",
  accessionNumber = "",
  collection = "",
  genus = "",
  species = "",
  studyDbId = "",
  synonym = "",
  parentDbId = "",
  progenyDbId = "",
  externalReferenceID = "",
  externalReferenceSource = "",
  page = 0,
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{germplasmPUI}{character; required: FALSE; Permanent unique identifier
(DOI, URI, \emph{etc.}).}

\item{germplasmDbId}{character; required: FALSE; Unique internal germplasm
database identifier to filter by.}

\item{germplasmName}{character; required: FALSE; Name of the germplasm}

\item{commonCropName}{character; required: FALSE; The common crop name
related to this germplasm.}

\item{accessionNumber}{character; required: FALSE; Unique identifiers for
accessions within a genebank.}

\item{collection}{character; required: FALSE; A specific
panel/collection/population name this germplasm belongs to.}

\item{genus}{character; required: FALSE; Genus name to identify germplasm.}

\item{species}{character; required: FALSE; Species name to identify germplasm.}

\item{studyDbId}{character; required: FALSE; Search for Germplasm that are
associated with a particular unique internal Study database identifier.}

\item{synonym}{character; required: FALSE; Alternative name or ID used to
reference this germplasm.}

\item{parentDbId}{character; required: FALSE; Search for Germplasm with
this parent.}

\item{progenyDbId}{character; required: FALSE; Search for Germplasm with this
child.}

\item{externalReferenceID}{character; required: FALSE; An external reference
ID. Could be a simple string or a URI. (use with \code{externalReferenceSource}
parameter)}

\item{externalReferenceSource}{character; required: FALSE; An identifier for
the source system or database of an external reference (use with
\code{externalReferenceID} parameter).}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is
\code{page = 0}). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Get a filtered list of Germplasm
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_germplasm(con = con)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Germplasm/2.0#/Germplasm/get_germplasm }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-germplasm: 
\code{\link{brapi_get_attributes}()},
\code{\link{brapi_get_attributevalues}()},
\code{\link{brapi_get_search_attributes_searchResultsDbId}()},
\code{\link{brapi_get_search_germplasm_searchResultsDbId}()},
\code{\link{brapi_post_search_attributes}()},
\code{\link{brapi_post_search_germplasm}()}

Other Germplasm: 
\code{\link{brapi_get_search_germplasm_searchResultsDbId}()},
\code{\link{brapi_post_search_germplasm}()}
}
\author{
Maikel Verouden
}
\concept{Germplasm}
\concept{brapi-germplasm}
