% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_images.R
\name{brapi_get_images}
\alias{brapi_get_images}
\title{get /images}
\usage{
brapi_get_images(
  con = NULL,
  imageDbId = "",
  imageName = "",
  observationUnitDbId = "",
  observationDbId = "",
  descriptiveOntologyTerm = "",
  externalReferenceID = "",
  externalReferenceSource = "",
  page = 0,
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{imageDbId}{character; required: FALSE; Unique identifier for an image.}

\item{imageName}{character; required: FALSE; Human readable name of an image.}

\item{observationUnitDbId}{character; required: FALSE; The unique identifier
of the observation unit an image is portraying.}

\item{observationDbId}{character; required: FALSE; The unique identifier of
the observation an image is associated with.}

\item{descriptiveOntologyTerm}{character; required: FALSE; A descriptive term
associated with an image. It can be thought of as Tags for the image.
These could be simple descriptive words, or ontology references, or full
ontology URI's.}

\item{externalReferenceID}{character; required: FALSE; An external reference
ID. Could be a simple string or a URI. (use with \code{externalReferenceSource}
parameter)}

\item{externalReferenceSource}{character; required: FALSE; An identifier for
the source system or database of an external reference (use with
\code{externalReferenceID} parameter)}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is
\code{page = 0}). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Get the image meta data summaries
}
\details{
Get filtered set of image meta data
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_images(con = con)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Phenotyping/2.0#/Images/get_images }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-phenotyping: 
\code{\link{brapi_get_events}()},
\code{\link{brapi_get_images_imageDbId}()},
\code{\link{brapi_get_methods_methodDbId}()},
\code{\link{brapi_get_methods}()},
\code{\link{brapi_get_observationlevels}()},
\code{\link{brapi_get_observations_observationDbId}()},
\code{\link{brapi_get_observations}()},
\code{\link{brapi_get_observationunits_observationUnitDbId}()},
\code{\link{brapi_get_observationunits}()},
\code{\link{brapi_get_ontologies}()},
\code{\link{brapi_get_scales_scaleDbId}()},
\code{\link{brapi_get_scales}()},
\code{\link{brapi_get_traits_traitDbId}()},
\code{\link{brapi_get_traits}()},
\code{\link{brapi_get_variables_observationVariableDbId}()},
\code{\link{brapi_get_variables}()},
\code{\link{brapi_post_methods}()},
\code{\link{brapi_post_observations}()},
\code{\link{brapi_post_observationunits}()},
\code{\link{brapi_post_traits}()},
\code{\link{brapi_put_methods_methodDbId}()},
\code{\link{brapi_put_observations_observationDbId}()},
\code{\link{brapi_put_observationunits_observationUnitDbId}()},
\code{\link{brapi_put_traits_traitDbId}()}

Other Images: 
\code{\link{brapi_get_images_imageDbId}()}
}
\author{
Maikel Verouden
}
\concept{Images}
\concept{brapi-phenotyping}
