% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_lists_listDbId.R
\name{brapi_get_lists_listDbId}
\alias{brapi_get_lists_listDbId}
\title{get /lists/\{listDbId\}}
\usage{
brapi_get_lists_listDbId(con = NULL, listDbId = "")
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{listDbId}{character; required: TRUE; The unique ID of this generic list}
}
\value{
data.frame
}
\description{
Get the details of a specific List
}
\details{
Get a specific generic lists
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_lists_listDbId(con = con,
                         listDbId = "list1")
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Core/2.0#/Lists/get_lists__listDbId_}{BrAPI SwaggerHub}
}
\seealso{
Other brapi-core: 
\code{\link{brapi_get_commoncropnames}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations_locationDbId}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_programs_programDbId}()},
\code{\link{brapi_get_programs}()},
\code{\link{brapi_get_search_lists_searchResultsDbId}()},
\code{\link{brapi_get_search_locations_searchResultsDbId}()},
\code{\link{brapi_get_search_people_searchResultsDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_trials_searchResultsDbId}()},
\code{\link{brapi_get_seasons_seasonDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_serverinfo}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_get_trials_trialDbId}()},
\code{\link{brapi_get_trials}()},
\code{\link{brapi_post_lists_listDbId_items}()},
\code{\link{brapi_post_lists}()},
\code{\link{brapi_post_locations}()},
\code{\link{brapi_post_people}()}

Other Lists: 
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_search_lists_searchResultsDbId}()},
\code{\link{brapi_post_lists_listDbId_items}()},
\code{\link{brapi_post_lists}()}
}
\author{
Maikel Verouden
}
\concept{Lists}
\concept{brapi-core}
