% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_locations_locationDbId.R
\name{brapi_get_locations_locationDbId}
\alias{brapi_get_locations_locationDbId}
\title{get /locations/\{locationDbId\}}
\usage{
brapi_get_locations_locationDbId(con = NULL, locationDbId = "")
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{locationDbId}{character; required: TRUE; The internal DB id for a
location}
}
\value{
data.frame
}
\description{
Get the details of a specific Location
}
\details{
Get details for a location.
\itemize{
\item The \code{countryCode} is as per
\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}{ISO_3166-1_alpha-3}
specification.
\item coordinates are specified as longitude, latitude (in this specific order).
\item \code{altitude} (optional third parameter to coordinates) is in meters.
}
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_locations_locationDbId(con = con,
                                 locationDbId = "location_01")
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Core/2.0#/Locations/get_locations__locationDbId_}{BrAPI SwaggerHub}
}
\seealso{
Other brapi-core: 
\code{\link{brapi_get_commoncropnames}()},
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_programs_programDbId}()},
\code{\link{brapi_get_programs}()},
\code{\link{brapi_get_search_lists_searchResultsDbId}()},
\code{\link{brapi_get_search_locations_searchResultsDbId}()},
\code{\link{brapi_get_search_people_searchResultsDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_trials_searchResultsDbId}()},
\code{\link{brapi_get_seasons_seasonDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_serverinfo}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_get_trials_trialDbId}()},
\code{\link{brapi_get_trials}()},
\code{\link{brapi_post_lists_listDbId_items}()},
\code{\link{brapi_post_lists}()},
\code{\link{brapi_post_locations}()},
\code{\link{brapi_post_people}()}

Other Locations: 
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_search_locations_searchResultsDbId}()},
\code{\link{brapi_post_locations}()}
}
\author{
Maikel Verouden
}
\concept{Locations}
\concept{brapi-core}
