% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_maps.R
\name{brapi_get_maps}
\alias{brapi_get_maps}
\title{get /maps}
\usage{
brapi_get_maps(
  con = NULL,
  commonCropName = "",
  mapDbId = "",
  mapPUI = "",
  scientificName = "",
  type = "",
  programDbId = "",
  trialDbId = "",
  studyDbId = "",
  page = 0,
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{commonCropName}{character; required: FALSE; The common name of the crop.}

\item{mapDbId}{character; required: FALSE; The unique database identifier for
a genomic map.}

\item{mapPUI}{character; required: FALSE; The DOI or other permanent
identifier for a genomic map.}

\item{scientificName}{character; required: FALSE; Full scientific binomial
format name. This includes Genus, Species, and Sub-species.}

\item{type}{character; required: FALSE; Type of map matches the supplied
string (case-sensitive, exact match) to filter on.}

\item{programDbId}{character; required: FALSE; Unique database identifier to
filter by program.}

\item{trialDbId}{character; required: FALSE; Unique database identifier to
filter by trial.}

\item{studyDbId}{character; required: FALSE; Unique database identifier to
filter by study.}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is
\code{page = 0}). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Get the Genomic Maps
}
\details{
Get list of maps
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_maps(con = con,
               type = "Physical Map")
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.0#/Genome\%20Maps/get_maps }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-genotyping: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets_callSetDbId}()},
\code{\link{brapi_get_callsets}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_maps_mapDbId_linkagegroups}()},
\code{\link{brapi_get_maps_mapDbId}()},
\code{\link{brapi_get_references_referenceDbId_bases}()},
\code{\link{brapi_get_references_referenceDbId}()},
\code{\link{brapi_get_referencesets_referenceSetDbId}()},
\code{\link{brapi_get_referencesets}()},
\code{\link{brapi_get_references}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_search_calls_searchResultsDbId}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_get_search_references_searchResultsDbId}()},
\code{\link{brapi_get_search_referencesets_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_variants_searchResultsDbId}()},
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId_calls}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variantsets_variantSetDbId_callsets}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets_variantSetDbId}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_samples}()},
\code{\link{brapi_post_search_callsets}()},
\code{\link{brapi_post_search_calls}()},
\code{\link{brapi_post_search_referencesets}()},
\code{\link{brapi_post_search_references}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_variantsets}()},
\code{\link{brapi_post_search_variants}()},
\code{\link{brapi_post_variantsets_extract}()},
\code{\link{brapi_put_samples_sampleDbId}()}

Other Genome Maps: 
\code{\link{brapi_get_maps_mapDbId_linkagegroups}()},
\code{\link{brapi_get_maps_mapDbId}()}
}
\author{
Maikel Verouden
}
\concept{Genome Maps}
\concept{brapi-genotyping}
