% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_methods_methodDbId.R
\name{brapi_get_methods_methodDbId}
\alias{brapi_get_methods_methodDbId}
\title{get /methods/\{methodDbId\}}
\usage{
brapi_get_methods_methodDbId(con = NULL, methodDbId = "")
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{methodDbId}{character; required: TRUE; Unique method identifier, of
which to retrieve the details.}
}
\value{
data.frame
}
\description{
Get the details for a specific Method
}
\details{
Retrieve details about a specific method
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_methods_methodDbId(con = con,
                             methodDbId = "method_variable1")
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Phenotyping/2.0#/Methods/get_methods__methodDbId_ }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-phenotyping: 
\code{\link{brapi_get_methods}()},
\code{\link{brapi_get_observations}()},
\code{\link{brapi_get_variables_observationVariableDbId}()},
\code{\link{brapi_get_variables}()},
\code{\link{brapi_post_observations}()}

Other Methods: 
\code{\link{brapi_get_methods}()}
}
\author{
Maikel Verouden
}
\concept{Methods}
\concept{brapi-phenotyping}
