% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/brapi_get_observationunits_observationUnitDbId.R
\name{brapi_get_observationunits_observationUnitDbId}
\alias{brapi_get_observationunits_observationUnitDbId}
\title{get /observationunits/\{observationUnitDbId\}}
\usage{
brapi_get_observationunits_observationUnitDbId(
  con = NULL,
  observationUnitDbId = ""
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{observationUnitDbId}{character; required: TRUE; The unique identifier
of the specific observation unit.}
}
\value{
data.frame
}
\description{
Get the details of a specific Observation Unit
}
\details{
Get the details of a specific Observation Unit
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_observationunits_observationUnitDbId(con = con,
                                               observationUnitDbId = "observation_unit1")
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Phenotyping/2.0#/Observation\%20Units/get_observationunits__observationUnitDbId_ }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-phenotyping: 
\code{\link{brapi_get_events}()},
\code{\link{brapi_get_images_imageDbId}()},
\code{\link{brapi_get_images}()},
\code{\link{brapi_get_methods_methodDbId}()},
\code{\link{brapi_get_methods}()},
\code{\link{brapi_get_observationlevels}()},
\code{\link{brapi_get_observations_observationDbId}()},
\code{\link{brapi_get_observations}()},
\code{\link{brapi_get_observationunits}()},
\code{\link{brapi_get_ontologies}()},
\code{\link{brapi_get_scales_scaleDbId}()},
\code{\link{brapi_get_scales}()},
\code{\link{brapi_get_traits_traitDbId}()},
\code{\link{brapi_get_traits}()},
\code{\link{brapi_get_variables_observationVariableDbId}()},
\code{\link{brapi_get_variables}()},
\code{\link{brapi_post_methods}()},
\code{\link{brapi_post_observations}()},
\code{\link{brapi_post_observationunits}()},
\code{\link{brapi_put_methods_methodDbId}()},
\code{\link{brapi_put_observations_observationDbId}()},
\code{\link{brapi_put_observationunits_observationUnitDbId}()}

Other Observation Units: 
\code{\link{brapi_get_observationlevels}()},
\code{\link{brapi_get_observationunits}()},
\code{\link{brapi_post_observationunits}()},
\code{\link{brapi_put_observationunits_observationUnitDbId}()}
}
\author{
Maikel Verouden
}
\concept{Observation Units}
\concept{brapi-phenotyping}
