% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_observationunits_table.R
\name{brapi_get_observationunits_table}
\alias{brapi_get_observationunits_table}
\title{get /observationunits/table}
\usage{
brapi_get_observationunits_table(
  con = NULL,
  Accept = "application/json",
  observationUnitDbId = "",
  germplasmDbId = "",
  observationVariableDbId = "",
  studyDbId = "",
  locationDbId = "",
  trialDbId = "",
  programDbId = "",
  seasonDbId = "",
  observationLevel = ""
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{Accept}{character; required: TRUE; The requested content type which
should be returned by the server; default: "application/json",
other possible values:
\itemize{
\item "text/csv"
\item "text/tsv"
}}

\item{observationUnitDbId}{character; required: FALSE; Unique observation
unit database identifier to filter on.}

\item{germplasmDbId}{character; required: FALSE; Unique germplasm (accession)
identifier to filter on.}

\item{observationVariableDbId}{character; required: FALSE; Unique observation
variable identifier to filter on.}

\item{studyDbId}{character; required: FALSE; Unique study database identifier
to filter on.}

\item{locationDbId}{character; required: FALSE; Unique location identifier,
where the observations were collected.}

\item{trialDbId}{character; required: FALSE; Unique trial identifier to
filter on}

\item{programDbId}{character; required: FALSE; Unique program database
identifier to filter on.}

\item{seasonDbId}{character; required: TRUE; The unique database identifier
for a season, e.g. the year or Phenotyping campaign of a multi-annual
study (trees, grape, ...). For backward compatibility it can be a string
like '2012', '1957-2004'.}

\item{observationLevel}{character; required: FALSE; The type of the
observationUnit. Returns only the observation unit of the specified type;
the parent levels ID can be accessed through \code{observationUnitStructure}.}
}
\value{
data.frame
}
\description{
Get a list of Observations in a table format
}
\details{
This service is designed to retrieve a table for observation values
as a matrix of Observation Units and Observation Variables.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_observationunits_table(con = con,
                                 Accept = "application/json")
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Phenotyping/2.0#/Observation\%20Units/get_observationunits_table}{BrAPI SwaggerHub}
}
\seealso{
Other brapi-phenotyping: 
\code{\link{brapi_get_events}()},
\code{\link{brapi_get_images_imageDbId}()},
\code{\link{brapi_get_images}()},
\code{\link{brapi_get_methods_methodDbId}()},
\code{\link{brapi_get_methods}()},
\code{\link{brapi_get_observationlevels}()},
\code{\link{brapi_get_observations_observationDbId}()},
\code{\link{brapi_get_observations_table}()},
\code{\link{brapi_get_observations}()},
\code{\link{brapi_get_observationunits_observationUnitDbId}()},
\code{\link{brapi_get_observationunits}()},
\code{\link{brapi_get_ontologies}()},
\code{\link{brapi_get_scales_scaleDbId}()},
\code{\link{brapi_get_scales}()},
\code{\link{brapi_get_search_images_searchResultsDbId}()},
\code{\link{brapi_get_search_observations_searchResultsDbId}()},
\code{\link{brapi_get_search_observationunits_searchResultsDbId}()},
\code{\link{brapi_get_search_variables_searchResultsDbId}()},
\code{\link{brapi_get_traits_traitDbId}()},
\code{\link{brapi_get_traits}()},
\code{\link{brapi_get_variables_observationVariableDbId}()},
\code{\link{brapi_get_variables}()},
\code{\link{brapi_post_images}()},
\code{\link{brapi_post_methods}()},
\code{\link{brapi_post_observations}()},
\code{\link{brapi_post_observationunits}()},
\code{\link{brapi_post_scales}()},
\code{\link{brapi_post_search_images}()},
\code{\link{brapi_post_search_observations}()},
\code{\link{brapi_post_search_observationunits}()},
\code{\link{brapi_post_search_variables}()},
\code{\link{brapi_post_traits}()},
\code{\link{brapi_post_variables}()},
\code{\link{brapi_put_images_imageDbId_imagecontent}()},
\code{\link{brapi_put_images_imageDbId}()},
\code{\link{brapi_put_methods_methodDbId}()},
\code{\link{brapi_put_observations_observationDbId}()},
\code{\link{brapi_put_observationunits_observationUnitDbId}()},
\code{\link{brapi_put_scales_scaleDbId}()},
\code{\link{brapi_put_traits_traitDbId}()},
\code{\link{brapi_put_variables_observationVariableDbId}()}

Other Observation Units: 
\code{\link{brapi_get_observationlevels}()},
\code{\link{brapi_get_observationunits_observationUnitDbId}()},
\code{\link{brapi_get_observationunits}()},
\code{\link{brapi_get_search_observationunits_searchResultsDbId}()},
\code{\link{brapi_post_observationunits}()},
\code{\link{brapi_post_search_observationunits}()},
\code{\link{brapi_put_observationunits_observationUnitDbId}()}
}
\author{
Maikel Verouden
}
\concept{Observation Units}
\concept{brapi-phenotyping}
