% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_ontologies.R
\name{brapi_get_ontologies}
\alias{brapi_get_ontologies}
\title{get /ontologies}
\usage{
brapi_get_ontologies(con = NULL, ontologyDbId = "", page = 0, pageSize = 1000)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{ontologyDbId}{character; required: FALSE; The unique identifier for an
ontology definition.}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is
\code{page = 0}). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Get the Ontologies
}
\details{
Call to retrieve observation variable ontologies available in the
database system.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_ontologies(con = con)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Phenotyping/2.0#/Ontologies/get_ontologies }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-phenotyping: 
\code{\link{brapi_get_events}()},
\code{\link{brapi_get_methods_methodDbId}()},
\code{\link{brapi_get_methods}()},
\code{\link{brapi_get_observations}()},
\code{\link{brapi_get_scales_scaleDbId}()},
\code{\link{brapi_get_scales}()},
\code{\link{brapi_get_traits_traitDbId}()},
\code{\link{brapi_get_traits}()},
\code{\link{brapi_get_variables_observationVariableDbId}()},
\code{\link{brapi_get_variables}()},
\code{\link{brapi_post_observations}()}
}
\author{
Maikel Verouden
}
\concept{Ontologies}
\concept{brapi-phenotyping}
