% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_people_personDbId.R
\name{brapi_get_people_personDbId}
\alias{brapi_get_people_personDbId}
\title{get /people/\{personDbId\}}
\usage{
brapi_get_people_personDbId(con = NULL, personDbId = "")
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{personDbId}{character; required: TRUE; The unique ID of a person}
}
\value{
data.frame
}
\description{
Get the details for a specific Person
}
\details{
Get the details for a specific Person
}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Core/2.0#/People/get_people__personDbId_ }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-core: 
\code{\link{brapi_get_commoncropnames}()},
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations_locationDbId}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_people}()}

Other People: 
\code{\link{brapi_get_people}()}
}
\author{
Maikel Verouden
}
\concept{People}
\concept{brapi-core}
