% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_programs.R
\name{brapi_get_programs}
\alias{brapi_get_programs}
\title{get /programs}
\usage{
brapi_get_programs(
  con = NULL,
  commonCropName = "",
  programDbId = "",
  programName = "",
  abbreviation = "",
  externalReferenceID = "",
  externalReferenceSource = "",
  page = 0,
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{commonCropName}{character; required: FALSE; Filter by the common crop
name. Exact match.}

\item{programDbId}{character; required: FALSE; Program filter to only return
trials associated with given program id. Exact match.}

\item{programName}{character; required: FALSE; Filter by program name.
Exact match.}

\item{abbreviation}{character; required: FALSE; Filter by program
abbreviation. Exact match.}

\item{externalReferenceID}{character; required: FALSE; An external reference
ID. Could be a simple string or a URI. (use with \code{externalReferenceSource}
parameter)}

\item{externalReferenceSource}{character; required: FALSE; An identifier for
the source system or database of an external reference (use with
\code{externalReferenceID} parameter)}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is 'page'=
0). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Get a filtered list of breeding Programs
}
\details{
Get a filtered list of breeding Programs. This list can be filtered
by common crop name to narrow results to a specific crop.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_programs(con = con)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Core/2.0#/Programs/get_programs }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-core: 
\code{\link{brapi_get_commoncropnames}()},
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations_locationDbId}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_programs_programDbId}()},
\code{\link{brapi_get_search_lists_searchResultsDbId}()},
\code{\link{brapi_get_search_locations_searchResultsDbId}()},
\code{\link{brapi_get_search_people_searchResultsDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_trials_searchResultsDbId}()},
\code{\link{brapi_get_seasons_seasonDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_serverinfo}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_get_trials_trialDbId}()},
\code{\link{brapi_get_trials}()},
\code{\link{brapi_post_lists_listDbId_items}()},
\code{\link{brapi_post_lists}()},
\code{\link{brapi_post_locations}()},
\code{\link{brapi_post_people}()}

Other Programs: 
\code{\link{brapi_get_programs_programDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()}
}
\author{
Maikel Verouden
}
\concept{Programs}
\concept{brapi-core}
