% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_references.R
\name{brapi_get_references}
\alias{brapi_get_references}
\title{get /references}
\usage{
brapi_get_references(
  con = NULL,
  referenceDbId = "",
  referenceSetDbId = "",
  accession = "",
  md5checksum = "",
  isDerived = NA,
  minLength = as.integer(NA),
  maxLength = as.integer(NA),
  page = 0,
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{referenceDbId}{character; required: FALSE; The unique database
identifier of the \code{Reference} to be retrieved.}

\item{referenceSetDbId}{character; required: FALSE; The unique database
identifier of the \code{ReferenceSet} to be retrieved.}

\item{accession}{character; required: FALSE; If set, return the reference
sets for which the \code{accession} matches this string (case-sensitive,
exact match).}

\item{md5checksum}{character; required: FALSE; If specified, return the
references for which the \code{md5checksum} matches this string (case-sensitive
, exact match).}

\item{isDerived}{logical; required: FALSE; If the reference is derived (TRUE) from a
source sequence or not (FALSE). If unspecified (NA), both will be
displayed.}

\item{minLength}{integer; required: FALSE; The minimum length of the
reference sequences to be retrieved.}

\item{maxLength}{integer; required: FALSE; The maximum length of the
reference sequences to be retrieved.}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is
\code{page = 0}). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Gets a filtered list of \code{Reference} objects.
}
\details{
\code{GET /references} will return a filtered data.frame of \code{Reference}
objects.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_references(con = con)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.0#/References/get_references }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-genotyping: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets_callSetDbId}()},
\code{\link{brapi_get_callsets}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_references_referenceDbId_bases}()},
\code{\link{brapi_get_references_referenceDbId}()},
\code{\link{brapi_get_referencesets}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_search_calls_searchResultsDbId}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_get_search_references_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_variants_searchResultsDbId}()},
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId_calls}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variantsets_variantSetDbId_callsets}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets_variantSetDbId}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_samples}()},
\code{\link{brapi_post_search_callsets}()},
\code{\link{brapi_post_search_calls}()},
\code{\link{brapi_post_search_references}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_variantsets}()},
\code{\link{brapi_post_search_variants}()},
\code{\link{brapi_post_variantsets_extract}()},
\code{\link{brapi_put_samples_sampleDbId}()}

Other References: 
\code{\link{brapi_get_references_referenceDbId_bases}()},
\code{\link{brapi_get_references_referenceDbId}()},
\code{\link{brapi_get_search_references_searchResultsDbId}()},
\code{\link{brapi_post_search_references}()}
}
\author{
Maikel Verouden
}
\concept{References}
\concept{brapi-genotyping}
