% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_references_referenceDbId.R
\name{brapi_get_references_referenceDbId}
\alias{brapi_get_references_referenceDbId}
\title{get /references/\{referenceDbId\}}
\usage{
brapi_get_references_referenceDbId(con = NULL, referenceDbId = "")
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{referenceDbId}{character; required: TRUE; The unique identifier of the
\code{Reference} to be retrieved.}
}
\value{
data.frame
}
\description{
Gets a \code{Reference} by ID.
}
\details{
Retrieve details for a reference specified by its \code{referenceDbId}.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_references_referenceDbId(con = con,
                                   referenceDbId = "reference1")
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.0#/References/get_references__referenceDbId_ }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-genotyping: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets_callSetDbId}()},
\code{\link{brapi_get_callsets}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_references_referenceDbId_bases}()},
\code{\link{brapi_get_references}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_search_calls_searchResultsDbId}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_get_search_references_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_variants_searchResultsDbId}()},
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId_calls}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variantsets_variantSetDbId_callsets}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets_variantSetDbId}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_samples}()},
\code{\link{brapi_post_search_callsets}()},
\code{\link{brapi_post_search_calls}()},
\code{\link{brapi_post_search_references}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_variantsets}()},
\code{\link{brapi_post_search_variants}()},
\code{\link{brapi_post_variantsets_extract}()},
\code{\link{brapi_put_samples_sampleDbId}()}

Other References: 
\code{\link{brapi_get_references_referenceDbId_bases}()},
\code{\link{brapi_get_references}()},
\code{\link{brapi_get_search_references_searchResultsDbId}()},
\code{\link{brapi_post_search_references}()}
}
\author{
Maikel Verouden
}
\concept{References}
\concept{brapi-genotyping}
