% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_references_referenceDbId_bases.R
\name{brapi_get_references_referenceDbId_bases}
\alias{brapi_get_references_referenceDbId_bases}
\title{get /references/\{referenceDbId\}/bases}
\usage{
brapi_get_references_referenceDbId_bases(
  con = NULL,
  referenceDbId = "",
  start = as.integer(NA),
  end = as.integer(NA),
  pageToken = ""
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{referenceDbId}{character; required: TRUE; The unique identifier of the
\code{Reference} to be retrieved.}

\item{start}{integer; required: FALSE; The start position (0-based) of this
query. Defaults to 0. Genomic positions are non-negative integers less
than reference length. Requests spanning the join of circular genomes are
represented as two requests one on each side of the join (position 0).}

\item{end}{integer; required: FALSE; The end position (0-based, exclusive) of
this query. Defaults to the length of this \code{Reference}.}

\item{pageToken}{character; required: FALSE; Used to request a specific page
of data to be returned. Tokenized pages are for large data sets, which can
not be efficiently broken into indexed pages. Use the \code{nextPageToken} and
\code{prevPageToken} from a prior response to construct a query and move to the
next or previous page respectively.}
}
\value{
data.frame
}
\description{
Lists \code{Reference} bases by database identifier and optional range.
}
\details{
Lists \code{Reference} bases by ID and optional range.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_references_referenceDbId_bases(con = con,
                                         referenceDbId = "reference1")
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.0#/References/get_references__referenceDbId__bases }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-genotyping: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets_callSetDbId}()},
\code{\link{brapi_get_callsets}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_maps_mapDbId_linkagegroups}()},
\code{\link{brapi_get_maps_mapDbId}()},
\code{\link{brapi_get_maps}()},
\code{\link{brapi_get_markerpositions}()},
\code{\link{brapi_get_references_referenceDbId}()},
\code{\link{brapi_get_referencesets_referenceSetDbId}()},
\code{\link{brapi_get_referencesets}()},
\code{\link{brapi_get_references}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_search_calls_searchResultsDbId}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_get_search_markerpositions_searchResultsDbId}()},
\code{\link{brapi_get_search_references_searchResultsDbId}()},
\code{\link{brapi_get_search_referencesets_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_variants_searchResultsDbId}()},
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId_calls}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variantsets_variantSetDbId_callsets}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets_variantSetDbId}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_get_vendor_orders}()},
\code{\link{brapi_post_samples}()},
\code{\link{brapi_post_search_callsets}()},
\code{\link{brapi_post_search_calls}()},
\code{\link{brapi_post_search_markerpositions}()},
\code{\link{brapi_post_search_referencesets}()},
\code{\link{brapi_post_search_references}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_variantsets}()},
\code{\link{brapi_post_search_variants}()},
\code{\link{brapi_post_variantsets_extract}()},
\code{\link{brapi_put_samples_sampleDbId}()}

Other References: 
\code{\link{brapi_get_references_referenceDbId}()},
\code{\link{brapi_get_references}()},
\code{\link{brapi_get_search_references_searchResultsDbId}()},
\code{\link{brapi_post_search_references}()}
}
\author{
Maikel Verouden
}
\concept{References}
\concept{brapi-genotyping}
