% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_referencesets.R
\name{brapi_get_referencesets}
\alias{brapi_get_referencesets}
\title{get /referencesets}
\usage{
brapi_get_referencesets(
  con = NULL,
  referenceSetDbId = "",
  accession = "",
  assemblyPUI = "",
  md5checksum = "",
  page = 0,
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{referenceSetDbId}{character; required: FALSE; The unique database
identifier of the \code{ReferenceSet} to be retrieved.}

\item{accession}{character; required: FALSE; If set, return the reference
sets for which the \code{accession} matches this string (case-sensitive,
exact match).}

\item{assemblyPUI}{character; required: FALSE; If set, return the reference
sets for which the \code{assemblyId} matches this string (case-sensitive, exact
match).}

\item{md5checksum}{character; required: FALSE; If set, return the reference
sets for which the \code{md5checksum} matches this string (case-sensitive,
exact match).}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is
\code{page = 0}). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Gets a list of \code{ReferenceSets}.
}
\details{
Gets a filtered list of \code{ReferenceSets}.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_referencesets(con = con)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.0#/Reference\%20Sets/get_referencesets }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-genotyping: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets_callSetDbId}()},
\code{\link{brapi_get_callsets}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_references_referenceDbId_bases}()},
\code{\link{brapi_get_references_referenceDbId}()},
\code{\link{brapi_get_referencesets_referenceSetDbId}()},
\code{\link{brapi_get_references}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_search_calls_searchResultsDbId}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_get_search_references_searchResultsDbId}()},
\code{\link{brapi_get_search_referencesets_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_variants_searchResultsDbId}()},
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId_calls}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variantsets_variantSetDbId_callsets}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets_variantSetDbId}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_samples}()},
\code{\link{brapi_post_search_callsets}()},
\code{\link{brapi_post_search_calls}()},
\code{\link{brapi_post_search_referencesets}()},
\code{\link{brapi_post_search_references}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_variantsets}()},
\code{\link{brapi_post_search_variants}()},
\code{\link{brapi_post_variantsets_extract}()},
\code{\link{brapi_put_samples_sampleDbId}()}

Other Reference Sets: 
\code{\link{brapi_get_referencesets_referenceSetDbId}()},
\code{\link{brapi_get_search_referencesets_searchResultsDbId}()},
\code{\link{brapi_post_search_referencesets}()}
}
\author{
Maikel Verouden
}
\concept{Reference Sets}
\concept{brapi-genotyping}
