% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_samples.R
\name{brapi_get_samples}
\alias{brapi_get_samples}
\title{get /samples}
\usage{
brapi_get_samples(
  con = NULL,
  sampleDbId = "",
  observationUnitDbId = "",
  plateDbId = "",
  germplasmDbId = "",
  studyDbId = "",
  externalReferenceID = "",
  externalReferenceSource = "",
  page = 0,
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{sampleDbId}{character; required: FALSE; the unique internal database
identifier for a sample}

\item{observationUnitDbId}{character; required: FALSE; the unique internal
database identifier for an observation unit, where a sample was taken from.}

\item{plateDbId}{character; required: FALSE; the unique internal database
identifier for a plate of samples.}

\item{germplasmDbId}{character; required: FALSE; the unique internal database
identifier for a germplasm.}

\item{studyDbId}{character; required: FALSE; The unique internal database
study identifier to filter by.}

\item{externalReferenceID}{character; required: FALSE; An external reference
identifier. Could be a simple string or a URI (use with
\code{externalReferenceSource} parameter).}

\item{externalReferenceSource}{character; required: FALSE; An identifier for
the source system or database of an external reference (use with
\code{externalReferenceID} parameter).}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is
\code{page = 0}). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Get the Samples
}
\details{
Used to retrieve list of Samples from a Sample Tracking system based
on some search criteria.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_samples(con = con)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.0#/Samples/get_samples }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-genotyping: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets_callSetDbId}()},
\code{\link{brapi_get_callsets}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_references}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_search_calls_searchResultsDbId}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_variants_searchResultsDbId}()},
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId_calls}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variantsets_variantSetDbId_callsets}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets_variantSetDbId}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_samples}()},
\code{\link{brapi_post_search_callsets}()},
\code{\link{brapi_post_search_calls}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_variantsets}()},
\code{\link{brapi_post_search_variants}()},
\code{\link{brapi_post_variantsets_extract}()},
\code{\link{brapi_put_samples_sampleDbId}()}

Other Samples: 
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_post_samples}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_put_samples_sampleDbId}()}
}
\author{
Maikel Verouden
}
\concept{Samples}
\concept{brapi-genotyping}
