% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_samples_sampleDbId.R
\name{brapi_get_samples_sampleDbId}
\alias{brapi_get_samples_sampleDbId}
\title{get /samples/\{sampleDbId\}}
\usage{
brapi_get_samples_sampleDbId(con = NULL, sampleDbId = "")
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{sampleDbId}{character; required: TRUE; the unique internal database
identifier for a sample.}
}
\value{
data.frame
}
\description{
Get the details of a specific Sample
}
\details{
Used to retrieve the details of a single Sample from a Sample
Tracking system.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_samples_sampleDbId(con = con,
                             sampleDbId = "sample1")
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.0#/Samples/get_samples__sampleDbId_ }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-genotyping: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets_callSetDbId}()},
\code{\link{brapi_get_callsets}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_search_calls_searchResultsDbId}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_get_search_variants_searchResultsDbId}()},
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId_calls}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variantsets_variantSetDbId_callsets}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets_variantSetDbId}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_search_callsets}()},
\code{\link{brapi_post_search_calls}()},
\code{\link{brapi_post_search_variantsets}()},
\code{\link{brapi_post_search_variants}()},
\code{\link{brapi_post_variantsets_extract}()}

Other Samples: 
\code{\link{brapi_get_samples}()}
}
\author{
Maikel Verouden
}
\concept{Samples}
\concept{brapi-genotyping}
