% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_scales_scaleDbId.R
\name{brapi_get_scales_scaleDbId}
\alias{brapi_get_scales_scaleDbId}
\title{get /scales/\{scaleDbId\}}
\usage{
brapi_get_scales_scaleDbId(con = NULL, scaleDbId = "")
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{scaleDbId}{character; required: TRUE; Unique identifier of the scale to
retrieve details of.}
}
\value{
data.frame
}
\description{
Get the details of a specific Scale
}
\details{
Retrieve details about a specific scale
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_scales_scaleDbId(con = con,
                           scaleDbId = "scale_variable1")
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Phenotyping/2.0#/Scales/get_scales__scaleDbId_ }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-phenotyping: 
\code{\link{brapi_get_events}()},
\code{\link{brapi_get_images_imageDbId}()},
\code{\link{brapi_get_images}()},
\code{\link{brapi_get_methods_methodDbId}()},
\code{\link{brapi_get_methods}()},
\code{\link{brapi_get_observationlevels}()},
\code{\link{brapi_get_observations_observationDbId}()},
\code{\link{brapi_get_observations}()},
\code{\link{brapi_get_observationunits_observationUnitDbId}()},
\code{\link{brapi_get_observationunits}()},
\code{\link{brapi_get_ontologies}()},
\code{\link{brapi_get_scales}()},
\code{\link{brapi_get_traits_traitDbId}()},
\code{\link{brapi_get_traits}()},
\code{\link{brapi_get_variables_observationVariableDbId}()},
\code{\link{brapi_get_variables}()},
\code{\link{brapi_post_observations}()},
\code{\link{brapi_post_observationunits}()},
\code{\link{brapi_put_observations_observationDbId}()},
\code{\link{brapi_put_observationunits_observationUnitDbId}()}

Other Scales: 
\code{\link{brapi_get_scales}()}
}
\author{
Maikel Verouden
}
\concept{Scales}
\concept{brapi-phenotyping}
