% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/brapi_get_search_attributevalues_searchResultsDbId.R
\name{brapi_get_search_attributevalues_searchResultsDbId}
\alias{brapi_get_search_attributevalues_searchResultsDbId}
\title{get /search/attributevalues/\{searchResultsDbId\}}
\usage{
brapi_get_search_attributevalues_searchResultsDbId(
  con = NULL,
  searchResultsDbId = "",
  page = 0,
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{searchResultsDbId}{character; required: TRUE; Unique identifier which
references the search results}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is
\code{page = 0}). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Get the results of a Germplasm \code{AttributeValues} search request
}
\details{
Get the results of a Germplasm \code{AttributeValues} search request.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
con[["token"]] <- "YYYY"
# Saved or Asynchronous Search Response Example
out <-
  brapi_post_search_attributevalues(
      con = con,
      attributeDbIds = c("attribute1",
                         "attribute2"),
      attributeNames = "Rht-B1b",
      attributeValueDbIds = "attribute_val1",
      externalReferenceIDs = "https://brapi.org/specification",
      externalReferenceSources = "BrAPI Doc",
      page = 0,
      pageSize = 1000)
searchResultsDbId <- out$searchResultsDbId
brapi_get_search_attributevalues_searchResultsDbId(con = con, searchResultsDbId = searchResultsDbId)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Germplasm/2.0#/Germplasm\%20Attribute\%20Values/get_search_attributevalues__searchResultsDbId_}{BrAPI SwaggerHub}
}
\seealso{
Other brapi-germplasm: 
\code{\link{brapi_get_attributes}()},
\code{\link{brapi_get_attributevalues}()},
\code{\link{brapi_get_germplasm}()},
\code{\link{brapi_get_search_attributes_searchResultsDbId}()},
\code{\link{brapi_get_search_germplasm_searchResultsDbId}()},
\code{\link{brapi_post_search_attributes}()},
\code{\link{brapi_post_search_attributevalues}()},
\code{\link{brapi_post_search_germplasm}()}

Other Germplasm Attribute Values: 
\code{\link{brapi_get_attributevalues}()},
\code{\link{brapi_post_search_attributevalues}()}
}
\author{
Maikel Verouden
}
\concept{Germplasm Attribute Values}
\concept{brapi-germplasm}
