% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_search_calls_searchResultsDbId.R
\name{brapi_get_search_calls_searchResultsDbId}
\alias{brapi_get_search_calls_searchResultsDbId}
\title{get /search/calls/\{searchResultsDbId\}}
\usage{
brapi_get_search_calls_searchResultsDbId(
  con = NULL,
  searchResultsDbId = "",
  pageToken = "",
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{searchResultsDbId}{character; required: TRUE; Unique identifier which
references the search results}

\item{pageToken}{character; required: FALSE; Used to request a specific page
of data to be returned. Tokenized pages are for large data sets, which can
not be efficiently broken into indexed pages. Use the \code{nextPageToken} and
\code{prevPageToken} from a prior response to construct a query and move to the
next or previous page respectively.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Get the results of a \code{Calls} search request
}
\details{
Returns the result of the advanced searching for the calls resource.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
con[["token"]] <- "YYYY"
# Saved or Asynchronous Search Response Example
out <-
 brapi_post_search_calls(con = con,
                         callSetDbIds = c("callset12",
                                          "callset21"),
                         expandHomozygotes = TRUE,
                         pageSize = 1000,
                         sepPhased = "/",
                         sepUnphased = "|",
                         unknownString = "-",
                         variantDbIds = c("variant01",
                                          "variant23"))
searchResultsDbId <- out$searchResultsDbId
brapi_get_search_calls_searchResultsDbId(con = con,
                                         searchResultsDbId = searchResultsDbId)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.0#/Calls/get_search_calls__searchResultsDbId_ }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-genotyping: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets_callSetDbId}()},
\code{\link{brapi_get_callsets}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_search_callsets}()},
\code{\link{brapi_post_search_calls}()}

Other Calls: 
\code{\link{brapi_get_calls}()},
\code{\link{brapi_post_search_calls}()}
}
\author{
Maikel Verouden
}
\concept{Calls}
\concept{brapi-genotyping}
