% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/brapi_get_search_observations_searchResultsDbId.R
\name{brapi_get_search_observations_searchResultsDbId}
\alias{brapi_get_search_observations_searchResultsDbId}
\title{get /search/observations/\{searchResultsDbId\}}
\usage{
brapi_get_search_observations_searchResultsDbId(
  con = NULL,
  Accept = "application/json",
  searchResultsDbId = "",
  page = 0,
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{Accept}{character; required: TRUE; The requested content type which
should be returned by the server; default: "application/json",
other possible values:
\itemize{
\item "text/csv"
\item "text/tsv"
\item "application/flapjack"
}}

\item{searchResultsDbId}{character; required: TRUE; Unique database
identifier which references the search results}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is
\code{page = 0}). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Get the results of a \code{Observations} search request
}
\details{
Returns the result of the advanced searching for the
\code{Observations} resource.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
con[["token"]] <- "YYYY"
# Saved or Asynchronous Search Response Example
out <- brapi_post_search_observations(
  con = con,
  externalReferenceIDs = "https://brapi.org/specification",
  externalReferenceSources = "BrAPI Doc",
  germplasmDbIds = c("germplasm1",
                     "germplasm2",
                     "germplasm3"),
  page = 0,
  pageSize = 1000)
searchResultsDbId <- out$searchResultsDbId
brapi_get_search_observations_searchResultsDbId(
  con = con,
  searchResultsDbId = searchResultsDbId)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Phenotyping/2.0#/Observations/get_search_observations__searchResultsDbId_ }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-phenotyping: 
\code{\link{brapi_get_events}()},
\code{\link{brapi_get_images_imageDbId}()},
\code{\link{brapi_get_images}()},
\code{\link{brapi_get_methods_methodDbId}()},
\code{\link{brapi_get_methods}()},
\code{\link{brapi_get_observationlevels}()},
\code{\link{brapi_get_observations_observationDbId}()},
\code{\link{brapi_get_observations}()},
\code{\link{brapi_get_observationunits_observationUnitDbId}()},
\code{\link{brapi_get_observationunits}()},
\code{\link{brapi_get_ontologies}()},
\code{\link{brapi_get_scales_scaleDbId}()},
\code{\link{brapi_get_scales}()},
\code{\link{brapi_get_search_images_searchResultsDbId}()},
\code{\link{brapi_get_search_observationunits_searchResultsDbId}()},
\code{\link{brapi_get_search_variables_searchResultsDbId}()},
\code{\link{brapi_get_traits_traitDbId}()},
\code{\link{brapi_get_traits}()},
\code{\link{brapi_get_variables_observationVariableDbId}()},
\code{\link{brapi_get_variables}()},
\code{\link{brapi_post_images}()},
\code{\link{brapi_post_methods}()},
\code{\link{brapi_post_observations}()},
\code{\link{brapi_post_observationunits}()},
\code{\link{brapi_post_scales}()},
\code{\link{brapi_post_search_images}()},
\code{\link{brapi_post_search_observations}()},
\code{\link{brapi_post_search_observationunits}()},
\code{\link{brapi_post_search_variables}()},
\code{\link{brapi_post_traits}()},
\code{\link{brapi_post_variables}()},
\code{\link{brapi_put_images_imageDbId_imagecontent}()},
\code{\link{brapi_put_images_imageDbId}()},
\code{\link{brapi_put_methods_methodDbId}()},
\code{\link{brapi_put_observations_observationDbId}()},
\code{\link{brapi_put_observationunits_observationUnitDbId}()},
\code{\link{brapi_put_scales_scaleDbId}()},
\code{\link{brapi_put_traits_traitDbId}()},
\code{\link{brapi_put_variables_observationVariableDbId}()}

Other Observations: 
\code{\link{brapi_get_observations_observationDbId}()},
\code{\link{brapi_get_observations}()},
\code{\link{brapi_post_observations}()},
\code{\link{brapi_post_search_observations}()},
\code{\link{brapi_put_observations_observationDbId}()}
}
\author{
Maikel Verouden
}
\concept{Observations}
\concept{brapi-phenotyping}
