% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_search_programs_searchResultsDbId.R
\name{brapi_get_search_programs_searchResultsDbId}
\alias{brapi_get_search_programs_searchResultsDbId}
\title{get /search/programs/\{searchResultsDbId\}}
\usage{
brapi_get_search_programs_searchResultsDbId(
  con = NULL,
  searchResultsDbId = "",
  page = 0,
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{searchResultsDbId}{character; required: TRUE; Unique identifier which references the search results}

\item{page}{integer; required: FALSE; Used to request a specific page of data to be returned.
The page indexing starts at 0 (the first page is 'page'= 0). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Get the results of a Programs search request
}
\details{
Advanced searching for the programs resource.
See Search Services for additional implementation details.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_search_programs_searchResultsDbId(con = con, searchResultsDbId = "")
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Core/2.0#/Programs/get_search_programs__searchResultsDbId_ }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-core: 
\code{\link{brapi_get_commoncropnames}()},
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations_locationDbId}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_programs_programDbId}()},
\code{\link{brapi_get_programs}()},
\code{\link{brapi_get_search_lists_searchResultsDbId}()},
\code{\link{brapi_get_search_locations_searchResultsDbId}()},
\code{\link{brapi_get_search_people_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_trials_searchResultsDbId}()}

Other Programs: 
\code{\link{brapi_get_programs_programDbId}()},
\code{\link{brapi_get_programs}()}
}
\author{
Maikel Verouden
}
\concept{Programs}
\concept{brapi-core}
