% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_search_variants_searchResultsDbId.R
\name{brapi_get_search_variants_searchResultsDbId}
\alias{brapi_get_search_variants_searchResultsDbId}
\title{get /search/variants/\{searchResultsDbId\}}
\usage{
brapi_get_search_variants_searchResultsDbId(
  con = NULL,
  searchResultsDbId = "",
  pageToken = "",
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{searchResultsDbId}{character; required: TRUE; Unique identifier which
references the search results}

\item{pageToken}{character; required: FALSE; Used to request a specific page
of data to be returned. Tokenized pages are for large data sets, which can
not be efficiently broken into indexed pages. Use the \code{nextPageToken} and
\code{prevPageToken} from a prior response to construct a query and move to the
next or previous page respectively.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Get the results of a \code{Variants} search request
}
\details{
Returns the result of the advanced searching for the \code{Variants}
resource.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
con[["token"]] <- "YYYY"
# Saved or Asynchronous Search Response Example
out <-
 brapi_post_search_variants(con = con,
                            callSetDbIds = c("callset01",
                                             "callset02"),
                            start = 0,
                            end = 1000,
                            variantDbIds = c("variant01",
                                             "variant02"),
                            variantSetDbIds = "variantset1",
                            pageSize = 1000)
searchResultsDbId <- out$searchResultsDbId
brapi_get_search_variants_searchResultsDbId(con = con,
                                            searchResultsDbId = searchResultsDbId)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.0#/Variants/get_search_variants__searchResultsDbId_ }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-genotyping: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets_callSetDbId}()},
\code{\link{brapi_get_callsets}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_references_referenceDbId_bases}()},
\code{\link{brapi_get_references_referenceDbId}()},
\code{\link{brapi_get_referencesets_referenceSetDbId}()},
\code{\link{brapi_get_referencesets}()},
\code{\link{brapi_get_references}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_search_calls_searchResultsDbId}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_get_search_references_searchResultsDbId}()},
\code{\link{brapi_get_search_referencesets_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId_calls}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variantsets_variantSetDbId_callsets}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets_variantSetDbId}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_samples}()},
\code{\link{brapi_post_search_callsets}()},
\code{\link{brapi_post_search_calls}()},
\code{\link{brapi_post_search_referencesets}()},
\code{\link{brapi_post_search_references}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_variantsets}()},
\code{\link{brapi_post_search_variants}()},
\code{\link{brapi_post_variantsets_extract}()},
\code{\link{brapi_put_samples_sampleDbId}()}

Other Variants: 
\code{\link{brapi_get_variants_variantDbId_calls}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_search_variants}()}
}
\author{
Maikel Verouden
}
\concept{Variants}
\concept{brapi-genotyping}
