% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_seasons.R
\name{brapi_get_seasons}
\alias{brapi_get_seasons}
\title{get /seasons}
\usage{
brapi_get_seasons(
  con = NULL,
  seasonDbId = "",
  season = "",
  year = as.integer(NA),
  page = 0,
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{seasonDbId}{character; required: FALSE; The unique identifier for a
season. For backward compatibility it can be a string like '2012',
'1957-2004'}

\item{season}{character; required: FALSE; The term to describe a given
season. Example "Spring" OR "May" OR "
Planting_Time_7".}

\item{year}{integer; required: FALSE; The 4 digit year of a season. Example
"2017"}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is
'page'= 0). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Get the Seasons
}
\details{
Call to retrieve all seasons in the database.

A season is made of 2 parts; the primary year and a term which defines
a segment of the year.
This could be a traditional season, like "Spring" or "Summer
" or this could be a month, like "May" or "June"
or this could be an arbitrary season name which is meaningful to the
breeding program like "PlantingTime_3" or "Season E"
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_seasons(con = con)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Core/2.0#/Seasons/get_seasons }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-core: 
\code{\link{brapi_get_commoncropnames}()},
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations_locationDbId}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_programs_programDbId}()},
\code{\link{brapi_get_programs}()},
\code{\link{brapi_get_search_lists_searchResultsDbId}()},
\code{\link{brapi_get_search_locations_searchResultsDbId}()},
\code{\link{brapi_get_search_people_searchResultsDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_trials_searchResultsDbId}()},
\code{\link{brapi_get_seasons_seasonDbId}()},
\code{\link{brapi_get_serverinfo}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_get_trials_trialDbId}()},
\code{\link{brapi_get_trials}()},
\code{\link{brapi_post_lists_listDbId_items}()},
\code{\link{brapi_post_lists}()},
\code{\link{brapi_post_locations}()},
\code{\link{brapi_post_people}()},
\code{\link{brapi_post_programs}()},
\code{\link{brapi_post_seasons}()}

Other Seasons: 
\code{\link{brapi_get_seasons_seasonDbId}()},
\code{\link{brapi_post_seasons}()}
}
\author{
Maikel Verouden
}
\concept{Seasons}
\concept{brapi-core}
