% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_seasons_seasonDbId.R
\name{brapi_get_seasons_seasonDbId}
\alias{brapi_get_seasons_seasonDbId}
\title{get /seasons/\{seasonDbId\}}
\usage{
brapi_get_seasons_seasonDbId(con = NULL, seasonDbId = "")
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{seasonDbId}{character; required: TRUE; The unique identifier for a season. For backward compatibility it can be a string like '2012', '1957-2004'}
}
\value{
data.frame
}
\description{
Get the a single Season
}
\details{
Get the a single Season
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_seasons_seasonDbId(con = con, seasonDbId = "winter_2012")
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Core/2.0#/Seasons/get_seasons__seasonDbId_ }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-core: 
\code{\link{brapi_get_commoncropnames}()},
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations_locationDbId}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_programs_programDbId}()},
\code{\link{brapi_get_programs}()},
\code{\link{brapi_get_search_lists_searchResultsDbId}()},
\code{\link{brapi_get_search_locations_searchResultsDbId}()},
\code{\link{brapi_get_search_people_searchResultsDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_trials_searchResultsDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_serverinfo}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_get_trials}()}

Other Seasons: 
\code{\link{brapi_get_seasons}()}
}
\author{
Maikel Verouden
}
\concept{Seasons}
\concept{brapi-core}
