% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_serverinfo.R
\name{brapi_get_serverinfo}
\alias{brapi_get_serverinfo}
\title{get /serverinfo}
\usage{
brapi_get_serverinfo(con = NULL, dataType = "")
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{dataType}{character; required: FALSE; The data format supported by the
call.; default: "", other possible values:
"application/json"|
"text/csv"|
"text/tsv"|
"application/flapjack"}
}
\value{
data.frame
}
\description{
Get the list of implemented Calls
}
\details{
Implementation Notes: Having a consistent structure for the path
string of each call is very important for teams to be able to connect and
find errors. Read more on Github.Here are the rules for the path of each
call that should be returned:
\itemize{
\item Every word in the call path should match the documentation exactly, both in
spelling and capitalization. Note that path strings are all lower case, but
path parameters are camel case.
\item Each path should start relative to "/" and therefore should not
include "/"
\item No leading or trailing slashes ("/")
\item Path parameters are wrapped in curly braces ("\{\}"). The name of
the path parameter should be spelled exactly as it is specified in the
documentation.
}

Examples:
\itemize{
\item GOOD "call": "germplasm/\{germplasmDbId\}/pedigree"
\item BAD "call": "germplasm/\{id\}/pedigree"
\item BAD "call": "germplasm/\{germplasmDBid\}/pedigree"
\item BAD "call": "brapi/v2/germplasm/\{germplasmDbId\}/pedigree"
\item BAD "call": "/germplasm/\{germplasmDbId\}/pedigree/"
\item BAD "call": "germplasm/<germplasmDbId>/pedigree"
}
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_serverinfo(con = con)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Core/2.0#/Server\%20Info/get_serverinfo }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-core: 
\code{\link{brapi_get_commoncropnames}()},
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations_locationDbId}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_programs_programDbId}()},
\code{\link{brapi_get_programs}()},
\code{\link{brapi_get_search_lists_searchResultsDbId}()},
\code{\link{brapi_get_search_locations_searchResultsDbId}()},
\code{\link{brapi_get_search_people_searchResultsDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_trials_searchResultsDbId}()},
\code{\link{brapi_get_seasons_seasonDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_get_trials_trialDbId}()},
\code{\link{brapi_get_trials}()},
\code{\link{brapi_post_lists_listDbId_items}()},
\code{\link{brapi_post_lists}()},
\code{\link{brapi_post_locations}()},
\code{\link{brapi_post_people}()},
\code{\link{brapi_post_programs}()},
\code{\link{brapi_post_seasons}()},
\code{\link{brapi_post_studies}()},
\code{\link{brapi_post_trials}()},
\code{\link{brapi_put_lists_listDbId}()},
\code{\link{brapi_put_locations_locationDbId}()},
\code{\link{brapi_put_people_personDbId}()},
\code{\link{brapi_put_programs_programDbId}()},
\code{\link{brapi_put_seasons_seasonDbId}()}
}
\author{
Maikel Verouden
}
\concept{Server Info}
\concept{brapi-core}
