% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_studies.R
\name{brapi_get_studies}
\alias{brapi_get_studies}
\title{get /studies}
\usage{
brapi_get_studies(
  con = NULL,
  commonCropName = "",
  studyType = "",
  programDbId = "",
  locationDbId = "",
  seasonDbId = "",
  trialDbId = "",
  studyDbId = "",
  studyName = "",
  studyCode = "",
  studyPUI = "",
  germplasmDbId = "",
  observationVariableDbId = "",
  active = NA,
  sortBy = "",
  sortOrder = "",
  externalReferenceID = "",
  externalReferenceSource = "",
  page = 0,
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{commonCropName}{character; required: FALSE; Common name for the crop associated with this study}

\item{studyType}{character; required: FALSE; Filter based on study type unique identifier}

\item{programDbId}{character; required: FALSE; Program filter to only return studies associated with given program id.}

\item{locationDbId}{character; required: FALSE; Filter by location}

\item{seasonDbId}{character; required: FALSE; Filter by season or year}

\item{trialDbId}{character; required: FALSE; Filter by trial}

\item{studyDbId}{character; required: FALSE; Filter by study DbId}

\item{studyName}{character; required: FALSE; Filter by study name}

\item{studyCode}{character; required: FALSE; Filter by study code}

\item{studyPUI}{character; required: FALSE; Filter by study PUI}

\item{germplasmDbId}{character; required: FALSE; Filter by germplasm DbId}

\item{observationVariableDbId}{character; required: FALSE; Filter by observation variable DbId}

\item{active}{logical; required: FALSE; Filter active status true/false.; default: NA, other possible values: TRUE | FALSE}

\item{sortBy}{character; required: FALSE; Name of the field to sort by.}

\item{sortOrder}{character; required: FALSE; Sort order direction. Ascending/Descending.; default: "", other possible values: "asc"|"ASC"|"desc"|"DESC"}

\item{externalReferenceID}{character; required: FALSE; An external reference ID. Could be a simple string or a URI. (use with \code{externalReferenceSource} parameter)}

\item{externalReferenceSource}{character; required: FALSE; An identifier for the source system or database of an external reference (use with \code{externalReferenceID} parameter)}

\item{page}{integer; required: FALSE; Used to request a specific page of data to be returned. The page indexing starts at 0 (the first page is 'page'= 0). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Get a filtered list of Studies
}
\details{
Get list of studies, where StartDate and endDate should be ISO-8601 format for dates.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_studies(con = con)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Core/2.0#/Studies/get_studies }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-core: 
\code{\link{brapi_get_commoncropnames}()},
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations_locationDbId}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_programs_programDbId}()},
\code{\link{brapi_get_programs}()},
\code{\link{brapi_get_search_lists_searchResultsDbId}()},
\code{\link{brapi_get_search_locations_searchResultsDbId}()},
\code{\link{brapi_get_search_people_searchResultsDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_trials_searchResultsDbId}()},
\code{\link{brapi_get_seasons_seasonDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_serverinfo}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_get_trials_trialDbId}()},
\code{\link{brapi_get_trials}()},
\code{\link{brapi_post_lists}()}

Other Studies: 
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studytypes}()}
}
\author{
Maikel Verouden
}
\concept{Studies}
\concept{brapi-core}
