% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_traits_traitDbId.R
\name{brapi_get_traits_traitDbId}
\alias{brapi_get_traits_traitDbId}
\title{get /traits/\{traitDbId\}}
\usage{
brapi_get_traits_traitDbId(con = NULL, traitDbId = "")
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{traitDbId}{character; required: TRUE; Unique identifier of the trait to
retrieve details of.}
}
\value{
data.frame
}
\description{
Get the details of a specific Trait
}
\details{
Retrieve the details of a single trait
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_traits_traitDbId(con = con,
                           traitDbId = "trait_variable1")
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Phenotyping/2.0#/Traits/get_traits__traitDbId_ }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-phenotyping: 
\code{\link{brapi_get_events}()},
\code{\link{brapi_get_images_imageDbId}()},
\code{\link{brapi_get_images}()},
\code{\link{brapi_get_methods_methodDbId}()},
\code{\link{brapi_get_methods}()},
\code{\link{brapi_get_observationlevels}()},
\code{\link{brapi_get_observations_observationDbId}()},
\code{\link{brapi_get_observations}()},
\code{\link{brapi_get_observationunits_observationUnitDbId}()},
\code{\link{brapi_get_observationunits}()},
\code{\link{brapi_get_ontologies}()},
\code{\link{brapi_get_scales_scaleDbId}()},
\code{\link{brapi_get_scales}()},
\code{\link{brapi_get_traits}()},
\code{\link{brapi_get_variables_observationVariableDbId}()},
\code{\link{brapi_get_variables}()},
\code{\link{brapi_post_observations}()},
\code{\link{brapi_put_observations_observationDbId}()}

Other Traits: 
\code{\link{brapi_get_traits}()}
}
\author{
Maikel Verouden
}
\concept{Traits}
\concept{brapi-phenotyping}
