% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_variants_variantDbId.R
\name{brapi_get_variants_variantDbId}
\alias{brapi_get_variants_variantDbId}
\title{get /variants/\{variantDbId\}}
\usage{
brapi_get_variants_variantDbId(con = NULL, variantDbId = "")
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{variantDbId}{character; required: TRUE; The unique database identifier
of the \code{Variant} to be retrieved.}
}
\value{
data.frame
}
\description{
Gets a \code{Variant} by unique database identifier.
}
\details{
Returns a data.frame version of \code{Variant} as received from the
server.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_variants_variantDbId(con = con,
                               variantDbId = "variant01")
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.0#/Variants/get_variants__variantDbId_ }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-genotyping: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets_callSetDbId}()},
\code{\link{brapi_get_callsets}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_references_referenceDbId_bases}()},
\code{\link{brapi_get_references_referenceDbId}()},
\code{\link{brapi_get_references}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_search_calls_searchResultsDbId}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_variants_searchResultsDbId}()},
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_callsets}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets_variantSetDbId}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_samples}()},
\code{\link{brapi_post_search_callsets}()},
\code{\link{brapi_post_search_calls}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_variantsets}()},
\code{\link{brapi_post_search_variants}()},
\code{\link{brapi_post_variantsets_extract}()},
\code{\link{brapi_put_samples_sampleDbId}()}

Other Variants: 
\code{\link{brapi_get_search_variants_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId_calls}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_search_variants}()}
}
\author{
Maikel Verouden
}
\concept{Variants}
\concept{brapi-genotyping}
