% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_variants_variantDbId_calls.R
\name{brapi_get_variants_variantDbId_calls}
\alias{brapi_get_variants_variantDbId_calls}
\title{get /variants/\{variantDbId\}/calls}
\usage{
brapi_get_variants_variantDbId_calls(
  con = NULL,
  variantDbId = "",
  expandHomozygotes = NA,
  unknownString = "",
  sepPhased = "",
  sepUnphased = "",
  pageToken = "",
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{variantDbId}{character; required: TRUE; The unique database identifier
of the \code{Variant} to be retrieved.}

\item{expandHomozygotes}{logical; required: FALSE; Should homozygotes be
expanded (TRUE) or collapsed into a single occurrence (FALSE); default: NA
, other possible values: TRUE | FALSE}

\item{unknownString}{character; required: FALSE; The string to use as a
representation for missing data.}

\item{sepPhased}{character; required: FALSE; The string to use as a separator
for phased allele calls.}

\item{sepUnphased}{character; required: FALSE; The string to use as a
separator for unphased allele calls.}

\item{pageToken}{character; required: FALSE; Used to request a specific page
of data to be returned. Tokenized pages are for large data sets, which can
not be efficiently broken into indexed pages. Use the \code{nextPageToken} and
\code{prevPageToken} from a prior response to construct a query and move to the
next or previous page respectively.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Gets a list of \code{Calls} associated with a \code{Variant}.
}
\details{
The variant calls for this particular variant. Each one represents
the determination of genotype with respect to this variant. \code{Calls} in
this array are implicitly associated with this \code{Variant}.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_variants_variantDbId_calls(
  con = con,
  variantDbId = "variant01",
  pageSize = 1000)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.0#/Variants/get_variants__variantDbId__calls }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-genotyping: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets_callSetDbId}()},
\code{\link{brapi_get_callsets}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_search_calls_searchResultsDbId}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_get_search_variants_searchResultsDbId}()},
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variantsets_variantSetDbId_callsets}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets_variantSetDbId}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_search_callsets}()},
\code{\link{brapi_post_search_calls}()},
\code{\link{brapi_post_search_variantsets}()},
\code{\link{brapi_post_search_variants}()},
\code{\link{brapi_post_variantsets_extract}()}

Other Variants: 
\code{\link{brapi_get_search_variants_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_search_variants}()}
}
\author{
Maikel Verouden
}
\concept{Variants}
\concept{brapi-genotyping}
