% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_variantsets.R
\name{brapi_get_variantsets}
\alias{brapi_get_variantsets}
\title{get /variantsets}
\usage{
brapi_get_variantsets(
  con = NULL,
  variantSetDbId = "",
  variantDbId = "",
  callSetDbId = "",
  studyDbId = "",
  studyName = "",
  page = 0,
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{variantSetDbId}{character; required: FALSE; The unique database
identifier of the \code{VariantSet} to be retrieved.}

\item{variantDbId}{character; required: FALSE; The unique database identifier
of the \code{Variant} to be retrieved.}

\item{callSetDbId}{character; required: FALSE; The unique database identifier
of the \code{CallSet} to be retrieved.}

\item{studyDbId}{character; required: FALSE; Filter by unique study database
identifier.}

\item{studyName}{character; required: FALSE; Filter by study name.}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is
\code{page = 0}). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Gets a filtered list of \code{VariantSets}.
}
\details{
Will return a filtered list of \code{VariantSet}.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_variantsets(con = con)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.0#/Variant\%20Sets/get_variantsets }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-genotyping: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets_callSetDbId}()},
\code{\link{brapi_get_callsets}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_search_calls_searchResultsDbId}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_get_search_variants_searchResultsDbId}()},
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId_calls}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_search_callsets}()},
\code{\link{brapi_post_search_calls}()},
\code{\link{brapi_post_search_variantsets}()},
\code{\link{brapi_post_search_variants}()}

Other Variant Sets: 
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_post_search_variantsets}()}
}
\author{
Maikel Verouden
}
\concept{Variant Sets}
\concept{brapi-genotyping}
