% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_variantsets_variantSetDbId.R
\name{brapi_get_variantsets_variantSetDbId}
\alias{brapi_get_variantsets_variantSetDbId}
\title{get /variantsets/\{variantSetDbId\}}
\usage{
brapi_get_variantsets_variantSetDbId(con = NULL, variantSetDbId = "")
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{variantSetDbId}{character; required: TRUE; The unique database
identifier of the \code{variantSet} to be retrieved.}
}
\value{
data.frame
}
\description{
Gets a \code{variantSet} by unique database identifier.
}
\details{
This call will return a JSON version of a \code{VariantSet}.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_variantsets_variantSetDbId(con = con,
                                     variantSetDbId = "variantset1")
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.0#/Variant\%20Sets/get_variantsets__variantSetDbId_ }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-genotyping: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets_callSetDbId}()},
\code{\link{brapi_get_callsets}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_references_referenceDbId_bases}()},
\code{\link{brapi_get_references_referenceDbId}()},
\code{\link{brapi_get_referencesets_referenceSetDbId}()},
\code{\link{brapi_get_referencesets}()},
\code{\link{brapi_get_references}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_search_calls_searchResultsDbId}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_get_search_references_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_variants_searchResultsDbId}()},
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId_calls}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variantsets_variantSetDbId_callsets}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_samples}()},
\code{\link{brapi_post_search_callsets}()},
\code{\link{brapi_post_search_calls}()},
\code{\link{brapi_post_search_references}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_variantsets}()},
\code{\link{brapi_post_search_variants}()},
\code{\link{brapi_post_variantsets_extract}()},
\code{\link{brapi_put_samples_sampleDbId}()}

Other Variant Sets: 
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variantsets_variantSetDbId_callsets}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_post_search_variantsets}()},
\code{\link{brapi_post_variantsets_extract}()}
}
\author{
Maikel Verouden
}
\concept{Variant Sets}
\concept{brapi-genotyping}
