% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_variantsets_variantSetDbId_callsets.R
\name{brapi_get_variantsets_variantSetDbId_callsets}
\alias{brapi_get_variantsets_variantSetDbId_callsets}
\title{get /variantsets/\{variantSetDbId\}/callsets}
\usage{
brapi_get_variantsets_variantSetDbId_callsets(
  con = NULL,
  variantSetDbId = "",
  callSetDbId = "",
  callSetName = "",
  page = 0,
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{variantSetDbId}{character; required: TRUE; The unique database
identifier of the \code{VariantSet} to be retrieved.}

\item{callSetDbId}{character; required: FALSE; The unique database identifier
of the \code{CallSet} to be retrieved.}

\item{callSetName}{character; required: FALSE; The human readable name of the
\code{CallSet} to be retrieved.}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is
\code{page = 0}). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Gets a list of \code{CallSets} associated with a \code{VariantSet}.
}
\details{
Gets a list of \code{CallSets} associated with a \code{VariantSet}.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_variantsets_variantSetDbId_callsets(con = con,
                                              variantSetDbId = "variantset1",
                                              callSetDbId = "callset01",
                                              callSetName = "P2")
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.0#/Variant\%20Sets/get_variantsets__variantSetDbId__callsets }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-genotyping: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets_callSetDbId}()},
\code{\link{brapi_get_callsets}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_references}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_search_calls_searchResultsDbId}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_variants_searchResultsDbId}()},
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId_calls}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets_variantSetDbId}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_samples}()},
\code{\link{brapi_post_search_callsets}()},
\code{\link{brapi_post_search_calls}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_variantsets}()},
\code{\link{brapi_post_search_variants}()},
\code{\link{brapi_post_variantsets_extract}()},
\code{\link{brapi_put_samples_sampleDbId}()}

Other Variant Sets: 
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets_variantSetDbId}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_post_search_variantsets}()},
\code{\link{brapi_post_variantsets_extract}()}
}
\author{
Maikel Verouden
}
\concept{Variant Sets}
\concept{brapi-genotyping}
