% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_imagebreed_login.R
\name{brapi_imagebreed_login}
\alias{brapi_imagebreed_login}
\title{post /token for ImageBreed.org}
\usage{
brapi_imagebreed_login(user = "", password = "")
}
\arguments{
\item{user}{character; Username to be used, when logging in is required. Default is "".}

\item{password}{character; Password to be used for the specified user, when
logging in is required. Default is "".}
}
\value{
A list object of class 'brapi_con' with the authentication token
filled to make BrAPI calls in R from the ImageBreed database.
}
\description{
Login into ImageBreed.
}
\details{
Using the POST /token call for authentication a user can login to
ImageBreed.org to obtain a valid token from within R and use with additional
BrAPI R functions. The password is \strong{NOT} stored in the connection object
for safety.
}
\examples{
\dontrun{

# Login to ImageBreed.org with R
con <- brapi_imagebreed_login(user = "USERNAME_HERE",
                              password = "PASSWORD_HERE")
}

}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.1/Specification/Authentication/Authentication.md}{BrAPI Specs GitHub}
}
\seealso{
Other brapiutils: 
\code{\link{as.brapi_db}()},
\code{\link{brapi_canInternet}()},
\code{\link{brapi_checkCon}()},
\code{\link{brapi_connect}()},
\code{\link{brapi_db}()},
\code{\link{brapi_imagebreed_logout}()},
\code{\link{brapi_login}()},
\code{\link{brapi_serverinfo_show}()},
\code{\link{is.brapi_con}()},
\code{\link{print.brapi_con}()},
\code{\link{print.brapi_db}()}
}
\author{
brapir generator package
}
\concept{brapiutils}
