% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_imagebreed_logout.R
\name{brapi_imagebreed_logout}
\alias{brapi_imagebreed_logout}
\title{delete /token for ImageBreed.org}
\usage{
brapi_imagebreed_logout(con = NULL)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}
}
\value{
A list object of class 'brapi_con' from which the authentication
token has been removed.
}
\description{
Logout from ImageBreed.
}
\details{
Using the DELETE /token call for authentication a user can logout
from ImageBreed.org and release the valid access token from within R.
}
\examples{
\dontrun{

# Login to ImageBreed.org with R
con <- brapi_imagebreed_login(user = "USERNAME_HERE",
                              password = "PASSWORD_HERE")

# Logout from ImageBreed.org with R
con <- brapi_imagebreed_logout(con = con)
}

}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.1/Specification/Authentication/Authentication.md}{BrAPI Specs GitHub}
}
\seealso{
Other brapiutils: 
\code{\link{as.brapi_db}()},
\code{\link{brapi_canInternet}()},
\code{\link{brapi_checkCon}()},
\code{\link{brapi_connect}()},
\code{\link{brapi_db}()},
\code{\link{brapi_imagebreed_login}()},
\code{\link{brapi_login}()},
\code{\link{brapi_serverinfo_show}()},
\code{\link{is.brapi_con}()},
\code{\link{print.brapi_con}()},
\code{\link{print.brapi_db}()}
}
\author{
brapir generator package
}
\concept{brapiutils}
