% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_login.R
\name{brapi_login}
\alias{brapi_login}
\title{post /token}
\usage{
brapi_login(con = NULL)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}
}
\value{
A list object of class 'brapi_con' with the authentication token filled.
}
\description{
Login into a BrAPI compatible database.
}
\details{
Although the POST /token call for authentication has been deprecated after version 1.1 of the BrAPI specification, its functionality is still retained for compatibility reasons with the \href{https://bmspro.io/}{Breeding Management System (BMS Pro)} of \href{https://integratedbreeding.net/}{Integrated Breeding Platform} to allow users to obtain a valid token from within R and use with additional BrAPI R functions.
}
\examples{
\dontrun{

# Create a connection object
con <- brapi_db()$testserver
}

}
\references{
\href{https://github.com/plantbreeding/API/blob/V1.1/Specification/Authentication/Authentication.md}{BrAPI Specs GitHub}
}
\seealso{
Other brapiutils: 
\code{\link{as.brapi_db}()},
\code{\link{brapi_canInternet}()},
\code{\link{brapi_checkCon}()},
\code{\link{brapi_connect}()},
\code{\link{brapi_db}()},
\code{\link{brapi_serverinfo_show}()},
\code{\link{is.brapi_con}()},
\code{\link{print.brapi_con}()},
\code{\link{print.brapi_db}()}
}
\author{
brapir generator package
}
\concept{brapiutils}
