% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_post_lists.R
\name{brapi_post_lists}
\alias{brapi_post_lists}
\title{post /lists}
\usage{
brapi_post_lists(
  con = NULL,
  data = "",
  additionalInfo = list(),
  dateCreated = "",
  dateModified = "",
  externalReferences = "",
  listDescription = "",
  listName = "",
  listOwnerName = "",
  listOwnerPersonDbId = "",
  listSize = 0,
  listSource = "",
  listType = ""
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{data}{vector of type character; required: FALSE; The list of DbIds
contained in this list; default: "", when using multiple values
supply as c("value1", "value2").}

\item{additionalInfo}{list; required: FALSE; Additional arbitrary information.
If provided use the following construct list(additionalProp1 = "string",
additionalProp2 =  "string", additionalProp3 = "string"). The Examples
Section shows an example on how to construct the \code{additionalInfo} argument
as a list.}

\item{dateCreated}{character; required: FALSE; Timestamp when the entity was
first created. Coded in the ISO 8601 standard extended format, where date,
time and time zone information needs to be provided (check for example
https://www.w3.org/TR/NOTE-datetime).}

\item{dateModified}{character; required: FALSE; Timestamp when the entity was
last updated. Coded in the ISO 8601 standard extended format, where date,
time and time zone information needs to be provided (check for example
https://www.w3.org/TR/NOTE-datetime).}

\item{externalReferences}{data.frame; required: FALSE; A data.frame of
external reference ids. These are references to this piece of data in an
external system. Could be a simple string or a URI. The \code{externalReferences}
argument data.frame should contain the following columns:
\itemize{
\item \code{referenceID} character; required: TRUE; The external reference ID. Could
be a simple string or a URI.
\item \code{referenceSource} character; required: TRUE; An identifier for the source
system or database of this reference.
}

The Examples Section shows an example of how to construct the \code{externalReferences}
argument as a data.frame.}

\item{listDescription}{character; required: FALSE; Description of a List.}

\item{listName}{character; required: FALSE; Human readable name of a List.}

\item{listOwnerName}{character; required: FALSE; Human readable name of a List
Owner (usually a user or person).}

\item{listOwnerPersonDbId}{character; required: FALSE; The unique identifier
for a List Owner (usually a user or person).}

\item{listSize}{integer; required: FALSE; The number of elements in a List.}

\item{listSource}{character; required: FALSE; The description of where a List
originated from.}

\item{listType}{character; required: FALSE; default: "", other
possible values: "germplasm"|"markers"|"observations
"|"observationUnits"|"observationVariables"|"
programs"|"samples"|"studies"|"trials"}
}
\value{
data.frame
}
\description{
Create New List Objects
}
\details{
Create new list objects in the database
}
\examples{
\dontrun{
con <- brapi_db()$testserver
con[["token"]] <- "YYYY"
data <- c("germ1", "germ2", "germ3")
additionalInfo <- list(dummyData = "True", example = "post_lists")
dateCreated <- "2021-04-12T09:01:03.183+02:00"
dateModified <- "2021-04-13T07:00:24.596Z"
externalReferences <-
  data.frame(referenceID = c("doi:10.155454/12341234",
                             "http://purl.obolibrary.org/obo/ro.owl",
                             "75a50e76"),
             referenceSource = c("DOI",
                                 "OBO Library",
                                 "Remote Data Collection Upload Tool"))
listDescription <-
 "This is a list of germplasm I would like to investigate next summer season"
listName <- "WURgermplasm_Apr_2021"
listOwnerName <- "Bob Robertson"
listOwnerPersonDbId <- "person1"
listSize <- 3
listSource <- "WUR GeneBank"
listType <- "germplasm"
brapi_post_lists(con = con,
                 data = data,
                 additionalInfo = additionalInfo,
                 dateCreated = dateCreated,
                 dateModified = dateModified,
                 externalReferences = externalReferences,
                 listDescription = listDescription,
                 listName = listName,
                 listOwnerName = listOwnerName,
                 listOwnerPersonDbId = listOwnerPersonDbId,
                 listSize = listSize,
                 listSource = listSource,
                 listType = listType)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Core/2.0#/Lists/post_lists }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-core: 
\code{\link{brapi_get_commoncropnames}()},
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations_locationDbId}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_programs_programDbId}()},
\code{\link{brapi_get_programs}()},
\code{\link{brapi_get_search_lists_searchResultsDbId}()},
\code{\link{brapi_get_search_locations_searchResultsDbId}()},
\code{\link{brapi_get_search_people_searchResultsDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_trials_searchResultsDbId}()},
\code{\link{brapi_get_seasons_seasonDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_serverinfo}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_get_trials_trialDbId}()},
\code{\link{brapi_get_trials}()},
\code{\link{brapi_post_lists_listDbId_items}()},
\code{\link{brapi_post_locations}()},
\code{\link{brapi_post_people}()}

Other Lists: 
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_search_lists_searchResultsDbId}()},
\code{\link{brapi_post_lists_listDbId_items}()}
}
\author{
Maikel Verouden
}
\concept{Lists}
\concept{brapi-core}
