% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_post_people.R
\name{brapi_post_people}
\alias{brapi_post_people}
\title{post /people}
\usage{
brapi_post_people(
  con = NULL,
  additionalInfo = list(),
  description = "",
  emailAddress = "",
  externalReferences = "",
  firstName = "",
  middleName = "",
  lastName = "",
  mailingAddress = "",
  phoneNumber = "",
  userID = ""
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{additionalInfo}{list; required: FALSE; Additional arbitrary information.
If provided use the following construct list(additionalProp1 = "string",
additionalProp2 =  "string", additionalProp3 = "string"). The Examples
Section shows an example on how to construct the \code{additionalInfo} argument
as a list.}

\item{description}{character; required: FALSE; A description of this person.}

\item{emailAddress}{character; required: FALSE; An email address for this
person.}

\item{externalReferences}{data.frame; required: FALSE; A data.frame of
external reference ids. These are references to this piece of data in an
external system. Could be a simple string or a URI. The \code{externalReferences}
argument data.frame should contain the following columns:
\itemize{
\item \code{referenceID} character; required: TRUE; The external reference ID. Could
be a simple string or a URI.
\item \code{referenceSource} character; required: TRUE; An identifier for the source
system or database of this reference.
}

The Examples Section shows an example of how to construct the
\code{externalReferences} argument as a data.frame.}

\item{firstName}{character; required: FALSE; Persons first name.}

\item{middleName}{character; required: FALSE; Persons middle name.}

\item{lastName}{character; required: FALSE; Persons last name.}

\item{mailingAddress}{character; required: FALSE; Physical address of this
person.}

\item{phoneNumber}{character; required: FALSE; phone number of this person.}

\item{userID}{character; required: FALSE; A systems user ID associated with
this person. Different from \code{personDbId} because you could have a person,
who is not a user of the system.}
}
\value{
data.frame
}
\description{
Create new People
}
\details{
Create new People entities. \code{personDbId} is generated and managed by the server.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
con[["token"]] <- "YYYY"
additionalInfo <- list(additionalProp1 = "string",
                       additionalProp2 = "string",
                       additionalProp3 = "string")
description <- "Bob likes pina coladas and getting caught in the rain."
emailAddress <- "bob@bob.com"
externalReferences <-
  data.frame(referenceID = c("doi:10.155454/12341234",
                             "http://purl.obolibrary.org/obo/ro.owl",
                             "75a50e76"),
             referenceSource = c("DOI",
                                 "OBO Library",
                                 "Remote Data Collection Upload Tool"))
firstName <- "Bob"
middleName <- "Danger"
lastName <- "Robertson"
mailingAddress <- "123 Street Ave, City, State, Country"
phoneNumber <- "+1-555-555-5555"
userID <- "bob-23"

brapi_post_people(con = con,
                  additionalInfo = additionalInfo,
                  description = description,
                  emailAddress = emailAddress,
                  externalReferences = externalReferences,
                  firstName = firstName,
                  middleName = middleName,
                  lastName = lastName,
                  mailingAddress = mailingAddress,
                  phoneNumber = phoneNumber,
                  userID = usedID)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Core/2.0#/People/post_people }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-core: 
\code{\link{brapi_get_commoncropnames}()},
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations_locationDbId}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_programs_programDbId}()},
\code{\link{brapi_get_programs}()},
\code{\link{brapi_get_search_lists_searchResultsDbId}()},
\code{\link{brapi_get_search_locations_searchResultsDbId}()},
\code{\link{brapi_get_search_people_searchResultsDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_trials_searchResultsDbId}()},
\code{\link{brapi_get_seasons_seasonDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_serverinfo}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_get_trials_trialDbId}()},
\code{\link{brapi_get_trials}()},
\code{\link{brapi_post_lists_listDbId_items}()},
\code{\link{brapi_post_lists}()},
\code{\link{brapi_post_locations}()}

Other People: 
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_search_people_searchResultsDbId}()}
}
\author{
Maikel Verouden
}
\concept{People}
\concept{brapi-core}
