% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_post_programs.R
\name{brapi_post_programs}
\alias{brapi_post_programs}
\title{post /programs}
\usage{
brapi_post_programs(
  con = NULL,
  abbreviation = "",
  additionalInfo = list(),
  commonCropName = "",
  documentationURL = "",
  externalReferences = "",
  leadPersonDbId = "",
  leadPersonName = "",
  objective = "",
  programName = ""
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{abbreviation}{character; required: FALSE; An abbreviation which
represents this program.}

\item{additionalInfo}{list; required: FALSE; Additional arbitrary information.
If provided use the following construct list(additionalProp1 = "string",
additionalProp2 =  "string", additionalProp3 = "string").

The Examples section shows an example on how to construct the
\code{additionalInfo} argument as a list.}

\item{commonCropName}{character; required: FALSE; Common name for the crop,
which this program is for.}

\item{documentationURL}{character; required: FALSE; A URL to the human
readable documentation of this object.}

\item{externalReferences}{data.frame; required: FALSE; A data.frame of
external reference ids. These are references to this piece of data in an
external system. Could be a simple string or a URI. The \code{externalReferences}
argument data.frame should contain the following columns:
\itemize{
\item \code{referenceID} character; required: TRUE; The external reference ID. Could
be a simple string or a URI.
\item \code{referenceSource} character; required: TRUE; An identifier for the source
system or database of this reference.
}

The Examples section shows an example of how to construct the
\code{externalReferences} argument as a data.frame.}

\item{leadPersonDbId}{character; required: FALSE; The unique identifier of
the program leader. Needs to be an existing \code{personDbId}.}

\item{leadPersonName}{character; required: FALSE; The name of the program
leader.}

\item{objective}{character; required: FALSE; The primary objective of the
program.}

\item{programName}{character; required: FALSE; Human readable name of the
program.}
}
\value{
data.frame
}
\description{
Add new breeding Programs to the database
}
\details{
Add new breeding Programs to the database. The \code{programDbId} is set
by the server, all other fields are take from the request body,
or a default value is used.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
con[["token"]] <- "YYYY"
abbreviation <- "P1"
additionalInfo <- list(additionalProp1 = "string",
                                additionalProp2 = "string",
                                additionalProp3 = "string")
commonCropName <- "Tomatillo"
documentationURL <- "https://wiki.brapi.org"
externalReferences <-
  data.frame(referenceID = c("doi:10.155454/12341234",
                             "http://purl.obolibrary.org/obo/ro.owl",
                             "75a50e76"),
             referenceSource = c("DOI",
                                 "OBO Library",
                                 "Remote Data Collection Upload Tool"))
leadPersonDbId <- "program_person_1"
leadPersonName <- "Bob Robertson"
objective <- "Make a better tomatillo"
programName <- "Tomatillo_Breeding_Program"

brapi_post_programs(con = con,
                    abbreviation = abbreviation,
                    additionalInfo = additionalInfo,
                    commonCropName = commonCropName,
                    documentationURL = documentationURL,
                    externalReferences = externalReferences,
                    leadPersonDbId = leadPersonDbId,
                    leadPersonName = leadPersonName,
                    objective = objective,
                    programName = programName)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Core/2.0#/Programs/post_programs }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-core: 
\code{\link{brapi_get_commoncropnames}()},
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations_locationDbId}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_programs_programDbId}()},
\code{\link{brapi_get_programs}()},
\code{\link{brapi_get_search_lists_searchResultsDbId}()},
\code{\link{brapi_get_search_locations_searchResultsDbId}()},
\code{\link{brapi_get_search_people_searchResultsDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_trials_searchResultsDbId}()},
\code{\link{brapi_get_seasons_seasonDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_serverinfo}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_get_trials_trialDbId}()},
\code{\link{brapi_get_trials}()},
\code{\link{brapi_post_lists_listDbId_items}()},
\code{\link{brapi_post_lists}()},
\code{\link{brapi_post_locations}()},
\code{\link{brapi_post_people}()},
\code{\link{brapi_post_seasons}()},
\code{\link{brapi_post_studies}()},
\code{\link{brapi_post_trials}()},
\code{\link{brapi_put_lists_listDbId}()},
\code{\link{brapi_put_locations_locationDbId}()},
\code{\link{brapi_put_people_personDbId}()},
\code{\link{brapi_put_programs_programDbId}()},
\code{\link{brapi_put_seasons_seasonDbId}()},
\code{\link{brapi_put_studies_studyDbId}()}

Other Programs: 
\code{\link{brapi_get_programs_programDbId}()},
\code{\link{brapi_get_programs}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_put_programs_programDbId}()}
}
\author{
Maikel Verouden
}
\concept{Programs}
\concept{brapi-core}
