% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_post_samples.R
\name{brapi_post_samples}
\alias{brapi_post_samples}
\title{post /samples}
\usage{
brapi_post_samples(
  con = NULL,
  additionalInfo = list(),
  column = as.integer(NA),
  externalReferences = "",
  germplasmDbId = "",
  observationUnitDbId = "",
  plateDbId = "",
  plateName = "",
  programDbId = "",
  row = "",
  sampleBarcode = "",
  sampleDescription = "",
  sampleGroupDbId = "",
  sampleName = "",
  samplePUI = "",
  sampleTimestamp = "",
  sampleType = "",
  studyDbId = "",
  takenBy = "",
  tissueType = "",
  trialDbId = "",
  well = ""
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{additionalInfo}{list; required: FALSE; Additional arbitrary information.
If provided use the following construct list(additionalProp1 = "string",
additionalProp2 =  "string", additionalProp3 = "string").

The Examples section shows an example on how to construct the
\code{additionalInfo} argument as a list.}

\item{column}{integer; required: FALSE; The Column identifier for this
sample's location in the plate.}

\item{externalReferences}{data.frame; required: FALSE; A data.frame of
external reference ids. These are references to this piece of data in an
external system. Could be a simple string or a URI. The \code{externalReferences}
argument data.frame should contain the following columns:
\itemize{
\item \code{referenceID} character; required: TRUE; The external reference ID. Could
be a simple string or a URI.
\item \code{referenceSource} character; required: TRUE; An identifier for the source
system or database of this reference.
}

The Examples section shows an example of how to construct the
\code{externalReferences} argument as a data.frame.}

\item{germplasmDbId}{character; required: FALSE; Unique database identifier
for a germplasm (accession).}

\item{observationUnitDbId}{character; required: FALSE; Database identifier,
which unique identifies a observation unit.}

\item{plateDbId}{character; required: FALSE; The database identifier, which
uniquely identifies a plate of samples.}

\item{plateName}{character; required: FALSE; The human readable name of a
plate.}

\item{programDbId}{character; required: FALSE; The datbase identifier, which
uniquely identifies a program.}

\item{row}{character; required: FALSE; The Row identifier for this sample's
location in the plate.}

\item{sampleBarcode}{character; required: FALSE; A unique identifier
physically attached to the sample.}

\item{sampleDescription}{character; required: FALSE; Description of a sample.
MIAPPE V1.1 (DM-79) Sample description - Any information not captured by
the other sample fields, including quantification, sample treatments and
processing.}

\item{sampleGroupDbId}{character; required: FALSE; The identifier, which
uniquely identifies a group of samples.}

\item{sampleName}{character; required: FALSE; The name of the sample.}

\item{samplePUI}{character; required: FALSE; A permanent unique identifier
for the sample (DOI, URL, UUID, \emph{etc.}). MIAPPE V1.1 (DM-81) External ID -
An identifier for the sample in a persistent repository, comprising the
name of the repository and the accession number of the observation unit
therein. Submission to the EBI Biosamples repository is recommended. URI
are recommended when possible.}

\item{sampleTimestamp}{character; required: FALSE; The date and time a sample
was collected from the field. MIAPPE V1.1 (DM-80) Collection date - The
date and time when the sample was collected / harvested.}

\item{sampleType}{character; required: FALSE; The type of sample taken. e.g.
'DNA', 'RNA', 'Tissue', \emph{etc.}.}

\item{studyDbId}{character; required: FALSE; The database identifier, which
uniquely identifies a study.}

\item{takenBy}{character; required: FALSE; The name or identifier of the
entity, which took the sample from the field.}

\item{tissueType}{character; required: FALSE; The type of tissue sampled,
e.g. 'Leaf', 'Root', \emph{etc.}. MIAPPE V1.1 (DM-78) Plant anatomical entity -
A description of  the plant part (e.g. leaf) or the plant product (e.g.
resin) from which the sample was taken, in the form of an accession number
to a suitable controlled vocabulary (Plant Ontology).}

\item{trialDbId}{character; required: FALSE; The database identifier, which
uniquely identifies a trial.}

\item{well}{character; required: FALSE; The Well identifier for this sample's
location in the plate. Usually a concatenation of Row and Column, or just
a number if the samples are not part of an ordered plate.}
}
\value{
data.frame
}
\description{
Add new Samples
}
\details{
Call to register the event of a sample being taken. Sample
identifier is assigned as a result of the operation and returned in
response.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
con[["token"]] <- "YYYY"
## Create function argument values
additionalInfo <- list(dummyData = "TRUE",
                       example = "post_samples")
column <- 3
externalReferences <-
  data.frame(referenceID = c("doi:10.155454/12341234",
                             "http://purl.obolibrary.org/obo/ro.owl",
                             "75a50e76"),
             referenceSource = c("DOI",
                                 "OBO Library",
                                 "Remote Data Collection Upload Tool"))
germplasmDbId <- "germplasm3"
observationUnitDbId <- "observation_unit3"
plateDbId <- "plate1"
plateName <- "Plate 1"
programDbId <- "program1"
row <- "B"
sampleBarcode <- "1234567890"
sampleDescription <- "Fake sample"
sampleGroupDbId <- "sample_group_02"
sampleName <- "Sample 4"
samplePUI <- "doi://40.12345/fake/4423"
sampleTimestamp <- "2012-12-12T12:12:12Z"
sampleType <- "Tissue"
studyDbId <- "study1"
takenBy <- "M. Verouden"
tissueType <- "Stem"
trialDbId <- "trial1"
well <- "B3"
brapi_post_samples(con = con,
                   additionalInfo = additionalInfo,
                   column = column,
                   externalReferences = externalReferences,
                   germplasmDbId = germplasmDbId,
                   observationUnitDbId = observationUnitDbId,
                   plateDbId = plateDbId,
                   plateName = plateName,
                   programDbId = programDbId,
                   row = row,
                   sampleBarcode = sampleBarcode,
                   sampleDescription = sampleDescription,
                   sampleGroupDbId = sampleGroupDbId,
                   sampleName = sampleName,
                   samplePUI = samplePUI,
                   sampleTimestamp = sampleTimestamp,
                   sampleType = sampleType,
                   studyDbId = studyDbId,
                   takenBy = takenBy,
                   tissueType = tissueType,
                   trialDbId = trialDbId,
                   well = well)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.0#/Samples/post_samples }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-genotyping: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets_callSetDbId}()},
\code{\link{brapi_get_callsets}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_references_referenceDbId_bases}()},
\code{\link{brapi_get_references_referenceDbId}()},
\code{\link{brapi_get_referencesets}()},
\code{\link{brapi_get_references}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_search_calls_searchResultsDbId}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_get_search_references_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_variants_searchResultsDbId}()},
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId_calls}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variantsets_variantSetDbId_callsets}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets_variantSetDbId}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_search_callsets}()},
\code{\link{brapi_post_search_calls}()},
\code{\link{brapi_post_search_references}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_variantsets}()},
\code{\link{brapi_post_search_variants}()},
\code{\link{brapi_post_variantsets_extract}()},
\code{\link{brapi_put_samples_sampleDbId}()}

Other Samples: 
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_put_samples_sampleDbId}()}
}
\author{
Maikel Verouden
}
\concept{Samples}
\concept{brapi-genotyping}
