% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_post_search_attributes.R
\name{brapi_post_search_attributes}
\alias{brapi_post_search_attributes}
\title{post /search/attributes}
\usage{
brapi_post_search_attributes(
  con = NULL,
  attributeDbIds = "",
  attributeNames = "",
  dataTypes = "",
  externalReferenceIDs = "",
  externalReferenceSources = "",
  methodDbIds = "",
  ontologyDbIds = "",
  scaleDbIds = "",
  studyDbId = "",
  traitClasses = "",
  traitDbIds = "",
  page = 0,
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{attributeDbIds}{vector of type character; required: FALSE; The unique
Germplasm Attribute database identifier(s) to search for; default:
"", when using multiple values supply as c("value1",
"value2").}

\item{attributeNames}{vector of type character; required: FALSE; The human
readable Germplasm Attribute names to search for; default: "",
when using multiple values supply as c("value1",
"value2").}

\item{dataTypes}{vector of type character; required: FALSE; The scale data
types to filter search results; default: "", when using multiple
values supply as c("value1", "value2").}

\item{externalReferenceIDs}{vector of type character; required: FALSE; The
external reference IDs to search for. Could be a simple strings or a URIs.
(use with \code{externalReferenceSources} parameter); default: "",
when using multiple values supply as c("value1",
"value2").}

\item{externalReferenceSources}{vector of type character; required: FALSE;
The identifiers for the source system or database of an external reference
to search for (use with \code{externalReferenceIDs} parameter); default:
"", when using multiple values supply as c("value1",
"value2").}

\item{methodDbIds}{vector of type character; required: FALSE; Unique method
database indentifier(s) to filter search results; default: "",
when using multiple values supply as c("value1",
"value2").}

\item{ontologyDbIds}{vector of type character; required: FALSE; Unique
ontology database identiefier(s) to search for; default: "",
when using multiple values supply as c("value1",
"value2").}

\item{scaleDbIds}{vector of type character; required: FALSE; List of unique
scale database identifier(s) to filter search results; default:
"", when using multiple values supply as c("value1",
"value2").}

\item{studyDbId}{vector of type character; required: FALSE; The unique
database identifier(s) of studies to filter on; default: "",
when using multiple values supply as c("value1",
"value2").}

\item{traitClasses}{vector of type character; required: FALSE; List of trait
classes to filter search results; default: "", when using
multiple values supply as c("value1", "value2").}

\item{traitDbIds}{vector of type character; required: FALSE; List of trait
unique database identifier(s) to filter search results; default:
"", when using multiple values supply as c("value1",
"value2").}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is
\code{page = 0}). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Submit a search request for Germplasm \code{Attributes}
}
\details{
Submit a search request for \code{Attributes}. Function will return
either the search results (Status 200 for an immediate response) or a
\code{searchResultsDbId} (Status 202 for both a saved and an asynchronous
search).
}
\examples{
\dontrun{
con <- brapi_db()$testserver
con[["token"]] <- "YYYY"

# Immediate Response Example
brapi_post_search_attributes(con = con, page = 0, pageSize = 1000)

# Saved or Asynchronous Search Response Example
brapi_post_search_attributes(con = con,
                             attributeDbIds = c("attribute1",
                                                 "attribute2"),
                             attributeNames = "Rht-B2b",
                             externalReferenceIDs = "https://brapi.org/specification",
                             externalReferenceSources = "BrAPI Doc",
                             traitClasses = "Numeric",
                             page = 0,
                             pageSize = 1000)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Germplasm/2.0#/Germplasm\%20Attributes/post_search_attributes }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-germplasm: 
\code{\link{brapi_get_attributes}()},
\code{\link{brapi_get_attributevalues}()},
\code{\link{brapi_get_germplasm}()},
\code{\link{brapi_get_search_attributes_searchResultsDbId}()},
\code{\link{brapi_get_search_germplasm_searchResultsDbId}()},
\code{\link{brapi_post_search_germplasm}()}

Other Germplasm Attributes: 
\code{\link{brapi_get_attributes}()},
\code{\link{brapi_get_search_attributes_searchResultsDbId}()}
}
\author{
Maikel Verouden
}
\concept{Germplasm Attributes}
\concept{brapi-germplasm}
