% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_post_search_attributevalues.R
\name{brapi_post_search_attributevalues}
\alias{brapi_post_search_attributevalues}
\title{post /search/attributevalues}
\usage{
brapi_post_search_attributevalues(
  con = NULL,
  attributeDbIds = "",
  attributeNames = "",
  attributeValueDbIds = "",
  dataTypes = "",
  externalReferenceIDs = "",
  externalReferenceSources = "",
  germplasmDbIds = "",
  germplasmNames = "",
  methodDbIds = "",
  ontologyDbIds = "",
  page = 0,
  pageSize = 1000,
  scaleDbIds = "",
  traitClasses = "",
  traitDbIds = ""
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{attributeDbIds}{vector of type character; required: FALSE; The unique
Germplasm Attribute database identifier(s) to search for; default:
"", when using multiple values supply as c("value1",
"value2").}

\item{attributeNames}{vector of type character; required: FALSE; Human
readable Germplasm Attribute names to search for; default: "",
when using multiple values supply as c("value1",
"value2").}

\item{attributeValueDbIds}{vector of type character; required: FALSE; The
unique Germplasm Attribute Value database identifiers to search for;
default: "", when using multiple values supply as
c("value1", "value2").}

\item{dataTypes}{vector of type character; required: FALSE; List of scale
data types to filter search results; default: "", when using
multiple values supply as c("value1", "value2").}

\item{externalReferenceIDs}{vector of type character; required: FALSE; List
of external reference IDs. Could be a simple strings or a URIs. (use with
\code{externalReferenceSources} parameter); default: "", when using
multiple values supply as c("value1", "value2").}

\item{externalReferenceSources}{vector of type character; required: FALSE;
List of identifiers for the source system or database of an external
reference (use with \code{externalReferenceIDs} parameter); default:
"", when using multiple values supply as c("value1",
"value2").}

\item{germplasmDbIds}{vector of type character; required: FALSE; List of IDs
which uniquely identify germplasm to search for; default: "",
when using multiple values supply as c("value1",
"value2").}

\item{germplasmNames}{vector of type character; required: FALSE; List of
human readable names to identify germplasm to search for; default:
"", when using multiple values supply as c("value1",
"value2").}

\item{methodDbIds}{vector of type character; required: FALSE; List of methods
to filter search results; default: "", when using multiple
values supply as c("value1", "value2").}

\item{ontologyDbIds}{vector of type character; required: FALSE; List of
ontology IDs to search for; default: "", when using multiple
values supply as c("value1", "value2").}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is
\code{page = 0}). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}

\item{scaleDbIds}{vector of type character; required: FALSE; List of scales
to filter search results; default: "", when using multiple
values supply as c("value1", "value2").}

\item{traitClasses}{vector of type character; required: FALSE; List of trait
classes to filter search results; default: "", when using
multiple values supply as c("value1", "value2").}

\item{traitDbIds}{vector of type character; required: FALSE; List of trait
unique ID to filter search results; default: "", when using
multiple values supply as c("value1", "value2").}
}
\value{
data.frame
}
\description{
Submit a search request for Germplasm \code{AttributeValues}
}
\details{
Submit a search request for Germplasm \code{Attributevalues}. Function
will return either the search results (Status 200 for an immediate
response) or a \code{searchResultsDbId} (Status 202 for both a saved and an
asynchronous search).
}
\examples{
\dontrun{
con <- brapi_db()$testserver
con[["token"]] <- "YYYY"

# Immediate Response Example
brapi_post_search_attributevalues(con = con, page = 0, pageSize = 1000)

# Saved or Asynchronous Search Response Example
brapi_post_search_attributevalues(con = con,
                                  attributeDbIds = c("attribute1",
                                                     "attribute2"),
                                  attributeNames = "Rht-B1b",
                                  attributeValueDbIds = "attribute_val1",
                                  externalReferenceIDs = "https://brapi.org/specification",
                                  externalReferenceSources = "BrAPI Doc",
                                  page = 0,
                                  pageSize = 1000)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Germplasm/2.0#/Germplasm\%20Attribute\%20Values/post_search_attributevalues }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-germplasm: 
\code{\link{brapi_get_attributes}()},
\code{\link{brapi_get_attributevalues}()},
\code{\link{brapi_get_germplasm}()},
\code{\link{brapi_get_search_attributes_searchResultsDbId}()},
\code{\link{brapi_get_search_attributevalues_searchResultsDbId}()},
\code{\link{brapi_get_search_germplasm_searchResultsDbId}()},
\code{\link{brapi_post_search_attributes}()},
\code{\link{brapi_post_search_germplasm}()}

Other Germplasm Attribute Values: 
\code{\link{brapi_get_attributevalues}()},
\code{\link{brapi_get_search_attributevalues_searchResultsDbId}()}
}
\author{
Maikel Verouden
}
\concept{Germplasm Attribute Values}
\concept{brapi-germplasm}
