% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_post_search_calls.R
\name{brapi_post_search_calls}
\alias{brapi_post_search_calls}
\title{post /search/calls}
\usage{
brapi_post_search_calls(
  con = NULL,
  callSetDbIds = "",
  expandHomozygotes = NA,
  pageSize = 1000,
  pageToken = "",
  sepPhased = "",
  sepUnphased = "",
  unknownString = "",
  variantDbIds = "",
  variantSetDbIds = ""
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{callSetDbIds}{vector of type character; required: FALSE; The CallSet to
search for specified as unique database callset identifier(s).; default:
"", when using multiple values supply as c("value1",
"value2").}

\item{expandHomozygotes}{logical; required: FALSE; Should homozygotes be
expanded (TRUE) or collapsed into a single occurrence (FALSE); default: NA
, other possible values: TRUE | FALSE.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}

\item{pageToken}{character; required: FALSE; Used to request a specific page
of data to be returned. Tokenized pages are for large data sets, which can
not be efficiently broken into indexed pages. Use the \code{nextPageToken} and
\code{prevPageToken} from a prior response to construct a query and move to the
next or previous page respectively.}

\item{sepPhased}{character; required: FALSE; The string to use as a separator
for phased allele calls.}

\item{sepUnphased}{character; required: FALSE; The string to use as a
separator for unphased allele calls.}

\item{unknownString}{character; required: FALSE; The string used as a
representation for missing data.; default: "".}

\item{variantDbIds}{vector of type character; required: FALSE; The unique
variant database identifier(s) to search for.; default: "",
when using multiple values supply as c("value1",
"value2").}

\item{variantSetDbIds}{vector of type character; required: FALSE; The unique
variantset database identifier(s) to search for.; default: "",
when using multiple values supply as c("value1",
"value2").}
}
\value{
data.frame
}
\description{
Submit a search request for \code{Calls}
}
\details{
Submit a search request for \code{Calls}. Function will return either the
search results (Status 200 for an immediate response) or a
\code{searchResultsDbId} (Status 202 for both a saved and an asynchronous
search).
}
\examples{
\dontrun{
con <- brapi_db()$testserver
con[["token"]] <- "YYYY"

# Immediate Response Example
brapi_post_search_calls(con = con, pageSize = 1000)

# Saved or Asynchronous Search Response Example
brapi_post_search_calls(con = con,
                        callSetDbIds = c("callset12",
                                         "callset21"),
                        expandHomozygotes = TRUE,
                        pageSize = 1000,
                        sepPhased = "/",
                        sepUnphased = "|",
                        unknownString = "-",
                        variantDbIds = c("variant01",
                                         "variant23"))
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.0#/Calls/post_search_calls }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-genotyping: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets_callSetDbId}()},
\code{\link{brapi_get_callsets}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_references_referenceDbId_bases}()},
\code{\link{brapi_get_references_referenceDbId}()},
\code{\link{brapi_get_referencesets}()},
\code{\link{brapi_get_references}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_search_calls_searchResultsDbId}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_get_search_references_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_variants_searchResultsDbId}()},
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId_calls}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variantsets_variantSetDbId_callsets}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets_variantSetDbId}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_samples}()},
\code{\link{brapi_post_search_callsets}()},
\code{\link{brapi_post_search_references}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_variantsets}()},
\code{\link{brapi_post_search_variants}()},
\code{\link{brapi_post_variantsets_extract}()},
\code{\link{brapi_put_samples_sampleDbId}()}

Other Calls: 
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_search_calls_searchResultsDbId}()}
}
\author{
Maikel Verouden
}
\concept{Calls}
\concept{brapi-genotyping}
