% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_post_search_markerpositions.R
\name{brapi_post_search_markerpositions}
\alias{brapi_post_search_markerpositions}
\title{post /search/markerpositions}
\usage{
brapi_post_search_markerpositions(
  con = NULL,
  linkageGroupNames = "",
  mapDbIds = "",
  minPosition = as.integer(NA),
  maxPosition = as.integer(NA),
  variantDbIds = "",
  page = 0,
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{linkageGroupNames}{vector of type character; required: FALSE; The
uniquely identifiable name(s) of the linkage group(s) to search for;
default: "", when using multiple values supply as c("
value1", "value2").}

\item{mapDbIds}{vector of type character; required: FALSE; The unique
database identifier(s) of the map(s) to search for; default: "",
when using multiple values supply as c("value1", "value2
").}

\item{minPosition}{integer; required: FALSE; The minimum position.}

\item{maxPosition}{integer; required: FALSE; The maximum position.}

\item{variantDbIds}{vector of type character; required: FALSE; Unique
internal database identifier(s) of the variant(s) to search for; default:
"", when using multiple values supply as c("value1",
"value2").}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is
\code{page = 0}). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Submit a search request for \code{MarkerPositions}
}
\details{
Submit a search request for \code{MarkerPositions}. Function will return
either the search results (Status 200 for an immediate response) or a
\code{searchResultsDbId} (Status 202 for both a saved and an asynchronous
search).
}
\examples{
\dontrun{
con <- brapi_db()$testserver
con[["token"]] <- "YYYY"

# Immediate Response Example
brapi_post_search_markerpositions(con = con,
                                  page = 0,
                                  pageSize = 1000)

# Saved or Asynchronous Search Response Example
brapi_post_search_markerpositions(con = con,
                                  linkageGroupNames = c("Chromosome 1",
                                                        "Chromosome 2"),
                                  mapDbIds = c("genome_map1",
                                               "genome_map2"),
                                  minPosition = 0,
                                  maxPosition = 60000,
                                  variantDbIds = c("variant03",
                                                   "variant06"),
                                  page = 0,
                                  pageSize = 1000)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.0#/Genome\%20Maps/post_search_markerpositions }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-genotyping: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets_callSetDbId}()},
\code{\link{brapi_get_callsets}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_maps_mapDbId_linkagegroups}()},
\code{\link{brapi_get_maps_mapDbId}()},
\code{\link{brapi_get_maps}()},
\code{\link{brapi_get_markerpositions}()},
\code{\link{brapi_get_references_referenceDbId_bases}()},
\code{\link{brapi_get_references_referenceDbId}()},
\code{\link{brapi_get_referencesets_referenceSetDbId}()},
\code{\link{brapi_get_referencesets}()},
\code{\link{brapi_get_references}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_search_calls_searchResultsDbId}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_get_search_markerpositions_searchResultsDbId}()},
\code{\link{brapi_get_search_references_searchResultsDbId}()},
\code{\link{brapi_get_search_referencesets_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_variants_searchResultsDbId}()},
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId_calls}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variantsets_variantSetDbId_callsets}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets_variantSetDbId}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_samples}()},
\code{\link{brapi_post_search_callsets}()},
\code{\link{brapi_post_search_calls}()},
\code{\link{brapi_post_search_referencesets}()},
\code{\link{brapi_post_search_references}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_variantsets}()},
\code{\link{brapi_post_search_variants}()},
\code{\link{brapi_post_variantsets_extract}()},
\code{\link{brapi_put_samples_sampleDbId}()}

Other Genome Maps: 
\code{\link{brapi_get_maps_mapDbId_linkagegroups}()},
\code{\link{brapi_get_maps_mapDbId}()},
\code{\link{brapi_get_maps}()},
\code{\link{brapi_get_markerpositions}()},
\code{\link{brapi_get_search_markerpositions_searchResultsDbId}()}
}
\author{
Maikel Verouden
}
\concept{Genome Maps}
\concept{brapi-genotyping}
