% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_post_search_observations.R
\name{brapi_post_search_observations}
\alias{brapi_post_search_observations}
\title{post /search/observations}
\usage{
brapi_post_search_observations(
  con = NULL,
  externalReferenceIDs = "",
  externalReferenceSources = "",
  germplasmDbIds = "",
  germplasmNames = "",
  locationDbIds = "",
  locationNames = "",
  observationDbIds = "",
  observationLevelRelationships = "",
  observationLevels = "",
  observationTimeStampRangeEnd = "",
  observationTimeStampRangeStart = "",
  observationUnitDbIds = "",
  observationVariableDbIds = "",
  observationVariableNames = "",
  page = 0,
  pageSize = 1000,
  programDbIds = "",
  programNames = "",
  seasonDbIds = "",
  studyDbIds = "",
  studyNames = "",
  trialDbIds = "",
  trialNames = ""
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{externalReferenceIDs}{vector of type character; required: FALSE; List
of external reference IDs. Could be a simple strings or a URIs. (use with
\code{externalReferenceSources} parameter); default: "", when using
multiple values supply as c("value1", "value2").}

\item{externalReferenceSources}{vector of type character; required: FALSE;
List of identifiers for the source system or database of an external
reference (use with \code{externalReferenceIDs} parameter); default:
"", when using multiple values supply as c("value1",
"value2").}

\item{germplasmDbIds}{vector of type character; required: FALSE; List of
unique database identifiers, which uniquely identify germplasm to search
for; default: "", when using multiple values supply as c(
"value1", "value2").}

\item{germplasmNames}{vector of type character; required: FALSE; List of
human readable names to identify germplasm to search for; default:
"", when using multiple values supply as c("value1",
"value2").}

\item{locationDbIds}{vector of type character; required: FALSE; The unique
location identifier(s) to search for; default: "", when using
multiple values supply as c("value1", "value2").}

\item{locationNames}{vector of type character; required: FALSE; A human
readable location name(s) to search for; default: "", when using
multiple values supply as c("value1", "value2").}

\item{observationDbIds}{vector of type character; required: FALSE; The unique
database identifier(s) of an observation(s) to search for; default:
"", when using multiple values supply as c("value1",
"value2").}

\item{observationLevelRelationships}{data.frame; required: FALSE; Data.frame
of observation levels to indicate the granularity level at which the
measurements are taken. Each row in the data.frame defines the level code,
level name (\code{levelName}), and the level order, as integer, where that
level exists in the hierarchy of levels. \code{levelOrders} lower numbers are
at the top of the hierarchy (i.e. field -> 0) and higher numbers are at
the bottom of the hierarchy (ie plant > 6). \code{levelCode} is an identifier
code for this level tag. Identify this observation unit by each level of
the hierarchy, where it exists.

The Examples section shows an example of how to construct the
\code{observationLevelRelationships} argument as a data.frame.}

\item{observationLevels}{data.frame; required: FALSE; A data.frame of
Observation levels, which indicate the granularity level at which the
measurements are taken. The \code{observationLevels} argument data.frame can
contain the following columns:
\itemize{
\item \code{levelName} character; required: FALSE; A name for this level.
\item \code{levelOrder} interger; required: FALSE; \code{levelOrder} defines, where that
level exists in the hierarchy of levels. \code{levelOrder}'s lower numbers
are at the top of the hierarchy (i.e. field -> 1) and higher numbers are
at the bottom of the hierarchy (i.e. plant -> 9).
}

The Examples section shows an example of how to construct the
\code{observationLevels} argument as a data.frame.}

\item{observationTimeStampRangeEnd}{character; required: FALSE; Timestamp
to end the range for observations to filter on. Coded in the ISO 8601
standard extended format, where date, time and time zone information needs
to be provided (check for example https://www.w3.org/TR/NOTE-datetime).}

\item{observationTimeStampRangeStart}{character; required: FALSE; Timestamp
to start the range for observations to filter on. Coded in the ISO 8601
standard extended format, where date, time and time zone information needs
to be provided (check for example https://www.w3.org/TR/NOTE-datetime).}

\item{observationUnitDbIds}{vector of type character; required: FALSE; The
unique database identifier of an observation unit; default: "",
when using multiple values supply as c("value1",
"value2").}

\item{observationVariableDbIds}{vector of type character; required: FALSE;
The unique database identifiers of Variables to search for; default:
"", when using multiple values supply as c("value1",
"value2").}

\item{observationVariableNames}{vector of type character; required: FALSE;
The names of Variables to search for; default: "", when using
multiple values supply as c("value1", "value2").}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is
\code{page = 0}). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}

\item{programDbIds}{vector of type character; required: FALSE; Unique program
database identifiers to search for; default: "", when using
multiple values supply as c("value1", "value2").}

\item{programNames}{vector of type character; required: FALSE; A program
names to search for; default: "", when using multiple values
supply as c("value1", "value2").}

\item{seasonDbIds}{vector of type character; required: FALSE; Unique season
database identifier(s) to search for; default: "", when using
multiple values supply as c("value1", "value2").}

\item{studyDbIds}{vector of type character; required: FALSE; List of unique
database study identifiers to search for; default: "", when
using multiple values supply as c("value1", "value2").}

\item{studyNames}{vector of type character; required: FALSE; List of study
names to filter search results; default: "", when using multiple
values supply as c("value1", "value2").}

\item{trialDbIds}{vector of type character; required: FALSE; The identifiers,
which uniquely identify trials to search for; default: "", when
using multiple values supply as c("value1", "value2").}

\item{trialNames}{vector of type character; required: FALSE; The human
readable names of trials to search for; default: "", when using
multiple values supply as c("value1", "value2").}
}
\value{
data.frame
}
\description{
Submit a search request for \code{Observations}
}
\details{
Submit a search request for \code{Observations}. Function will return
either the search results (Status 200 for an immediate response) or a
\code{searchResultsDbId} (Status 202 for both a saved and an asynchronous
search).
}
\examples{
\dontrun{
con <- brapi_db()$testserver
con[["token"]] <- "YYYY"
observationLevelRelationships <- data.frame(
  levelCode  = c("fieldA", "rep1", "block1"),
  levelName  = c("field", "rep", "block"),
  levelOrder = c(1, 2, 3))
observationLevels <-
  data.frame(levelName = c("field", "block", "plot"),
             levelOrder = c(0, 1, 2))

# Immediate Response Example
brapi_post_search_observations(con = con,
                               germplasmDbIds = c("germplasm1",
                                                  "germplasm2",
                                                  "germplasm3"))

# Saved or Asynchronous Search Response Example
brapi_post_search_observations(
  con = con,
  externalReferenceIDs = "https://brapi.org/specification",
  externalReferenceSources = "BrAPI Doc",
  germplasmDbIds = c("germplasm1",
                     "germplasm2",
                     "germplasm3"),
  page = 0,
  pageSize = 1000)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Phenotyping/2.0#/Observations/post_search_observations }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-phenotyping: 
\code{\link{brapi_get_events}()},
\code{\link{brapi_get_images_imageDbId}()},
\code{\link{brapi_get_images}()},
\code{\link{brapi_get_methods_methodDbId}()},
\code{\link{brapi_get_methods}()},
\code{\link{brapi_get_observationlevels}()},
\code{\link{brapi_get_observations_observationDbId}()},
\code{\link{brapi_get_observations}()},
\code{\link{brapi_get_observationunits_observationUnitDbId}()},
\code{\link{brapi_get_observationunits}()},
\code{\link{brapi_get_ontologies}()},
\code{\link{brapi_get_scales_scaleDbId}()},
\code{\link{brapi_get_scales}()},
\code{\link{brapi_get_search_images_searchResultsDbId}()},
\code{\link{brapi_get_search_observations_searchResultsDbId}()},
\code{\link{brapi_get_search_observationunits_searchResultsDbId}()},
\code{\link{brapi_get_search_variables_searchResultsDbId}()},
\code{\link{brapi_get_traits_traitDbId}()},
\code{\link{brapi_get_traits}()},
\code{\link{brapi_get_variables_observationVariableDbId}()},
\code{\link{brapi_get_variables}()},
\code{\link{brapi_post_methods}()},
\code{\link{brapi_post_observations}()},
\code{\link{brapi_post_observationunits}()},
\code{\link{brapi_post_scales}()},
\code{\link{brapi_post_search_images}()},
\code{\link{brapi_post_search_observationunits}()},
\code{\link{brapi_post_search_variables}()},
\code{\link{brapi_post_traits}()},
\code{\link{brapi_post_variables}()},
\code{\link{brapi_put_methods_methodDbId}()},
\code{\link{brapi_put_observations_observationDbId}()},
\code{\link{brapi_put_observationunits_observationUnitDbId}()},
\code{\link{brapi_put_scales_scaleDbId}()},
\code{\link{brapi_put_traits_traitDbId}()},
\code{\link{brapi_put_variables_observationVariableDbId}()}

Other Observations: 
\code{\link{brapi_get_observations_observationDbId}()},
\code{\link{brapi_get_observations}()},
\code{\link{brapi_get_search_observations_searchResultsDbId}()},
\code{\link{brapi_post_observations}()},
\code{\link{brapi_put_observations_observationDbId}()}
}
\author{
Maikel Verouden
}
\concept{Observations}
\concept{brapi-phenotyping}
