% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_post_search_referencesets.R
\name{brapi_post_search_referencesets}
\alias{brapi_post_search_referencesets}
\title{post /search/referencesets}
\usage{
brapi_post_search_referencesets(
  con = NULL,
  referenceSetDbIds = "",
  accessions = "",
  assemblyPUIs = "",
  md5checksums = "",
  page = 0,
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{referenceSetDbIds}{vector of type character; required: FALSE; The
unique database identifier(s) of a reference set or reference sets to
search.; default: "", when using multiple values supply as c(
"value1", "value2").}

\item{accessions}{vector of type character; required: FALSE; If specified,
return the reference sets for which the accession matches this string
(case-sensitive, exact match); default: "", when using multiple
values supply as c("value1", "value2").}

\item{assemblyPUIs}{vector of type character; required: FALSE; If set, return
the reference sets for which the \code{assemblyId} matches this string (case-
sensitive, exact match).}

\item{md5checksums}{vector of type character; required: FALSE; If specified,
return the reference sets for which the \code{md5checksum} matches this string
(case-sensitive, exact match); default: "", when using multiple
values supply as c("value1", "value2").}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is
\code{page = 0}). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Submit a search request for \code{ReferenceSets}
}
\details{
Submit a search request for \code{ReferenceSets}. Function will return
either the search results (Status 200 for an immediate response) or a
\code{searchResultsDbId} (Status 202 for both a saved and an asynchronous
search).
}
\examples{
\dontrun{
con <- brapi_db()$testserver
con[["token"]] <- "YYYY"

# Immediate Response Example
brapi_post_search_referencesets(con = con,
                                page = 0,
                                pageSize = 1000)

# Saved or Asynchronous Search Response Example
brapi_post_search_referencesets(
 con = con,
 referenceSetDbIds = c("reference_set1",
                       "reference_set2"),
 accessions = c("A0000001",
                "A0000002"),
 assemblyPUIs = "doi://10.12345/fake/9876",
 md5checksums = "0ba836092b9efada3a4aa72bb0eb0e04",
 page = 0,
 pageSize = 1000)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.0#/Reference\%20Sets/post_search_referencesets }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-genotyping: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets_callSetDbId}()},
\code{\link{brapi_get_callsets}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_maps_mapDbId_linkagegroups}()},
\code{\link{brapi_get_maps_mapDbId}()},
\code{\link{brapi_get_maps}()},
\code{\link{brapi_get_markerpositions}()},
\code{\link{brapi_get_references_referenceDbId_bases}()},
\code{\link{brapi_get_references_referenceDbId}()},
\code{\link{brapi_get_referencesets_referenceSetDbId}()},
\code{\link{brapi_get_referencesets}()},
\code{\link{brapi_get_references}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_search_calls_searchResultsDbId}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_get_search_references_searchResultsDbId}()},
\code{\link{brapi_get_search_referencesets_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_variants_searchResultsDbId}()},
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId_calls}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variantsets_variantSetDbId_callsets}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets_variantSetDbId}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_samples}()},
\code{\link{brapi_post_search_callsets}()},
\code{\link{brapi_post_search_calls}()},
\code{\link{brapi_post_search_references}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_variantsets}()},
\code{\link{brapi_post_search_variants}()},
\code{\link{brapi_post_variantsets_extract}()},
\code{\link{brapi_put_samples_sampleDbId}()}

Other Reference Sets: 
\code{\link{brapi_get_referencesets_referenceSetDbId}()},
\code{\link{brapi_get_referencesets}()},
\code{\link{brapi_get_search_referencesets_searchResultsDbId}()}
}
\author{
Maikel Verouden
}
\concept{Reference Sets}
\concept{brapi-genotyping}
