% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_post_search_trials.R
\name{brapi_post_search_trials}
\alias{brapi_post_search_trials}
\title{post /search/trials}
\usage{
brapi_post_search_trials(
  con = NULL,
  active = NA,
  commonCropNames = "",
  contactDbIds = "",
  externalReferenceIDs = "",
  externalReferenceSources = "",
  locationDbIds = "",
  locationNames = "",
  programDbIds = "",
  programNames = "",
  searchDateRangeEnd = "",
  searchDateRangeStart = "",
  studyDbIds = "",
  studyNames = "",
  trialDbIds = "",
  trialNames = "",
  trialPUIs = "",
  page = 0,
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{active}{logical; required: FALSE; Is this trial currently active;
default: NA, other possible values: TRUE | FALSE}

\item{commonCropNames}{vector of type character; required: FALSE; Common crop
name(s) for the trials to search for; default: "", when using
multiple values supply as c("value1", "value2").}

\item{contactDbIds}{vector of type character; required: FALSE; Unique contact
identifier(s) of trials to search for; default: "", when using
multiple values supply as c("value1", "value2").}

\item{externalReferenceIDs}{vector of type character; required: FALSE;
External reference identifier(s) to search for. Could be a simple strings
or a URIs (use with \code{externalReferenceSources} parameter).; default:
"", when using multiple values supply as c("value1",
"value2").}

\item{externalReferenceSources}{vector of type character; required: FALSE;
Source system or database identifier(s) of an external reference(s) to
search for (use with \code{externalReferenceIDs} parameter); default:
"", when using multiple values supply as c("value1",
"value2").}

\item{locationDbIds}{vector of type character; required: FALSE; The unique
location identifier(s) to search for; default: "", when using
multiple values supply as c("value1", "value2").}

\item{locationNames}{vector of type character; required: FALSE; A human
readable location name(s) to search for; default: "", when using
multiple values supply as c("value1", "value2").}

\item{programDbIds}{vector of type character; required: FALSE; Unique program
identifier(s) to search for; default: "", when using multiple
values supply as c("value1", "value2").}

\item{programNames}{vector of type character; required: FALSE; Program
name(s) to search for; default: "", when using multiple values
supply as c("value1", "value2").}

\item{searchDateRangeEnd}{character; required: FALSE; The start of the
overlapping search date range. \code{searchDateRangeStart} must be before
\code{searchDateRangeEnd}.

Return a Trial entity if any of the following cases are true:
\itemize{
\item \code{searchDateRangeStart} is before \code{trial.endDate} AND \code{searchDateRangeEnd}
is null
\item \code{searchDateRangeStart} is before \code{trial.endDate} AND \code{searchDateRangeEnd}
is after \code{trial.startDate}
\item \code{searchDateRangeEnd} is after \code{trial.startDate} AND \code{searchDateRangeStart}
is null
\item \code{searchDateRangeEnd} is after \code{trial.startDate} AND \code{searchDateRangeStart}
is before \code{trial.endDate}
}}

\item{searchDateRangeStart}{character; required: FALSE; The start of the
overlapping search date range. \code{searchDateRangeStart} must be before
\code{searchDateRangeEnd}.}

\item{studyDbIds}{vector of type character; required: FALSE; Unique study
identifier(s) to search for; default: "", when using multiple
values supply as c("value1", "value2").}

\item{studyNames}{vector of type character; required: FALSE; Study name(s)
to search for; default: "", when using multiple values supply as
c("value1", "value2").}

\item{trialDbIds}{vector of type character; required: FALSE; Unique trial
identifier(s) to search for; default: "", when using multiple
values supply as c("value1", "value2").}

\item{trialNames}{vector of type character; required: FALSE; The human
readable trial name(s)to search for; default: "", when using
multiple values supply as c("value1", "value2").}

\item{trialPUIs}{vector of type character; required: FALSE; A permanent trial
identifier(s) to search for. Could be DOI or other URI formatted
identifier.; default: "", when using multiple values supply as
c("value1", "value2").}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is
\code{page = 0}). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Submit a search request for Trials
}
\details{
Advanced searching for the programs resource. Function will return
either the search results (Status 200 for an immediate response) or a
\code{searchResultsDbId} (Status 202 for both a saved and an asynchronous
search).
}
\examples{
\dontrun{
con <- brapi_db()$testserver
con[["token"]] <- "YYYY"

TO BE ADDED STILL
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Core/2.0#/Trials/post_search_trials }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-core: 
\code{\link{brapi_get_commoncropnames}()},
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations_locationDbId}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_programs_programDbId}()},
\code{\link{brapi_get_programs}()},
\code{\link{brapi_get_search_lists_searchResultsDbId}()},
\code{\link{brapi_get_search_locations_searchResultsDbId}()},
\code{\link{brapi_get_search_people_searchResultsDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_trials_searchResultsDbId}()},
\code{\link{brapi_get_seasons_seasonDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_serverinfo}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_get_trials_trialDbId}()},
\code{\link{brapi_get_trials}()},
\code{\link{brapi_post_lists_listDbId_items}()},
\code{\link{brapi_post_lists}()},
\code{\link{brapi_post_locations}()},
\code{\link{brapi_post_people}()},
\code{\link{brapi_post_programs}()},
\code{\link{brapi_post_search_lists}()},
\code{\link{brapi_post_search_locations}()},
\code{\link{brapi_post_search_people}()},
\code{\link{brapi_post_search_programs}()},
\code{\link{brapi_post_search_studies}()},
\code{\link{brapi_post_seasons}()},
\code{\link{brapi_post_studies}()},
\code{\link{brapi_post_trials}()},
\code{\link{brapi_put_lists_listDbId}()},
\code{\link{brapi_put_locations_locationDbId}()},
\code{\link{brapi_put_people_personDbId}()},
\code{\link{brapi_put_programs_programDbId}()},
\code{\link{brapi_put_seasons_seasonDbId}()},
\code{\link{brapi_put_studies_studyDbId}()},
\code{\link{brapi_put_trials_trialDbId}()}

Other Trials: 
\code{\link{brapi_get_search_trials_searchResultsDbId}()},
\code{\link{brapi_get_trials_trialDbId}()},
\code{\link{brapi_get_trials}()},
\code{\link{brapi_post_trials}()},
\code{\link{brapi_put_trials_trialDbId}()}
}
\author{
Maikel Verouden
}
\concept{Trials}
\concept{brapi-core}
