% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_post_search_variantsets.R
\name{brapi_post_search_variantsets}
\alias{brapi_post_search_variantsets}
\title{post /search/variantsets}
\usage{
brapi_post_search_variantsets(
  con = NULL,
  callSetDbIds = "",
  studyDbIds = "",
  studyNames = "",
  variantDbIds = "",
  variantSetDbIds = "",
  page = 0,
  pageSize = 1000
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{callSetDbIds}{vector of type character; required: FALSE; The \code{CallSets}
to search as specified by the unique database identifier(s); default:
"", when using multiple values supply as c("value1",
"value2").}

\item{studyDbIds}{vector of type character; required: FALSE; List of unique
database study identifiers to search for; default: "", when
using multiple values supply as c("value1", "value2").}

\item{studyNames}{vector of type character; required: FALSE; List of study
names to filter search results; default: "", when using multiple
values supply as c("value1", "value2").}

\item{variantDbIds}{vector of type character; required: FALSE; The unique
Variant database identifiers to search.; default: "", when using
multiple values supply as c("value1", "value2").}

\item{variantSetDbIds}{vector of type character; required: FALSE; The unique
\code{VariantSet} database identifiers to search.; default: "", when
using multiple values supply as c("value1", "value2").}

\item{page}{integer; required: FALSE; Used to request a specific page of data
to be returned. The page indexing starts at 0 (the first page is
\code{page = 0}). Default is \code{0}.}

\item{pageSize}{integer; required: FALSE; The size of the pages to be
returned. Default is \code{1000}.}
}
\value{
data.frame
}
\description{
Submit a search request for \code{VariantSets}
}
\details{
Submit a search request for \code{VariantSets}. Function will return
either the search results (Status 200 for an immediate response) or a
\code{searchResultsDbId} (Status 202 for both a saved and an asynchronous
search).
}
\examples{
\dontrun{
con <- brapi_db()$testserver
con[["token"]] <- "YYYY"

# Immediate Response Example
brapi_post_search_variantsets(con = con,
                              page = 0,
                              pageSize = 1000)

# Saved or Asynchronous Search Response Example
brapi_post_search_variantsets(con = con,
                              studyDbIds = c("study1",
                                             "study2"),
                              studyNames = "Paw paw 2013 yield trial",
                              variantSetDbIds = c("variantset1",
                                                  "variantset2"),
                              page = 0,
                              pageSize = 1000)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Genotyping/2.0#/Variant\%20Sets/post_search_variantsets }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-genotyping: 
\code{\link{brapi_get_callsets_callSetDbId_calls}()},
\code{\link{brapi_get_callsets_callSetDbId}()},
\code{\link{brapi_get_callsets}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_references_referenceDbId_bases}()},
\code{\link{brapi_get_references_referenceDbId}()},
\code{\link{brapi_get_referencesets}()},
\code{\link{brapi_get_references}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_search_calls_searchResultsDbId}()},
\code{\link{brapi_get_search_callsets_searchResultsDbId}()},
\code{\link{brapi_get_search_references_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_variants_searchResultsDbId}()},
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variants_variantDbId_calls}()},
\code{\link{brapi_get_variants_variantDbId}()},
\code{\link{brapi_get_variantsets_variantSetDbId_callsets}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets_variantSetDbId}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_get_variants}()},
\code{\link{brapi_post_samples}()},
\code{\link{brapi_post_search_callsets}()},
\code{\link{brapi_post_search_calls}()},
\code{\link{brapi_post_search_references}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_variants}()},
\code{\link{brapi_post_variantsets_extract}()},
\code{\link{brapi_put_samples_sampleDbId}()}

Other Variant Sets: 
\code{\link{brapi_get_search_variantsets_searchResultsDbId}()},
\code{\link{brapi_get_variantsets_variantSetDbId_callsets}()},
\code{\link{brapi_get_variantsets_variantSetDbId_calls}()},
\code{\link{brapi_get_variantsets_variantSetDbId_variants}()},
\code{\link{brapi_get_variantsets_variantSetDbId}()},
\code{\link{brapi_get_variantsets}()},
\code{\link{brapi_post_variantsets_extract}()}
}
\author{
Maikel Verouden
}
\concept{Variant Sets}
\concept{brapi-genotyping}
