% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_post_seasons.R
\name{brapi_post_seasons}
\alias{brapi_post_seasons}
\title{post /seasons}
\usage{
brapi_post_seasons(
  con = NULL,
  seasonDbId = "",
  seasonName = "",
  year = as.integer(NA)
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{seasonDbId}{character; required: FALSE; The Id which uniquely
identifies a season. For backward compatibility it can be a string like
'2012', '1957-2004'.}

\item{seasonName}{character; required: FALSE; Name of the season. ex.
'Spring', 'Q2', 'Season A', \emph{etc.}}

\item{year}{integer; required: FALSE; The 4 digit year of the season.}
}
\value{
data.frame
}
\description{
POST new Seasons
}
\details{
Add new season entries to the database
}
\examples{
\dontrun{
con <- brapi_db()$testserver
con[["token"]] <- "YYYY"
seasonDbId <- "spring_2021"
seasonName <- "spring"
year <- 2021

brapi_post_seasons(con = con,
                   seasonDbId = seasonDbId,
                   seasonName = seasonName,
                   year = year)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Core/2.0#/Seasons/post_seasons }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-core: 
\code{\link{brapi_get_commoncropnames}()},
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations_locationDbId}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_programs_programDbId}()},
\code{\link{brapi_get_programs}()},
\code{\link{brapi_get_search_lists_searchResultsDbId}()},
\code{\link{brapi_get_search_locations_searchResultsDbId}()},
\code{\link{brapi_get_search_people_searchResultsDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_trials_searchResultsDbId}()},
\code{\link{brapi_get_seasons_seasonDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_serverinfo}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_get_trials_trialDbId}()},
\code{\link{brapi_get_trials}()},
\code{\link{brapi_post_lists_listDbId_items}()},
\code{\link{brapi_post_lists}()},
\code{\link{brapi_post_locations}()},
\code{\link{brapi_post_people}()},
\code{\link{brapi_post_programs}()},
\code{\link{brapi_post_search_lists}()},
\code{\link{brapi_post_studies}()},
\code{\link{brapi_post_trials}()},
\code{\link{brapi_put_lists_listDbId}()},
\code{\link{brapi_put_locations_locationDbId}()},
\code{\link{brapi_put_people_personDbId}()},
\code{\link{brapi_put_programs_programDbId}()},
\code{\link{brapi_put_seasons_seasonDbId}()},
\code{\link{brapi_put_studies_studyDbId}()},
\code{\link{brapi_put_trials_trialDbId}()}

Other Seasons: 
\code{\link{brapi_get_seasons_seasonDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_put_seasons_seasonDbId}()}
}
\author{
Maikel Verouden
}
\concept{Seasons}
\concept{brapi-core}
