% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_post_studies.R
\name{brapi_post_studies}
\alias{brapi_post_studies}
\title{post /studies}
\usage{
brapi_post_studies(
  con = NULL,
  active = NA,
  additionalInfo = list(),
  commonCropName = "",
  contacts = "",
  culturalPractices = "",
  dataLinks = "",
  documentationURL = "",
  endDate = "",
  environmentParameters = "",
  experimentalDesign = list(),
  externalReferences = "",
  growthFacility = list(),
  lastUpdate = list(),
  license = "",
  locationDbId = "",
  locationName = "",
  observationLevels = "",
  observationUnitsDescription = "",
  seasons = "",
  startDate = "",
  studyCode = "",
  studyDescription = "",
  studyName = "",
  studyPUI = "",
  studyType = "",
  trialDbId = "",
  trialName = ""
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{active}{logical; required: FALSE; Is this study currently active;
default: NA, other possible values: TRUE | FALSE}

\item{additionalInfo}{list; required: FALSE; Additional arbitrary information.
If provided use the following construct list(additionalProp1 = "string",
additionalProp2 =  "string", additionalProp3 = "string").

The Examples section shows an example on how to construct the
\code{additionalInfo} argument as a list.}

\item{commonCropName}{character; required: FALSE; Common name for the crop
associated with this study.}

\item{contacts}{data.frame; required: FALSE; A data.frame of contact entities
associated with this study. The \code{contacts} argument data.frame should or
can contain the following columns:
\itemize{
\item \code{contactDbId} character; required: TRUE; The ID which uniquely
identifies this contact. MIAPPE V1.1 (DM-33) Person ID - An identifier
for the data submitter. If that submitter is an individual, ORCID
identifiers are recommended.
\item \code{email}    character; required: FALSE; The contacts email address. MIAPPE
V1.1 (DM-32) Person email - The electronic mail address of the person.
\item \code{instituteName} character; required: FALSE; The name of the institution
which this contact is part of. MIAPPE V1.1 (DM-35) Person affiliation -
The institution the person belongs to.
\item \code{name}    character; required: FALSE; The full name of this contact person.
MIAPPE V1.1 (DM-31) Person name - The name of the person (either full
name or as used in scientific publications).
\item \code{orcid} character; required: FALSE; The Open Researcher and Contributor
ID for this contact person (orcid.org). MIAPPE V1.1 (DM-33) Person ID -
An identifier for the data submitter. If that submitter is an individual,
ORCID identifiers are recommended.
\item \code{type} character;    required: FALSE; The type of person this contact
represents (ex: Coordinator, Scientist, PI, etc.). MIAPPE V1.1 (DM-34)
Person role - Type of contribution of the person to the investigation.
}

The Examples section shows an example of how to construct the
\code{contacts} argument as a data.frame.}

\item{culturalPractices}{character; required: FALSE; MIAPPE V1.1 (DM-28)
Cultural practices - General description of the cultural practices of the
study.}

\item{dataLinks}{data.frame; required: FALSE; A data.frame of links to extra
data files associated with this study. Extra data could include notes,
images, and reference data. The \code{dataLinks} arguments data.frame can
contain the following columns:
\itemize{
\item \code{dataFormat} character; required: FALSE; The structure of the data
within a file. For example - VCF, table, image archive, multispectral
image archives in EDAM ontology (used in Galaxy). MIAPPE V1.1 (DM-38)
Data file description - Description of the format of the data file. May
be a standard file format name, or a description of organization of the
data in a tabular file.
\item \code{description} character; required: FALSE; The general description of
this data link. MIAPPE V1.1 (DM-38) Data file description - Description
of the format of the data file. May be a standard file format name, or a
description of organization of the data in a tabular file.
\item \code{fileFormat} character; required: FALSE; The MIME type of the file (i.e.
text/csv, application/excel, application/zip). MIAPPE V1.1 (DM-38) Data
file description - Description of the format of the data file. May be a
standard file format name, or a description of organization of the data
in a tabular file.
\item \code{name} character; required: FALSE; The name of the external data link.
MIAPPE V1.1 (DM-38) Data file description - Description of the format of
the data file. May be a standard file format name, or a description of
organization of the data in a tabular file.
\item \code{provenance} character; required: FALSE; The description of the origin
or ownership of this linked data. Could be a formal reference to
software, method, or workflow.
\item \code{scientificType} character; required: FALSE; The general type of data.
For example- Genotyping, Phenotyping raw data, Phenotyping reduced data,
Environmental, \emph{etc.}
\item \code{url} character; required: FALSE; URL describing the location of this
data file to view or download. MIAPPE V1.1 (DM-37) Data file link - Link
to the data file (or digital object) in a public database or in a
persistent institutional repository; or identifier of the data file when
submitted together with the MIAPPE submission.
\item \code{version} character; required: FALSE; The version number for this data.
MIAPPE V1.1 (DM-39) Data file version - The version of the data set (the
actual data).
}

The Examples section shows an example of how to construct the
\code{dataLinks} argument as a data.frame.}

\item{documentationURL}{character; required: FALSE; A URL to the human
readable documentation of this object.}

\item{endDate}{character; required: FALSE; The date the study will end or
ended. Coded in the ISO 8601 standard extended format, where date,
time and time zone information needs to be provided (check for example
\url{https://www.w3.org/TR/NOTE-datetime}
). MIAPPE V1.1 (DM-15) End date of study - Date and, if relevant, time
when the experiment ended.}

\item{environmentParameters}{data.frame; required: FALSE; A data.frame of
Environmental parameters that were kept constant throughout the study and
did not change between observation units. MIAPPE V1.1 (DM-57) Environment
\itemize{
\item Environmental parameters that were kept constant throughout the study
and did not change between observation units or assays. Environment
characteristics that vary over time, i.e. environmental variables, should
be recorded as Observed Variables (see below). The \code{environmentParameters}
argument data.frame should or can contain the following columns:
\item \code{description} character; required: TRUE; Human-readable value of the
environment parameter (defined above) constant within the experiment.
\item \code{parameterName} character; required: TRUE; Name of the environment
parameter constant within the experiment. MIAPPE V1.1 (DM-58) Environment
parameter - Name of the environment parameter constant within the
experiment.
\item \code{parameterPUI} character; required: FALSE; URI pointing to an ontology
class for the parameter.
\item \code{unit} character; required: FALSE; Unit of the value for this parameter
\item \code{unitPUI} character; required: FALSE; URI pointing to an ontology class
for the unit.
\item \code{value} character; required: FALSE; Numerical or categorical value.
MIAPPE V1.1 (DM-59) Environment parameter value - Value of the
environment parameter (defined above) constant within the experiment.
\item \code{valuePUI} character; required: FALSE; URI pointing to an ontology class
for the parameter value.
}

The Examples section shows an example of how to construct the
\code{environmentParameters} argument as a data.frame.}

\item{experimentalDesign}{list; required: FALSE; The experimental and
statistical design full description plus a category PUI taken from crop
research ontology or agronomy ontology. The \code{experimentalDesign} list
argument can contain the following elements:
\itemize{
\item \code{PUI} character; required: FALSE; MIAPPE V1.1 (DM-23) Type of experimental
design - Type of experimental design of the study, in the form of an
accession number from the Crop Ontology.
\item \code{description} character: required: FALSE; MIAPPE V1.1 (DM-22) Description
of the experimental design - Short description of the experimental
design, possibly including statistical design. In specific cases, e.g.
legacy datasets or data computed from several studies, the experimental
design can be "unknown"/"NA", "aggregated/reduced data", or simply 'none'.
}

The Examples section shows an example of how to construct the
\code{experimentalDesign} argument as a list.}

\item{externalReferences}{data.frame; required: FALSE; A data.frame of
external reference ids. These are references to this piece of data in an
external system. Could be a simple string or a URI. The \code{externalReferences}
argument data.frame should contain the following columns:
\itemize{
\item \code{referenceID} character; required: TRUE; The external reference ID. Could
be a simple string or a URI.
\item \code{referenceSource} character; required: TRUE; An identifier for the source
system or database of this reference.
}

The Examples section shows an example of how to construct the
\code{externalReferences} argument as a data.frame.}

\item{growthFacility}{list; required: FALSE; Short description of the
facility in which the study was carried out. The \code{growthFacility} list
argument can contain the following elements:
\itemize{
\item \code{PUI} character; required: FALSE; MIAPPE V1.1 (DM-27) Type of growth
facility - Type of growth facility in which the study was carried out,
in the form of an accession number from the Crop Ontology.
\item \code{description} charater; required: FALSE; MIAPPE V1.1 (DM-26) Description
of growth facility - Short description of the facility in which the
study was carried out.
}

The Examples section shows an example of how to construct the
\code{growthFacility} argument as a list.}

\item{lastUpdate}{list; required: FALSE; The date and time when this study
was last modified. The \code{lastUpdate} list argument can contain the
following elements:
\itemize{
\item \code{timestamp} character; required: FALSE; Coded in the ISO 8601 standard
extended format, where date, time and time zone information needs to be
provided (check for example
\url{https://www.w3.org/TR/NOTE-datetime}
).
\item \code{version} character; required: FALSE; Version number specifying the last
update.
}

The Examples section shows an example of how to construct the
\code{lastUpdate} argument as a list.}

\item{license}{character; required: FALSE; The usage license associated with
the study data.}

\item{locationDbId}{character; required: FALSE; The unique identifier for a
location.}

\item{locationName}{character; required: FALSE; A human readable name for
this location. MIAPPE V1.1 (DM-18) Experimental site name - The name of
the natural site, experimental field, greenhouse, phenotyping facility,
\emph{etc.} where the experiment took place.}

\item{observationLevels}{data.frame; required: FALSE; A data.frame of
Observation levels, which indicate the granularity level at which the
measurements are taken. The \code{observationLevels} argument data.frame can
contain the following columns:
\itemize{
\item \code{levelName} character; required: FALSE; A name for this level.
\item \code{levelOrder} interger; required: FALSE; \code{levelOrder} defines, where that
level exists in the hierarchy of levels. \code{levelOrder}'s lower numbers
are at the top of the hierarchy (i.e. field -> 1) and higher numbers are
at the bottom of the hierarchy (i.e. plant -> 9).
}

The Examples section shows an example of how to construct the
\code{observationLevels} argument as a data.frame.}

\item{observationUnitsDescription}{character; required: FALSE; MIAPPE V1.1
(DM-25) Observation unit description - General description of the observation
units in the study.}

\item{seasons}{vector of type character; required: FALSE; seasonDbId value(s)
over which this study was performed.; default: "", when using
multiple values supply as c("value1", "value2").}

\item{startDate}{character; required: FALSE; The date the study will start or
started. Coded in the ISO 8601 standard extended format, where date, time
and time zone information needs to be provided (check for example
\url{https://www.w3.org/TR/NOTE-datetime}
). MIAPPE V1.1 (DM-14) Start date of study - Date and, if relevant, time
when the experiment started.}

\item{studyCode}{character; required: FALSE; A short human readable code for
a study.}

\item{studyDescription}{character; required: FALSE; The description of this
study. MIAPPE V1.1 (DM-13) Study description - Human-readable text
describing the study.}

\item{studyName}{character; required: FALSE; The human readable name for a
study. MIAPPE V1.1 (DM-12) Study title - Human-readable text summarizing
the study.}

\item{studyPUI}{character; required: FALSE; A permanent unique identifier
associated with this study data. For example, a URI or DOI.}

\item{studyType}{character; required: FALSE; The type of study being
performed. ex. "Yield Trial", \emph{etc.}}

\item{trialDbId}{character; required: FALSE; The ID which uniquely identifies
a trial.}

\item{trialName}{character; required: FALSE; The human readable name of a
trial.}
}
\value{
data.frame
}
\description{
Create new Studies.
}
\details{
Create new studies, where the \code{studDbId} is generated by the server.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
con[["token"]] <- "YYYY"
active <- TRUE
additionalInfo <- list(additionalProp1 = "string",
                       additionalProp2 = "string",
                       additionalProp3 = "string")
commonCropName <- "Tomatillo"
contacts <- data.frame(contactDbId = "trial_contact_1",
                       email = "a.breeder@brapi.org",
                       instituteName = "Plant Science Institute",
                       name = "A. Breeder",
                       orcid = "http://orcid.org/0000-0002-0607-8728",
                       type = "Breeder")
culturalPractices <- "Irrigation was applied according needs during summer to
                      prevent water stress."
dataLinks <- data.frame(dataFormat = "Image Archive",
                        description = "Raw drone images collected for this study",
                        fileFormat = "application/zip",
                        name = "image-archive.zip",
                        provenance = "Image Processing Pipeline, built at the
                                      University of Antarctica:
                                      https://github.com/antarctica/pipeline",
                        scientificType = "Environmental",
                        url = "https://brapi.org/image-archive.zip",
                        version = "1.0.3")
documentationURL <- "https://wiki.brapi.org"
endDate <- "2021-04-19T12:34:13.494Z"
environmentParameters <- data.frame(description = "the soil type was clay",
                                    parameterName =  "soil type",
                                    parameterPUI = "PECO:0007155",
                                    unit = "pH",
                                    unitPUI = "PECO:0007059",
                                    value = "clay soil",
                                    valuePUI = "ENVO:00002262")
experimentalDesign <- list(PUI = "CO_715:0000145",
                           description = "Lines were repeated twice at each
                                          location using a complete block
                                          design. In order to limit competition
                                          effects, each block was organized into
                                          four sub-blocks corresponding to
                                          earliest groups based on a prior
                                          information.")
externalReferences <-
  data.frame(referenceID = c("doi:10.155454/12341234",
                             "http://purl.obolibrary.org/obo/ro.owl",
                             "75a50e76"),
             referenceSource = c("DOI",
                                 "OBO Library",
                                 "Remote Data Collection Upload Tool"))
growthFacility <- list(PUI = "CO_715:0000162",
                       description = "field environment condition, greenhouse")
lastUpdate <- list(timestamp = "2021-04-19T12:34:13.494Z",
                   version = "1.2.3")
license <- "MIT License"
locationDbId <- "location_01"
locationName <- "Location 1"
observationLevels <-
  data.frame(levelName = c("field", "block", "plot"),
             levelOrder = c(0, 1, 2))
observationUnitsDescription <- "Observation units consisted in individual
                                plots themselves consisting of a row of 15
                                plants at a density of approximately six
                                plants per square meter."
seasons <- c("spring_2012", "summer_2012")
startDate <- "2021-04-19T12:34:13.494Z"
studyCode <- "Tomatillo_Yield_Spring_2012"
studyDescription <- "This is a yield study for Spring 2012"
studyName <- "INRA's Tomatillo Genetic Resources Observation at Peru"
studyPUI <- "doi:10.155454/12349537312"
studyType <- "Phenotyping"
trialDbId <- "trial1"
trialName <- "Peru Yield Trial 1"

brapi_post_studies(con = con,
                   active = active,
                   additionalInfo = additionalInfo,
                   commonCropName = commonCropName,
                   contacts = contacts,
                   culturalPractices = culturalPractices,
                   dataLinks = dataLinks,
                   documentationURL = documentationURL,
                   endDate = endDate,
                   environmentParameters = environmentParameters,
                   experimentalDesign = experimentalDesign,
                   externalReferences = externalReferences,
                   growthFacility = growthFacility,
                   lastUpdate = lastUpdate,
                   license = license,
                   locationDbId = locationDbId,
                   locationName = locationName,
                   observationLevels = observationLevels,
                   observationUnitsDescription = observationUnitsDescription,
                   seasons = seasons,
                   startDate = startDate,
                   studyCode = studyCode,
                   studyDescription = studyDescription,
                   studyName = studyName,
                   studyPUI = studyPUI,
                   studyType = studyType,
                   trialDbId = trialDbId,
                   trialName = trialName)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Core/2.0#/Studies/post_studies }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-core: 
\code{\link{brapi_get_commoncropnames}()},
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations_locationDbId}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_programs_programDbId}()},
\code{\link{brapi_get_programs}()},
\code{\link{brapi_get_search_lists_searchResultsDbId}()},
\code{\link{brapi_get_search_locations_searchResultsDbId}()},
\code{\link{brapi_get_search_people_searchResultsDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_trials_searchResultsDbId}()},
\code{\link{brapi_get_seasons_seasonDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_serverinfo}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_get_trials_trialDbId}()},
\code{\link{brapi_get_trials}()},
\code{\link{brapi_post_lists_listDbId_items}()},
\code{\link{brapi_post_lists}()},
\code{\link{brapi_post_locations}()},
\code{\link{brapi_post_people}()},
\code{\link{brapi_post_programs}()},
\code{\link{brapi_post_search_lists}()},
\code{\link{brapi_post_search_locations}()},
\code{\link{brapi_post_seasons}()},
\code{\link{brapi_post_trials}()},
\code{\link{brapi_put_lists_listDbId}()},
\code{\link{brapi_put_locations_locationDbId}()},
\code{\link{brapi_put_people_personDbId}()},
\code{\link{brapi_put_programs_programDbId}()},
\code{\link{brapi_put_seasons_seasonDbId}()},
\code{\link{brapi_put_studies_studyDbId}()},
\code{\link{brapi_put_trials_trialDbId}()}

Other Studies: 
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_put_studies_studyDbId}()}
}
\author{
Maikel Verouden
}
\concept{Studies}
\concept{brapi-core}
