% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_post_studies_studyDbId_table.R
\name{brapi_post_studies_studyDbId_table}
\alias{brapi_post_studies_studyDbId_table}
\title{post /studies/\{studyDbId\}/table}
\usage{
brapi_post_studies_studyDbId_table(
  con = NULL,
  studyDbId = "",
  data = "",
  headerRow = "",
  observationVariableDbIds = ""
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{studyDbId}{character; required: TRUE; Identifier of the study. Usually a number, could be alphanumeric.}

\item{data}{data.frame; required: TRUE; data frame of observation data recorded for different observation variables across different observation units; default: "". The Examples Section shows an example of how to construct the data argument as a data.frame.}

\item{headerRow}{vector of type character; required: TRUE; The header row describing the data frame. Append 'observationVariableDbIds' for complete header row of the table. This vector of type character should contain any or all of the following strings; year, studyDbId, studyName, locationDbId, locationName, germplasmDbId, germplasmName, observationUnitDbId, plotNumber, replicate, blockNumber, entryType, X, Y; default: "", when using multiple values supply as c("year", "studyDbId", ...).}

\item{observationVariableDbIds}{vector of type character; required: TRUE; The list of observation variable identifiers which have values recorded for them in the data frame. Append to the 'headerRow' for complete header row; default: "", when using multiple values supply as c("value1", "value2").}
}
\value{
data.frame
}
\description{
Submit new Observations in a table format for a specific Study
}
\details{
This call can be used to create new observations in bulk.

Note: If you need to update any existing observation, please use \code{PUT /studies/{studyDbId}/observations}. This call should only be used to create \strong{NEW} observations.
}
\examples{
\dontrun{

# Create a connection object
con <- brapi_db()$testserver

# Create an example data.frame data object
data <- data.frame(
  year = c("2019", "2019", "2020", "2020"),
  studyDbId = c("1001", "1001", "1001", "1001"),
  studyName = c("Study 1", "Study 1", "Study 1", "Study 1"),
  locationDbId = c("1", "1", "1", "1"),
  locationName = c("Peru", "Peru", "Peru", "Peru"),
  germplasmDbId = c("1", "1", "2", "2"),
  germplasmName  = c("Name001", "Name001", "Name002", "Name002"),
  observationUnitDbId = c("1", "2", "3", "4"),
  plotNumber = c("1", "1", "2", "2"),
  replicate = c("0", "0", "0", "0"),
  blockNumber = c("1", "1", "1", "1"),
  entryType = c("FILLER", "TEST", "FILLER", "FILLER"),
  X = c("1", "1", "1", "1"),
  Y = c("1", "2", "3", "4"),
  'MO_123:100002' = c("1.0", "1.1", "1.2", "1.3"),
  'MO_123:100003' = c("2.1", "2.2", "2.3", "2.4"),
  'MO_123:100005' = c("light blue", "blue", "dark blue", "blue"),
  'MO_123:100006' = c("4.3", "4.2", "4.0", "4.1"))

# Create a headerRow character vector
headerRow <- c("year",
               "studyDbId",
               "studyName",
               "locationDbId",
               "locationName",
               "germplasmDbId",
               "germplasmName",
               "observationUnitDbId",
               "plotNumber",
               "replicate",
               "blockNumber",
               "entryType",
               "X",
               "Y")

# Create an observationVariableDbIds character vector
observationVariableDbIds <- c("MO_123:100002",
                              "MO_123:100003",
                              "MO_123:100005",
                              "MO_123:100006")

# Make post /studies/{studyDbId}/table call
brapi_post_studies_studyDbId_table(con = con,
                                   studyDbId = "1001",
                                   data = data,
                                   headerRow = headerRow,
                                   observationVariableDbIds = observationVariableDbIds)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI/1.3#/Studies/post_studies__studyDbId__table }{BrAPI SwaggerHub}
}
\seealso{
Other brapi_1.3: 
\code{\link{brapi_get_allelematrices_search}()},
\code{\link{brapi_get_allelematrices}()},
\code{\link{brapi_get_attributes_categories}()},
\code{\link{brapi_get_attributes}()},
\code{\link{brapi_get_breedingmethods_breedingMethodDbId}()},
\code{\link{brapi_get_breedingmethods}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_commoncropnames}()},
\code{\link{brapi_get_germplasm_germplasmDbId_attributes}()},
\code{\link{brapi_get_germplasm_germplasmDbId_markerprofiles}()},
\code{\link{brapi_get_germplasm_germplasmDbId_mcpd}()},
\code{\link{brapi_get_germplasm_germplasmDbId_pedigree}()},
\code{\link{brapi_get_germplasm_germplasmDbId_progeny}()},
\code{\link{brapi_get_germplasm_germplasmDbId}()},
\code{\link{brapi_get_germplasm}()},
\code{\link{brapi_get_images_imageDbId}()},
\code{\link{brapi_get_images}()},
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations_locationDbId}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_maps_mapDbId_positions_linkageGroupName}()},
\code{\link{brapi_get_maps_mapDbId_positions}()},
\code{\link{brapi_get_maps_mapDbId}()},
\code{\link{brapi_get_maps}()},
\code{\link{brapi_get_markerprofiles_markerProfileDbId}()},
\code{\link{brapi_get_markerprofiles}()},
\code{\link{brapi_get_markers_markerDbId}()},
\code{\link{brapi_get_markers}()},
\code{\link{brapi_get_methods_methodDbId}()},
\code{\link{brapi_get_methods}()},
\code{\link{brapi_get_observationlevels}()},
\code{\link{brapi_get_observationunits}()},
\code{\link{brapi_get_ontologies}()},
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_programs}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_scales_scaleDbId}()},
\code{\link{brapi_get_scales}()},
\code{\link{brapi_get_search_germplasm_searchResultsDbId}()},
\code{\link{brapi_get_search_images_searchResultsDbId}()},
\code{\link{brapi_get_search_markers_searchResultsDbId}()},
\code{\link{brapi_get_search_observationtables_searchResultsDbId}()},
\code{\link{brapi_get_search_observationunits_searchResultsDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_variables_searchResultsDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_studies_studyDbId_germplasm}()},
\code{\link{brapi_get_studies_studyDbId_layouts}()},
\code{\link{brapi_get_studies_studyDbId_observations}()},
\code{\link{brapi_get_studies_studyDbId_observationunits}()},
\code{\link{brapi_get_studies_studyDbId_observationvariables}()},
\code{\link{brapi_get_studies_studyDbId_table}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_get_traits_traitDbId}()},
\code{\link{brapi_get_traits}()},
\code{\link{brapi_get_trials_trialDbId}()},
\code{\link{brapi_get_trials}()},
\code{\link{brapi_get_variables_observationVariableDbId}()},
\code{\link{brapi_get_variables}()},
\code{\link{brapi_get_vendor_orders_orderId_plates}()},
\code{\link{brapi_get_vendor_orders_orderId_results}()},
\code{\link{brapi_get_vendor_orders_orderId_status}()},
\code{\link{brapi_get_vendor_orders}()},
\code{\link{brapi_get_vendor_plates_submissionId}()},
\code{\link{brapi_get_vendor_specifications}()},
\code{\link{brapi_post_allelematrices_search}()},
\code{\link{brapi_post_images}()},
\code{\link{brapi_post_lists_listDbId_items}()},
\code{\link{brapi_post_lists}()},
\code{\link{brapi_post_methods}()},
\code{\link{brapi_post_people}()},
\code{\link{brapi_post_phenotypes}()},
\code{\link{brapi_post_scales}()},
\code{\link{brapi_post_search_germplasm}()},
\code{\link{brapi_post_search_images}()},
\code{\link{brapi_post_search_markers}()},
\code{\link{brapi_post_search_observationtables}()},
\code{\link{brapi_post_search_observationunits}()},
\code{\link{brapi_post_search_programs}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_studies}()},
\code{\link{brapi_post_search_variables}()},
\code{\link{brapi_post_traits}()},
\code{\link{brapi_post_vendor_orders}()},
\code{\link{brapi_post_vendor_plates}()},
\code{\link{brapi_put_images_imageDbId_imagecontent}()},
\code{\link{brapi_put_images_imageDbId}()},
\code{\link{brapi_put_lists_listDbId}()},
\code{\link{brapi_put_methods_methodDbId}()}

Other Studies: 
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_studies_studyDbId_germplasm}()},
\code{\link{brapi_get_studies_studyDbId_layouts}()},
\code{\link{brapi_get_studies_studyDbId_observations}()},
\code{\link{brapi_get_studies_studyDbId_observationunits}()},
\code{\link{brapi_get_studies_studyDbId_observationvariables}()},
\code{\link{brapi_get_studies_studyDbId_table}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_post_search_studies}()}

Other Observations: 
\code{\link{brapi_get_observationlevels}()},
\code{\link{brapi_get_observationunits}()},
\code{\link{brapi_get_search_observationtables_searchResultsDbId}()},
\code{\link{brapi_get_search_observationunits_searchResultsDbId}()},
\code{\link{brapi_get_studies_studyDbId_observations}()},
\code{\link{brapi_get_studies_studyDbId_observationunits}()},
\code{\link{brapi_get_studies_studyDbId_table}()},
\code{\link{brapi_post_phenotypes}()},
\code{\link{brapi_post_search_observationtables}()},
\code{\link{brapi_post_search_observationunits}()}
}
\author{
brapir generator package
}
\concept{Observations}
\concept{Studies}
\concept{brapi_1.3}
