% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_post_trials.R
\name{brapi_post_trials}
\alias{brapi_post_trials}
\title{post /trials}
\usage{
brapi_post_trials(
  con = NULL,
  active = NA,
  additionalInfo = list(),
  commonCropName = "",
  contacts = "",
  datasetAuthorships = "",
  documentationURL = "",
  endDate = "",
  externalReferences = "",
  programDbId = "",
  programName = "",
  publications = "",
  startDate = "",
  trialDescription = "",
  trialName = "",
  trialPUI = ""
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{active}{logical; required: FALSE; Is this trial currently active;
default: NA, other possible values: TRUE | FALSE}

\item{additionalInfo}{list; required: FALSE; Additional arbitrary information.
If provided use the following construct list(additionalProp1 = "string",
additionalProp2 =  "string", additionalProp3 = "string").

The Examples section shows an example on how to construct the
\code{additionalInfo} argument as a list.}

\item{commonCropName}{character; required: FALSE; Common name for the crop
associated with this trial.}

\item{contacts}{data.frame; required: FALSE; A data.frame of contact entities
associated with this trial. The \code{contacts} argument data.frame should or
can contain the following columns:
\itemize{
\item \code{contactDbId} character; required: TRUE; The ID which uniquely
identifies this contact. MIAPPE V1.1 (DM-33) Person ID - An identifier
for the data submitter. If that submitter is an individual, ORCID
identifiers are recommended.
\item \code{email}    character; required: FALSE; The contacts email address. MIAPPE
V1.1 (DM-32) Person email - The electronic mail address of the person.
\item \code{instituteName} character; required: FALSE; The name of the institution
which this contact is part of. MIAPPE V1.1 (DM-35) Person affiliation -
The institution the person belongs to.
\item \code{name}    character; required: FALSE; The full name of this contact person.
MIAPPE V1.1 (DM-31) Person name - The name of the person (either full
name or as used in scientific publications).
\item \code{orcid} character; required: FALSE; The Open Researcher and Contributor
ID for this contact person (orcid.org). MIAPPE V1.1 (DM-33) Person ID -
An identifier for the data submitter. If that submitter is an individual,
ORCID identifiers are recommended.
\item \code{type} character;    required: FALSE; The type of person this contact
represents (ex: Coordinator, Scientist, PI, etc.). MIAPPE V1.1 (DM-34)
Person role - Type of contribution of the person to the investigation.
}

The Examples section shows an example of how to construct the
\code{contacts} argument as a data.frame.}

\item{datasetAuthorships}{data.frame; required: FALSE; A data.frame of
License and citation information for the data in this trial. The
\code{datasetAuthorships} argument data.frame can contain the following
columns:
\itemize{
\item \code{datasetPUI} character; required: FALSE; For example a DOI of a
publication in which the dataset is used.
\item \code{license} character; required: FALSE; MIAPPE V1.1 (DM-7) License -
License for the reuse of the data associated with this investigation.
The Creative Commons licenses cover most use cases and are recommended.
\item \code{publicReleaseDate} character; required: FALSE; MIAPPE V1.1 (DM-6)
Public release date - Date of first public release of the dataset
presently being described.
\item \code{submissionDate} character; required: FALSE; MIAPPE V1.1 (DM-5)
Submission date - Date of submission of the data set presently being
described to a host repository.
}

The Examples section shows an example of how to construct the
\code{datasetAuthorships} argument as a data.frame.}

\item{documentationURL}{character; required: FALSE; A URL to the human
readable documentation of this object.}

\item{endDate}{character; required: FALSE; The date the trial will end or
ended. Coded in the ISO 8601 standard extended format, where date,
time and time zone information needs to be provided (check for example
\url{https://www.w3.org/TR/NOTE-datetime}
).}

\item{externalReferences}{data.frame; required: FALSE; A data.frame of
external reference ids. These are references to this piece of data in an
external system. Could be a simple string or a URI. The \code{externalReferences}
argument data.frame should contain the following columns:
\itemize{
\item \code{referenceID} character; required: TRUE; The external reference ID. Could
be a simple string or a URI.
\item \code{referenceSource} character; required: TRUE; An identifier for the source
system or database of this reference.
}

The Examples section shows an example of how to construct the
\code{externalReferences} argument as a data.frame.}

\item{programDbId}{character; required: FALSE; A program identifier, to which
this trial belongs to. Exact match.}

\item{programName}{character; required: FALSE; Human readable name of the
program, to which this trial belongs to. Exact match.}

\item{publications}{data.frame; required: FALSE; A data.frame of publications
in which the trial is described. MIAPPE V1.1 (DM-9) Associated publication
\itemize{
\item An identifier for a literature publication where the investigation is
described. Use of DOIs is recommended. The \code{publications} data.frame
should contain the following columns:
\item \code{publicationPUI} character; required: FALSE; Preferably a DOI of the
publication
\item \code{publicationReference} character; required: FALSE;    A reference for the
publication, e.g. "Selby, BrAPI - An application programming interface
for plant breeding applications, Bioinformatics,
https://doi.org/10.1093/bioinformatics/190"
}

The Examples section shows an example of how to construct the
\code{publications} argument as a data.frame.}

\item{startDate}{character; required: FALSE; The date the trial will start or
started. Coded in the ISO 8601 standard extended format, where date, time
and time zone information needs to be provided (check for example
\url{https://www.w3.org/TR/NOTE-datetime}
).}

\item{trialDescription}{character; required: FALSE; The human readable
description of a trial. MIAPPE V1.1 (DM-4) Investigation description -
Human-readable text describing the investigation in more detail.}

\item{trialName}{character; required: FALS; The human readable name of a
trial. MIAPPE V1.1 (DM-3) Investigation title - Human-readable string
summarizing the investigation.}

\item{trialPUI}{character; required: FALSE; A permanent identifier for a
trial. Could be DOI or other URI formatted identifier.}
}
\value{
data.frame
}
\description{
Create new trials
}
\details{
Create new breeding Trials. A Trial represents a collection of
related Studies. \code{trialDbId} is generated by the server.
}
\examples{
\dontrun{
con <- brapi_db()$testserver
con[["token"]] <- "YYYY"
active <- TRUE
additionalInfo <- list(additionalProp1 = "string",
                       additionalProp2 = "string",
                       additionalProp3 = "string")
commonCropName <- "Tomatillo"
contacts <- data.frame(contactDbId = "trial_contact_1",
                       email = "a.breeder@brapi.org",
                       instituteName = "Plant Science Institute",
                       name = "A. Breeder",
                       orcid = "http://orcid.org/0000-0002-0607-8728",
                       type = "Breeder")
datasetAuthorships <-
  data.frame(datasetPUI = "doi:10.15454/312953986E3",
             license = "https://CreativeCommons.org/licenses/by/4.0/",
             publicReleaseDate = "2021-04-28",
             submissionDate = "2021-04-28")
documentationURL <- "https://wiki.brapi.org"
endDate <- "2021-04-28T12:34:13.494Z"
externalReferences <-
  data.frame(referenceID = c("doi:10.155454/12341234",
                             "http://purl.obolibrary.org/obo/ro.owl",
                             "75a50e76"),
             referenceSource = c("DOI",
                                 "OBO Library",
                                 "Remote Data Collection Upload Tool"))
programDbId <- "program2"
programName <- "The Fake Tomatillo Breeding Program"
publications <-
  data.frame(publicationPUI = "doi:10.15454/312953986E3",
             publicationReference = "Selby, BrAPI - An application programming interface for plant breeding applications, Bioinformatics, https://doi.org/10.1093/bioinformatics/190")
startDate <- "2021-04-28T12:34:13.494Z"
trialDescription <- "General drought resistance trial initiated in Peru before duplication in Africa"
trialName <- "Peru Yield Trial 1"
trialPUI <- "https://doi.org/101093190"

brapi_post_trials(con = con,
                  active = active,
                  additionalInfo = additionalInfo,
                  commonCropName = commonCropName,
                  contacts = contacts,
                  datasetAuthorships = datasetAuthorships,
                  documentationURL = documentationURL,
                  endDate = endDate,
                  externalReferences = externalReferences,
                  programDbId = programDbId,
                  programName = programName,
                  publications = publications,
                  startDate = startDate,
                  trialDescription = trialDescription,
                  trialName = trialName,
                  trialPUI = trialPUI)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Core/2.0#/Trials/post_trials }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-core: 
\code{\link{brapi_get_commoncropnames}()},
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations_locationDbId}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_programs_programDbId}()},
\code{\link{brapi_get_programs}()},
\code{\link{brapi_get_search_lists_searchResultsDbId}()},
\code{\link{brapi_get_search_locations_searchResultsDbId}()},
\code{\link{brapi_get_search_people_searchResultsDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_trials_searchResultsDbId}()},
\code{\link{brapi_get_seasons_seasonDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_serverinfo}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_get_trials_trialDbId}()},
\code{\link{brapi_get_trials}()},
\code{\link{brapi_post_lists_listDbId_items}()},
\code{\link{brapi_post_lists}()},
\code{\link{brapi_post_locations}()},
\code{\link{brapi_post_people}()},
\code{\link{brapi_post_programs}()},
\code{\link{brapi_post_search_lists}()},
\code{\link{brapi_post_seasons}()},
\code{\link{brapi_post_studies}()},
\code{\link{brapi_put_lists_listDbId}()},
\code{\link{brapi_put_locations_locationDbId}()},
\code{\link{brapi_put_people_personDbId}()},
\code{\link{brapi_put_programs_programDbId}()},
\code{\link{brapi_put_seasons_seasonDbId}()},
\code{\link{brapi_put_studies_studyDbId}()},
\code{\link{brapi_put_trials_trialDbId}()}

Other Trials: 
\code{\link{brapi_get_search_trials_searchResultsDbId}()},
\code{\link{brapi_get_trials_trialDbId}()},
\code{\link{brapi_get_trials}()},
\code{\link{brapi_put_trials_trialDbId}()}
}
\author{
Maikel Verouden
}
\concept{Trials}
\concept{brapi-core}
