% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_put_samples.R
\name{brapi_put_samples}
\alias{brapi_put_samples}
\title{put /samples}
\usage{
brapi_put_samples(
  con = NULL,
  germplasmDbId = "",
  notes = "",
  observationUnitDbId = "",
  plantDbId = "",
  plateDbId = "",
  plateIndex = 0,
  plotDbId = "",
  sampleDbId = "",
  sampleTimestamp = "",
  sampleType = "",
  studyDbId = "",
  takenBy = "",
  tissueType = ""
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{germplasmDbId}{character; required: FALSE;  The internal database identifier which uniquely identifies a germplasm; default: "".}

\item{notes}{character; required: FALSE; Additional notes about a sample; default: "".}

\item{observationUnitDbId}{character; required: FALSE; The internal database identifier, which uniquely identifies an observation unit; default: "".}

\item{plantDbId}{character; required: FALSE; The internal database identifier, which uniquely identifies a plant. Usually 'plantNumber'; default: "".}

\item{plateDbId}{character; required: FALSE; The internal database identifier, which uniquely identifies a plate of samples; default: "".}

\item{plateIndex}{integer; required: FALSE; The index number of this sample on a plate; default: 0}

\item{plotDbId}{character; required: FALSE;  The internal database indentifier, which uniquely identifies a plot. Usually 'plotNumber'; default: "".}

\item{sampleDbId}{character; required: FALSE; The internal database identifier, which uniquely identifies a sample; default: "".}

\item{sampleTimestamp}{character; required: FALSE; The date and time when the sample was collected from the field and specified in the ISO 8601 format; default: "".}

\item{sampleType}{character; required: FALSE; The type of sample taken. e.g. 'DNA', 'RNA', 'Tissue', \emph{etc.} ; default: "".}

\item{studyDbId}{character; required: FALSE; The internal database identifier, which uniquely identifies a study within the given database server; default: "".}

\item{takenBy}{character; required: FALSE; The name or identifier of the entity which took the sample from the field; default: "".}

\item{tissueType}{character; required: FALSE; The type of tissue sampled, e.g. 'Leaf', 'Root', \emph{etc.}; default: "".}
}
\value{
data.frame
}
\description{
Add a new Sample
}
\details{
Call to register the event of a sample being taken, where sampleDbId is assigned as a result of the operation and returned in response.
}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI/1.3#/Samples/put_samples }{BrAPI SwaggerHub}
}
\seealso{
Other brapi_1.3: 
\code{\link{brapi_get_allelematrices_search}()},
\code{\link{brapi_get_allelematrices}()},
\code{\link{brapi_get_attributes_categories}()},
\code{\link{brapi_get_attributes}()},
\code{\link{brapi_get_breedingmethods_breedingMethodDbId}()},
\code{\link{brapi_get_breedingmethods}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_commoncropnames}()},
\code{\link{brapi_get_germplasm_germplasmDbId_attributes}()},
\code{\link{brapi_get_germplasm_germplasmDbId_markerprofiles}()},
\code{\link{brapi_get_germplasm_germplasmDbId_mcpd}()},
\code{\link{brapi_get_germplasm_germplasmDbId_pedigree}()},
\code{\link{brapi_get_germplasm_germplasmDbId_progeny}()},
\code{\link{brapi_get_germplasm_germplasmDbId}()},
\code{\link{brapi_get_germplasm}()},
\code{\link{brapi_get_images_imageDbId}()},
\code{\link{brapi_get_images}()},
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations_locationDbId}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_maps_mapDbId_positions_linkageGroupName}()},
\code{\link{brapi_get_maps_mapDbId_positions}()},
\code{\link{brapi_get_maps_mapDbId}()},
\code{\link{brapi_get_maps}()},
\code{\link{brapi_get_markerprofiles_markerProfileDbId}()},
\code{\link{brapi_get_markerprofiles}()},
\code{\link{brapi_get_markers_markerDbId}()},
\code{\link{brapi_get_markers}()},
\code{\link{brapi_get_methods_methodDbId}()},
\code{\link{brapi_get_methods}()},
\code{\link{brapi_get_observationlevels}()},
\code{\link{brapi_get_observationunits}()},
\code{\link{brapi_get_ontologies}()},
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_programs}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_scales_scaleDbId}()},
\code{\link{brapi_get_scales}()},
\code{\link{brapi_get_search_germplasm_searchResultsDbId}()},
\code{\link{brapi_get_search_images_searchResultsDbId}()},
\code{\link{brapi_get_search_markers_searchResultsDbId}()},
\code{\link{brapi_get_search_observationtables_searchResultsDbId}()},
\code{\link{brapi_get_search_observationunits_searchResultsDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_variables_searchResultsDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_studies_studyDbId_germplasm}()},
\code{\link{brapi_get_studies_studyDbId_layouts}()},
\code{\link{brapi_get_studies_studyDbId_observations}()},
\code{\link{brapi_get_studies_studyDbId_observationunits}()},
\code{\link{brapi_get_studies_studyDbId_observationvariables}()},
\code{\link{brapi_get_studies_studyDbId_table}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_get_traits_traitDbId}()},
\code{\link{brapi_get_traits}()},
\code{\link{brapi_get_trials_trialDbId}()},
\code{\link{brapi_get_trials}()},
\code{\link{brapi_get_variables_observationVariableDbId}()},
\code{\link{brapi_get_variables}()},
\code{\link{brapi_get_vendor_orders_orderId_plates}()},
\code{\link{brapi_get_vendor_orders_orderId_results}()},
\code{\link{brapi_get_vendor_orders_orderId_status}()},
\code{\link{brapi_get_vendor_orders}()},
\code{\link{brapi_get_vendor_plates_submissionId}()},
\code{\link{brapi_get_vendor_specifications}()},
\code{\link{brapi_post_allelematrices_search}()},
\code{\link{brapi_post_images}()},
\code{\link{brapi_post_lists_listDbId_items}()},
\code{\link{brapi_post_lists}()},
\code{\link{brapi_post_methods}()},
\code{\link{brapi_post_people}()},
\code{\link{brapi_post_phenotypes}()},
\code{\link{brapi_post_scales}()},
\code{\link{brapi_post_search_germplasm}()},
\code{\link{brapi_post_search_images}()},
\code{\link{brapi_post_search_markers}()},
\code{\link{brapi_post_search_observationtables}()},
\code{\link{brapi_post_search_observationunits}()},
\code{\link{brapi_post_search_programs}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_studies}()},
\code{\link{brapi_post_search_variables}()},
\code{\link{brapi_post_studies_studyDbId_table}()},
\code{\link{brapi_post_traits}()},
\code{\link{brapi_post_vendor_orders}()},
\code{\link{brapi_post_vendor_plates}()},
\code{\link{brapi_put_images_imageDbId_imagecontent}()},
\code{\link{brapi_put_images_imageDbId}()},
\code{\link{brapi_put_lists_listDbId}()},
\code{\link{brapi_put_methods_methodDbId}()},
\code{\link{brapi_put_people_personDbId}()},
\code{\link{brapi_put_scales_scaleDbId}()},
\code{\link{brapi_put_studies_studyDbId_layouts}()}

Other Samples: 
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_post_search_samples}()}
}
\author{
brapir generator package
}
\concept{Samples}
\concept{brapi_1.3}
