% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_put_studies_studyDbId_layouts.R
\name{brapi_put_studies_studyDbId_layouts}
\alias{brapi_put_studies_studyDbId_layouts}
\title{put /studies/\{studyDbId\}/layouts}
\usage{
brapi_put_studies_studyDbId_layouts(con = NULL, studyDbId = "", layout = "")
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{studyDbId}{character; required: TRUE; Identifier of the study. Usually a number, could be alphanumeric.}

\item{layout}{data.frame; required: TRUE, with default: ""; data.frame of observation unit position data entities which need to be updated. Each row in the \code{layout} argument data .frame corresponds to a single entity and requires to contain the following columns:
\itemize{
\item \code{blockNumber} vector of type character; to specify the blocknumber positions of each observation unit
\item \code{entryType} vector of type character; to specify the entrytype of each observation unit with possible values: "CHECK" | "FILLER" | "TEST
\item \code{observationUnitDbId} vector of type character; to specify the internal database identifier of each observation unit.
\item \code{positionCoordinateX} vector of type character; to specify the X position coordinate for each observation unit. Different systems may use different coordinate systems.
\item \code{positionCoordinateXType} vector of type character; to specify the type of X positional coordinate used for each observation unit. Must be one of the following values:
\itemize{
\item \code{LONGITUDE} ISO 6709 standard, WGS84 geodetic datum. See "Location Coordinate Encoding" for details
\item \code{LATITUDE} ISO 6709 standard, WGS84 geodetic datum. See "Location Coordinate Encoding" for details
\item \code{PLANTED_ROW} The physical planted row number
\item \code{PLANTED_INDIVIDUAl} The physical counted number, could be independEnt or within a planted row
\item \code{GRID_ROW} The row index number of a square grid overlay
\item \code{GRID_COL} The column index number of a square grid overlay
\item \code{MEASURED_ROW} The distance in meters from a defined 0th row
\item \code{MEASURED_COL} The distance in meters from a defined 0th column
}
\item \code{positionCoordinateY} vector of type character; to specify the Y position coordinate for each observation unit. Different systems may use different coordinate systems.
\item \code{positionCoordinateYType} vector of type character; to specify the type of positional Y coordinate used for each observation unit. Must be one of the following values:
\itemize{
\item \code{LONGITUDE} ISO 6709 standard, WGS84 geodetic datum. See "Location Coordinate Encoding" for details
\item \code{LATITUDE} ISO 6709 standard, WGS84 geodetic datum. See "Location Coordinate Encoding" for details
\item \code{PLANTED_ROW} The physical planted row number
\item \code{PLANTED_INDIVIDUAl} The physical counted number, could be independEnt or within a planted row
\item \code{GRID_ROW} The row index number of a square grid overlay
\item \code{GRID_COL} The column index number of a square grid overlay
\item \code{MEASURED_ROW} The distance in meters from a defined 0th row
\item \code{MEASURED_COL} The distance in meters from a defined 0th column
}
\item \code{replicate} vector of type character; to specify the replicate number of each observation unit.
}

The Examples Section shows an example of how to construct the \code{layout} argument as a data.frame object.}
}
\value{
data.frame
}
\description{
Update an existing Study with new layout details
}
\details{
Update the layout data for a set of observation units within a study. A layout data.frame is a subset of fields within an observationUnit. If any of the fields in the request data.frame (as defined in the \code{layout} function argument) is missing, that piece of data will not be updated. The response should match the structure of the response from \verb{GET studies/\{studyDbId\}/layout} as implemented in the \code{brapi_get_studies_studyDbId_layout()} function, but it should only contain the layout objects which have been updated by the PUT request.
}
\examples{
\dontrun{

# Create a connection object
con <- brapi_db()$testserver

# Create a layout data.frame
# containing the fields to be updated
layout <- data.frame(
  blockNumber = c("12"),
  entryType = c("CHECK"),
  observationUnitDbId = c("1"),
  positionCoordinateX = c("12"),
  positionCoordinateXType = c("GRID_COL"),
  positionCoordinateY = c("12"),
  positionCoordinateYType = c("GRID_ROW"),
  replicate = c("1"))

# Make put /studies/{studyDbId}/layout call
brapi_put_studies_studyDbId_layouts(con = con, studyDbId = '1001', layout = layout)
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI/1.3#/Studies/put_studies__studyDbId__layouts }{BrAPI SwaggerHub}
}
\seealso{
Other brapi_1.3: 
\code{\link{brapi_get_allelematrices_search}()},
\code{\link{brapi_get_allelematrices}()},
\code{\link{brapi_get_attributes_categories}()},
\code{\link{brapi_get_attributes}()},
\code{\link{brapi_get_breedingmethods_breedingMethodDbId}()},
\code{\link{brapi_get_breedingmethods}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_commoncropnames}()},
\code{\link{brapi_get_germplasm_germplasmDbId_attributes}()},
\code{\link{brapi_get_germplasm_germplasmDbId_markerprofiles}()},
\code{\link{brapi_get_germplasm_germplasmDbId_mcpd}()},
\code{\link{brapi_get_germplasm_germplasmDbId_pedigree}()},
\code{\link{brapi_get_germplasm_germplasmDbId_progeny}()},
\code{\link{brapi_get_germplasm_germplasmDbId}()},
\code{\link{brapi_get_germplasm}()},
\code{\link{brapi_get_images_imageDbId}()},
\code{\link{brapi_get_images}()},
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations_locationDbId}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_maps_mapDbId_positions_linkageGroupName}()},
\code{\link{brapi_get_maps_mapDbId_positions}()},
\code{\link{brapi_get_maps_mapDbId}()},
\code{\link{brapi_get_maps}()},
\code{\link{brapi_get_markerprofiles_markerProfileDbId}()},
\code{\link{brapi_get_markerprofiles}()},
\code{\link{brapi_get_markers_markerDbId}()},
\code{\link{brapi_get_markers}()},
\code{\link{brapi_get_methods_methodDbId}()},
\code{\link{brapi_get_methods}()},
\code{\link{brapi_get_observationlevels}()},
\code{\link{brapi_get_observationunits}()},
\code{\link{brapi_get_ontologies}()},
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_programs}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_scales_scaleDbId}()},
\code{\link{brapi_get_scales}()},
\code{\link{brapi_get_search_germplasm_searchResultsDbId}()},
\code{\link{brapi_get_search_images_searchResultsDbId}()},
\code{\link{brapi_get_search_markers_searchResultsDbId}()},
\code{\link{brapi_get_search_observationtables_searchResultsDbId}()},
\code{\link{brapi_get_search_observationunits_searchResultsDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_variables_searchResultsDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_studies_studyDbId_germplasm}()},
\code{\link{brapi_get_studies_studyDbId_layouts}()},
\code{\link{brapi_get_studies_studyDbId_observations}()},
\code{\link{brapi_get_studies_studyDbId_observationunits}()},
\code{\link{brapi_get_studies_studyDbId_observationvariables}()},
\code{\link{brapi_get_studies_studyDbId_table}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_get_traits_traitDbId}()},
\code{\link{brapi_get_traits}()},
\code{\link{brapi_get_trials_trialDbId}()},
\code{\link{brapi_get_trials}()},
\code{\link{brapi_get_variables_observationVariableDbId}()},
\code{\link{brapi_get_variables}()},
\code{\link{brapi_get_vendor_orders_orderId_plates}()},
\code{\link{brapi_get_vendor_orders_orderId_results}()},
\code{\link{brapi_get_vendor_orders_orderId_status}()},
\code{\link{brapi_get_vendor_orders}()},
\code{\link{brapi_get_vendor_plates_submissionId}()},
\code{\link{brapi_get_vendor_specifications}()},
\code{\link{brapi_post_allelematrices_search}()},
\code{\link{brapi_post_images}()},
\code{\link{brapi_post_lists_listDbId_items}()},
\code{\link{brapi_post_lists}()},
\code{\link{brapi_post_methods}()},
\code{\link{brapi_post_people}()},
\code{\link{brapi_post_phenotypes}()},
\code{\link{brapi_post_scales}()},
\code{\link{brapi_post_search_germplasm}()},
\code{\link{brapi_post_search_images}()},
\code{\link{brapi_post_search_markers}()},
\code{\link{brapi_post_search_observationtables}()},
\code{\link{brapi_post_search_observationunits}()},
\code{\link{brapi_post_search_programs}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_studies}()},
\code{\link{brapi_post_search_variables}()},
\code{\link{brapi_post_studies_studyDbId_table}()},
\code{\link{brapi_post_traits}()},
\code{\link{brapi_post_vendor_orders}()},
\code{\link{brapi_post_vendor_plates}()},
\code{\link{brapi_put_images_imageDbId_imagecontent}()},
\code{\link{brapi_put_images_imageDbId}()},
\code{\link{brapi_put_lists_listDbId}()},
\code{\link{brapi_put_methods_methodDbId}()},
\code{\link{brapi_put_people_personDbId}()},
\code{\link{brapi_put_samples}()},
\code{\link{brapi_put_scales_scaleDbId}()},
\code{\link{brapi_put_studies_studyDbId_observations}()}

Other Studies: 
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_studies_studyDbId_germplasm}()},
\code{\link{brapi_get_studies_studyDbId_layouts}()},
\code{\link{brapi_get_studies_studyDbId_observations}()},
\code{\link{brapi_get_studies_studyDbId_observationunits}()},
\code{\link{brapi_get_studies_studyDbId_observationvariables}()},
\code{\link{brapi_get_studies_studyDbId_table}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_post_search_studies}()},
\code{\link{brapi_post_studies_studyDbId_table}()},
\code{\link{brapi_put_studies_studyDbId_observations}()}
}
\author{
brapir generator package
}
\concept{Studies}
\concept{brapi_1.3}
