% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_put_studies_studyDbId_observationunits.R
\name{brapi_put_studies_studyDbId_observationunits}
\alias{brapi_put_studies_studyDbId_observationunits}
\title{put /studies/\{studyDbId\}/observationunits}
\usage{
brapi_put_studies_studyDbId_observationunits(
  con = NULL,
  studyDbId = "",
  observationUnits = "",
  observationUnitXrefList = list(),
  treatmentList = list(),
  observationsList = list(),
  seasonList = list()
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{studyDbId}{character; required: TRUE; The internal study database identifier these observation units are related to.}

\item{observationUnits}{data.frame; required: TRUE; data.frame of observation units. Each row in the \code{observationUnits} data.frame corresponds to a single observation unit. The following columns are required in the \code{observationUnits} argument data.frame:
\itemize{
\item \code{blockNumber} vector of type character; The block number for each observation unit. Different systems may use different block designs.
\item \code{entryNumber} vector of type character; The entry number for each observation unit. Different systems may use different entry systems.
\item \code{entryType} vector of type character; The type of entry for each observation unit with possible values: "CHECK" | "FILLER" | "TEST"
\item \code{germplasmDbId} vector of type character; The internal database identifier of a germplasm for each observation unit
\item \code{observationLevel} vector of type character; The level of each observation unit, e.g. "plot", "plant"
\item \code{observationUnitDbId} vector of type character; The internal database identifier for each observation unit in case of updating
\item \code{observationUnitName} vector of type character; A human readable name for each observation unit
\item \code{plantNumber} vector of type character; The plant number in a field. Applicable for observationLevel: "plant"
\item \code{plotNumber} vector of type character: The plot number in a field. Applicable for observationLevel: "plot"
\item \code{positionCoordinateX} vector of type character; to specify the X position coordinate for each observation unit. Different systems may use different coordinate systems.
\item \code{positionCoordinateXType} vector of type character; to specify the type of X positional coordinate used for each observation unit. Must be one of the following values:
\itemize{
\item \code{LONGITUDE} ISO 6709 standard, WGS84 geodetic datum. See "Location Coordinate Encoding" for details
\item \code{LATITUDE} ISO 6709 standard, WGS84 geodetic datum. See "Location Coordinate Encoding" for details
\item \code{PLANTED_ROW} The physical planted row number
\item \code{PLANTED_INDIVIDUAl} The physical counted number, could be independEnt or within a planted row
\item \code{GRID_ROW} The row index number of a square grid overlay
\item \code{GRID_COL} The column index number of a square grid overlay
\item \code{MEASURED_ROW} The distance in meters from a defined 0th row
\item \code{MEASURED_COL} The distance in meters from a defined 0th column
}
\item \code{positionCoordinateY} vector of type character; to specify the Y position coordinate for each observation unit. Different systems may use different coordinate systems.
\item \code{positionCoordinateYType} vector of type character; to specify the type of positional Y coordinate used for each observation unit. Must be one of the following values:
\itemize{
\item \code{LONGITUDE} ISO 6709 standard, WGS84 geodetic datum. See "Location Coordinate Encoding" for details
\item \code{LATITUDE} ISO 6709 standard, WGS84 geodetic datum. See "Location Coordinate Encoding" for details
\item \code{PLANTED_ROW} The physical planted row number
\item \code{PLANTED_INDIVIDUAl} The physical counted number, could be independEnt or within a planted row
\item \code{GRID_ROW} The row index number of a square grid overlay
\item \code{GRID_COL} The column index number of a square grid overlay
\item \code{MEASURED_ROW} The distance in meters from a defined 0th row
\item \code{MEASURED_COL} The distance in meters from a defined 0th column
}
\item \code{replicate} vector of type character; To specify the replicate number of each observation unit. May be the same as blockNumber.
\item \code{studyDbId} vector of type character; The internal study database identifier each observation unit is related to. Must be the same as provided as the \code{studyDbId} argument in the function.
}

The Examples Section shows an example of how to construct the \code{observationUnits} argument as a data.frame object.}
}
\value{
data.frame
}
\description{
Update the Observation Units or upload new Observation Units for a specific Study
}
\details{
Use this call for updating the Observation Units or uploading new Observation Units for a specific Study. The following rules are used:
\itemize{
\item If \code{observationUnitDbId} or \code{observationDbId} is populated, they should be considered updates to existing records.
\item If an existing record of a \code{observationUnitDbId} or \code{observationDbId} is not found, the entered data should be treated as new records and assign new internal database identifiers.
\item If \code{observationUnitDbId} or \code{observationDbId} is un-populated (empty string or null) the entered data should be treated as new records and assig new internal database identifiers.
}
}
\examples{
\dontrun{

# Create a connection object
con <- brapi_db()$testserver
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI/1.3#/Studies/put_studies__studyDbId__observationunits }{BrAPI SwaggerHub}
}
\seealso{
Other brapi_1.3: 
\code{\link{brapi_get_allelematrices_search}()},
\code{\link{brapi_get_allelematrices}()},
\code{\link{brapi_get_attributes_categories}()},
\code{\link{brapi_get_attributes}()},
\code{\link{brapi_get_breedingmethods_breedingMethodDbId}()},
\code{\link{brapi_get_breedingmethods}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_commoncropnames}()},
\code{\link{brapi_get_germplasm_germplasmDbId_attributes}()},
\code{\link{brapi_get_germplasm_germplasmDbId_markerprofiles}()},
\code{\link{brapi_get_germplasm_germplasmDbId_mcpd}()},
\code{\link{brapi_get_germplasm_germplasmDbId_pedigree}()},
\code{\link{brapi_get_germplasm_germplasmDbId_progeny}()},
\code{\link{brapi_get_germplasm_germplasmDbId}()},
\code{\link{brapi_get_germplasm}()},
\code{\link{brapi_get_images_imageDbId}()},
\code{\link{brapi_get_images}()},
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations_locationDbId}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_maps_mapDbId_positions_linkageGroupName}()},
\code{\link{brapi_get_maps_mapDbId_positions}()},
\code{\link{brapi_get_maps_mapDbId}()},
\code{\link{brapi_get_maps}()},
\code{\link{brapi_get_markerprofiles_markerProfileDbId}()},
\code{\link{brapi_get_markerprofiles}()},
\code{\link{brapi_get_markers_markerDbId}()},
\code{\link{brapi_get_markers}()},
\code{\link{brapi_get_methods_methodDbId}()},
\code{\link{brapi_get_methods}()},
\code{\link{brapi_get_observationlevels}()},
\code{\link{brapi_get_observationunits}()},
\code{\link{brapi_get_ontologies}()},
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_programs}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_scales_scaleDbId}()},
\code{\link{brapi_get_scales}()},
\code{\link{brapi_get_search_germplasm_searchResultsDbId}()},
\code{\link{brapi_get_search_images_searchResultsDbId}()},
\code{\link{brapi_get_search_markers_searchResultsDbId}()},
\code{\link{brapi_get_search_observationtables_searchResultsDbId}()},
\code{\link{brapi_get_search_observationunits_searchResultsDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_variables_searchResultsDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_studies_studyDbId_germplasm}()},
\code{\link{brapi_get_studies_studyDbId_layouts}()},
\code{\link{brapi_get_studies_studyDbId_observations}()},
\code{\link{brapi_get_studies_studyDbId_observationunits}()},
\code{\link{brapi_get_studies_studyDbId_observationvariables}()},
\code{\link{brapi_get_studies_studyDbId_table}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_get_traits_traitDbId}()},
\code{\link{brapi_get_traits}()},
\code{\link{brapi_get_trials_trialDbId}()},
\code{\link{brapi_get_trials}()},
\code{\link{brapi_get_variables_observationVariableDbId}()},
\code{\link{brapi_get_variables}()},
\code{\link{brapi_get_vendor_orders_orderId_plates}()},
\code{\link{brapi_get_vendor_orders_orderId_results}()},
\code{\link{brapi_get_vendor_orders_orderId_status}()},
\code{\link{brapi_get_vendor_orders}()},
\code{\link{brapi_get_vendor_plates_submissionId}()},
\code{\link{brapi_get_vendor_specifications}()},
\code{\link{brapi_post_allelematrices_search}()},
\code{\link{brapi_post_images}()},
\code{\link{brapi_post_lists_listDbId_items}()},
\code{\link{brapi_post_lists}()},
\code{\link{brapi_post_methods}()},
\code{\link{brapi_post_people}()},
\code{\link{brapi_post_phenotypes}()},
\code{\link{brapi_post_scales}()},
\code{\link{brapi_post_search_germplasm}()},
\code{\link{brapi_post_search_images}()},
\code{\link{brapi_post_search_markers}()},
\code{\link{brapi_post_search_observationtables}()},
\code{\link{brapi_post_search_observationunits}()},
\code{\link{brapi_post_search_programs}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_studies}()},
\code{\link{brapi_post_search_variables}()},
\code{\link{brapi_post_studies_studyDbId_table}()},
\code{\link{brapi_post_traits}()},
\code{\link{brapi_post_vendor_orders}()},
\code{\link{brapi_post_vendor_plates}()},
\code{\link{brapi_put_images_imageDbId_imagecontent}()},
\code{\link{brapi_put_images_imageDbId}()},
\code{\link{brapi_put_lists_listDbId}()},
\code{\link{brapi_put_methods_methodDbId}()},
\code{\link{brapi_put_people_personDbId}()},
\code{\link{brapi_put_samples}()},
\code{\link{brapi_put_scales_scaleDbId}()},
\code{\link{brapi_put_studies_studyDbId_layouts}()},
\code{\link{brapi_put_studies_studyDbId_observations}()},
\code{\link{brapi_put_traits_traitDbId}()}

Other Studies: 
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_studies_studyDbId_germplasm}()},
\code{\link{brapi_get_studies_studyDbId_layouts}()},
\code{\link{brapi_get_studies_studyDbId_observations}()},
\code{\link{brapi_get_studies_studyDbId_observationunits}()},
\code{\link{brapi_get_studies_studyDbId_observationvariables}()},
\code{\link{brapi_get_studies_studyDbId_table}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_post_search_studies}()},
\code{\link{brapi_post_studies_studyDbId_table}()},
\code{\link{brapi_put_studies_studyDbId_layouts}()},
\code{\link{brapi_put_studies_studyDbId_observations}()}

Other Observations: 
\code{\link{brapi_get_observationlevels}()},
\code{\link{brapi_get_observationunits}()},
\code{\link{brapi_get_search_observationtables_searchResultsDbId}()},
\code{\link{brapi_get_search_observationunits_searchResultsDbId}()},
\code{\link{brapi_get_studies_studyDbId_observations}()},
\code{\link{brapi_get_studies_studyDbId_observationunits}()},
\code{\link{brapi_get_studies_studyDbId_table}()},
\code{\link{brapi_post_phenotypes}()},
\code{\link{brapi_post_search_observationtables}()},
\code{\link{brapi_post_search_observationunits}()},
\code{\link{brapi_post_studies_studyDbId_table}()},
\code{\link{brapi_put_studies_studyDbId_observations}()}
}
\author{
brapir generator package
}
\concept{Observations}
\concept{Studies}
\concept{brapi_1.3}
