% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_serverinfo_expiration.R
\name{brapi_serverinfo_expiration}
\alias{brapi_serverinfo_expiration}
\title{brapi_serverinfo_expiration}
\usage{
brapi_serverinfo_expiration(con)
}
\arguments{
\item{con}{list; required: \strong{TRUE}; BrAPI connection object of class \code{brapi_con}.}
}
\description{
Display the expiration date and time of token if available.
}
\examples{
\dontrun{
if (interactive()) {
  con <- brapi_db()$testserver
  brapi_serverinfo_expiration(con)
}
}

}
\seealso{
Other brapiutils: 
\code{\link{as.brapi_db}()},
\code{\link{brapi_bms_login}()},
\code{\link{brapi_canInternet}()},
\code{\link{brapi_checkCon}()},
\code{\link{brapi_connect}()},
\code{\link{brapi_db}()},
\code{\link{brapi_serverinfo_show}()},
\code{\link{is.brapi_con}()},
\code{\link{print.brapi_con}()},
\code{\link{print.brapi_db}()}
}
\author{
Maikel Verouden
}
\concept{brapiutils}
