% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.brapi_con.R
\name{is.brapi_con}
\alias{is.brapi_con}
\title{is.brapi_con}
\usage{
is.brapi_con(obj)
}
\arguments{
\item{obj}{object to be tested}
}
\value{
boolean (TRUE of FALSE)
}
\description{
Check whether an object is of class brapi_con
}
\seealso{
Other brapiutils: 
\code{\link{as.brapi_db}()},
\code{\link{brapi_canInternet}()},
\code{\link{brapi_checkCon}()},
\code{\link{brapi_connect}()},
\code{\link{brapi_db}()},
\code{\link{brapi_login}()},
\code{\link{brapi_serverinfo_show}()},
\code{\link{print.brapi_con}()},
\code{\link{print.brapi_db}()}
}
\author{
brapir generator package
}
\concept{brapiutils}
