% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_get_observations_observationDbId.R
\name{brapi_get_observations_observationDbId}
\alias{brapi_get_observations_observationDbId}
\title{get /observations/\{observationDbId\}}
\usage{
brapi_get_observations_observationDbId(con = NULL, observationDbId = "")
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{observationDbId}{character; required: TRUE; Unique observation
identifier.}
}
\value{
data.frame
}
\description{
Get the details of a specific Observations
}
\details{
Get the details of a specific Observations
}
\examples{
\dontrun{
con <- brapi_db()$testserver
brapi_get_observations_observationDbId(con = con,
                                       observationDbId = "observation1")
}

}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI-Phenotyping/2.0#/Observations/get_observations__observationDbId_ }{BrAPI SwaggerHub}
}
\seealso{
Other brapi-phenotyping: 
\code{\link{brapi_get_events}()},
\code{\link{brapi_get_images_imageDbId}()},
\code{\link{brapi_get_images}()},
\code{\link{brapi_get_methods_methodDbId}()},
\code{\link{brapi_get_methods}()},
\code{\link{brapi_get_observationlevels}()},
\code{\link{brapi_get_observations}()},
\code{\link{brapi_get_ontologies}()},
\code{\link{brapi_get_scales_scaleDbId}()},
\code{\link{brapi_get_scales}()},
\code{\link{brapi_get_traits_traitDbId}()},
\code{\link{brapi_get_traits}()},
\code{\link{brapi_get_variables_observationVariableDbId}()},
\code{\link{brapi_get_variables}()},
\code{\link{brapi_post_observations}()}

Other Observations: 
\code{\link{brapi_get_observations}()},
\code{\link{brapi_post_observations}()}
}
\author{
Maikel Verouden
}
\concept{Observations}
\concept{brapi-phenotyping}
