% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_post_search_studies.R
\name{brapi_post_search_studies}
\alias{brapi_post_search_studies}
\title{post /search/studies}
\usage{
brapi_post_search_studies(
  con = NULL,
  active = NA,
  commonCropNames = "",
  germplasmDbIds = "",
  locationDbIds = "",
  observationVariableDbIds = "",
  page = 0,
  pageSize = 1000,
  programDbIds = "",
  programNames = "",
  seasonDbIds = "",
  sortBy = "",
  sortOrder = "",
  studyDbIds = "",
  studyNames = "",
  studyTypeDbIds = "",
  studyTypeNames = "",
  trialDbIds = ""
)
}
\arguments{
\item{con}{list; required: TRUE; BrAPI connection object}

\item{active}{logical; required: FALSE; Is this study currently active; default: NA, other possible values: TRUE | FALSE}

\item{commonCropNames}{vector of type character; required: FALSE; Common names for the crop associated with this study; default: "", when using multiple values supply as c("value1", "value2").}

\item{germplasmDbIds}{vector of type character; required: FALSE; List of IDs which uniquely identify germplasm; default: "", when using multiple values supply as c("value1", "value2").}

\item{locationDbIds}{vector of type character; required: FALSE; List of location names to filter search results; default: "", when using multiple values supply as c("value1", "value2").}

\item{observationVariableDbIds}{vector of type character; required: FALSE; List of observation variable IDs to search for; default: "", when using multiple values supply as c("value1", "value2").}

\item{page}{integer; required: FALSE; Which page of the "data" array to return. The page indexing starts at 0 (page=0 will return the first page). Default is 0.}

\item{pageSize}{integer; required: FALSE; The maximum number of items to return per page of the "data" array. Default is 1000.}

\item{programDbIds}{vector of type character; required: FALSE; List of program identifiers to filter search results; default: "", when using multiple values supply as c("value1", "value2").}

\item{programNames}{vector of type character; required: FALSE; List of program names to filter search results; default: "", when using multiple values supply as c("value1", "value2").}

\item{seasonDbIds}{vector of type character; required: FALSE; The ID which uniquely identifies a season; default: "", when using multiple values supply as c("value1", "value2").}

\item{sortBy}{character; required: FALSE; Name of one of the fields within the study object on which results can be sorted; default: "".}

\item{sortOrder}{character; required: FALSE; Order results should be sorted. ex. "ASC" or "DESC"; default: "".}

\item{studyDbIds}{vector of type character; required: FALSE; List of study identifiers to search for; default: "", when using multiple values supply as c("value1", "value2").}

\item{studyNames}{vector of type character; required: FALSE; List of study names to filter search results; default: "", when using multiple values supply as c("value1", "value2").}

\item{studyTypeDbIds}{vector of type character; required: FALSE; The unique identifier of the type of study being performed.; default: "", when using multiple values supply as c("value1", "value2").}

\item{studyTypeNames}{vector of type character; required: FALSE; The name of the type of study being performed. ex. "Yield Trial", etc; default: "", when using multiple values supply as c("value1", "value2").}

\item{trialDbIds}{vector of type character; required: FALSE; List of trial identifiers to filter search results; default: "", when using multiple values supply as c("value1", "value2").}
}
\value{
data.frame
}
\description{
Submit a search request for Studies
}
\details{
Get list of studies
StartDate and endDate should be ISO8601 format for dates
See Search Services for additional implementation details.
}
\references{
\href{https://app.swaggerhub.com/apis/PlantBreedingAPI/BrAPI/1.3#/Studies/post_search_studies }{BrAPI SwaggerHub}
}
\seealso{
Other brapi_1.3: 
\code{\link{brapi_get_allelematrices_search}()},
\code{\link{brapi_get_allelematrices}()},
\code{\link{brapi_get_attributes_categories}()},
\code{\link{brapi_get_attributes}()},
\code{\link{brapi_get_breedingmethods_breedingMethodDbId}()},
\code{\link{brapi_get_breedingmethods}()},
\code{\link{brapi_get_calls}()},
\code{\link{brapi_get_commoncropnames}()},
\code{\link{brapi_get_germplasm_germplasmDbId_attributes}()},
\code{\link{brapi_get_germplasm_germplasmDbId_markerprofiles}()},
\code{\link{brapi_get_germplasm_germplasmDbId_mcpd}()},
\code{\link{brapi_get_germplasm_germplasmDbId_pedigree}()},
\code{\link{brapi_get_germplasm_germplasmDbId_progeny}()},
\code{\link{brapi_get_germplasm_germplasmDbId}()},
\code{\link{brapi_get_germplasm}()},
\code{\link{brapi_get_images_imageDbId}()},
\code{\link{brapi_get_images}()},
\code{\link{brapi_get_lists_listDbId}()},
\code{\link{brapi_get_lists}()},
\code{\link{brapi_get_locations_locationDbId}()},
\code{\link{brapi_get_locations}()},
\code{\link{brapi_get_maps_mapDbId_positions_linkageGroupName}()},
\code{\link{brapi_get_maps_mapDbId_positions}()},
\code{\link{brapi_get_maps_mapDbId}()},
\code{\link{brapi_get_maps}()},
\code{\link{brapi_get_markerprofiles_markerProfileDbId}()},
\code{\link{brapi_get_markerprofiles}()},
\code{\link{brapi_get_markers_markerDbId}()},
\code{\link{brapi_get_markers}()},
\code{\link{brapi_get_methods_methodDbId}()},
\code{\link{brapi_get_methods}()},
\code{\link{brapi_get_observationlevels}()},
\code{\link{brapi_get_observationunits}()},
\code{\link{brapi_get_ontologies}()},
\code{\link{brapi_get_people_personDbId}()},
\code{\link{brapi_get_people}()},
\code{\link{brapi_get_programs}()},
\code{\link{brapi_get_samples_sampleDbId}()},
\code{\link{brapi_get_samples}()},
\code{\link{brapi_get_scales_scaleDbId}()},
\code{\link{brapi_get_scales}()},
\code{\link{brapi_get_search_germplasm_searchResultsDbId}()},
\code{\link{brapi_get_search_images_searchResultsDbId}()},
\code{\link{brapi_get_search_markers_searchResultsDbId}()},
\code{\link{brapi_get_search_observationtables_searchResultsDbId}()},
\code{\link{brapi_get_search_observationunits_searchResultsDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_variables_searchResultsDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_studies_studyDbId_germplasm}()},
\code{\link{brapi_get_studies_studyDbId_layouts}()},
\code{\link{brapi_get_studies_studyDbId_observations}()},
\code{\link{brapi_get_studies_studyDbId_observationunits}()},
\code{\link{brapi_get_studies_studyDbId_observationvariables}()},
\code{\link{brapi_get_studies_studyDbId_table}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()},
\code{\link{brapi_get_traits_traitDbId}()},
\code{\link{brapi_get_traits}()},
\code{\link{brapi_get_trials_trialDbId}()},
\code{\link{brapi_get_trials}()},
\code{\link{brapi_get_variables_observationVariableDbId}()},
\code{\link{brapi_get_variables}()},
\code{\link{brapi_get_vendor_orders_orderId_plates}()},
\code{\link{brapi_get_vendor_orders_orderId_results}()},
\code{\link{brapi_get_vendor_orders_orderId_status}()},
\code{\link{brapi_get_vendor_orders}()},
\code{\link{brapi_get_vendor_plates_submissionId}()},
\code{\link{brapi_get_vendor_specifications}()},
\code{\link{brapi_post_allelematrices_search}()},
\code{\link{brapi_post_search_germplasm}()},
\code{\link{brapi_post_search_markers}()},
\code{\link{brapi_post_search_observationtables}()},
\code{\link{brapi_post_search_observationunits}()},
\code{\link{brapi_post_search_programs}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_variables}()}

Other Studies: 
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_seasons}()},
\code{\link{brapi_get_studies_studyDbId_germplasm}()},
\code{\link{brapi_get_studies_studyDbId_layouts}()},
\code{\link{brapi_get_studies_studyDbId_observations}()},
\code{\link{brapi_get_studies_studyDbId_observationunits}()},
\code{\link{brapi_get_studies_studyDbId_observationvariables}()},
\code{\link{brapi_get_studies_studyDbId_table}()},
\code{\link{brapi_get_studies_studyDbId}()},
\code{\link{brapi_get_studies}()},
\code{\link{brapi_get_studytypes}()}

Other Search Services: 
\code{\link{brapi_get_search_germplasm_searchResultsDbId}()},
\code{\link{brapi_get_search_images_searchResultsDbId}()},
\code{\link{brapi_get_search_markers_searchResultsDbId}()},
\code{\link{brapi_get_search_observationtables_searchResultsDbId}()},
\code{\link{brapi_get_search_observationunits_searchResultsDbId}()},
\code{\link{brapi_get_search_programs_searchResultsDbId}()},
\code{\link{brapi_get_search_samples_searchResultsDbId}()},
\code{\link{brapi_get_search_studies_searchResultsDbId}()},
\code{\link{brapi_get_search_variables_searchResultsDbId}()},
\code{\link{brapi_post_search_germplasm}()},
\code{\link{brapi_post_search_markers}()},
\code{\link{brapi_post_search_observationtables}()},
\code{\link{brapi_post_search_observationunits}()},
\code{\link{brapi_post_search_programs}()},
\code{\link{brapi_post_search_samples}()},
\code{\link{brapi_post_search_variables}()}
}
\author{
brapir generator package
}
\concept{Search Services}
\concept{Studies}
\concept{brapi_1.3}
