% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_serverinfo_expiration.R
\name{brapi_serverinfo_expiration}
\alias{brapi_serverinfo_expiration}
\title{brapi_serverinfo_expiration}
\usage{
brapi_serverinfo_expiration(con)
}
\arguments{
\item{con}{list; required: \strong{TRUE}; BrAPI connection object of class \code{brapi_con}.}
}
\description{
Display the expiration date and time of token if available.
}
\examples{
library(brapirv1)

if (interactive()) {
  con <- brapi_db()$testserver
  brapi_serverinfo_expiration(con)
}

}
