% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brapi_serverinfo_show.R
\name{brapi_serverinfo_show}
\alias{brapi_serverinfo_show}
\title{brapi_serverinfo_show}
\usage{
brapi_serverinfo_show(show = TRUE)
}
\arguments{
\item{show}{logical; default is TRUE}
}
\description{
Show informative messages or not
}
\seealso{
Other brapiutils: 
\code{\link{as.brapi_db}()},
\code{\link{brapi_canInternet}()},
\code{\link{brapi_checkCon}()},
\code{\link{brapi_connect}()},
\code{\link{brapi_db}()},
\code{\link{brapi_login}()},
\code{\link{is.brapi_con}()},
\code{\link{print.brapi_con}()},
\code{\link{print.brapi_db}()}
}
\concept{brapiutils}
